### 2023-04-02 #5324 @Kevin-Mattheus-Moerman
@markus1978 I am the AEiC on this track and here to help with the initial steps. For the moment, can you please address those potentially missing DOIs? ☝️ You can call `@editorialbot check references` here to check them, and use `@editorialbot generate pdf` to update the paper if needed.
