+++ paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={1--9},
   year={2016},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/sdata.2016.18}
 }
 
 @article{scheffler:2022,
@@ -17,14 +18,16 @@
   number={7907},
   pages={635--642},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41586-022-04501-x}
 }
 
 @article{ghiringhelli:2022,
   title={Shared Metadata for Data-Centric Materials Science},
   author={Ghiringhelli, Luca M and Baldauf, Carsten and Bereau, Tristan and Brockhauser, Sandor and Carbogno, Christian and Chamanara, Javad and Cozzini, Stefano and Curtarolo, Stefano and Draxl, Claudia and Dwaraknath, Shyam and others},
   journal={arXiv preprint arXiv:2205.14774},
-  year={2022}
+  year={2022},
+  doi={10.48550/arXiv.2205.14774}
 }
 
 @article{andersen:2021,
@@ -35,7 +38,8 @@
   number={1},
   pages={1--10},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41597-021-00974-z}
 }
 
 @article{jain:2013,
@@ -46,7 +50,8 @@
   number={1},
   pages={011002},
   year={2013},
-  publisher={American Institute of PhysicsAIP}
+  publisher={American Institute of PhysicsAIP},
+  doi={10.1063/1.4812323}
 }
 
 @article{saal:2013,
@@ -57,7 +62,8 @@
   number={11},
   pages={1501--1509},
   year={2013},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11837-013-0755-4}
 }
 
 @article{curtarolo:2012,
@@ -67,7 +73,8 @@
   volume={58},
   pages={218--226},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2012.02.005}
 }
 
 @article{huber:2020,
@@ -78,7 +85,8 @@
   number={1},
   pages={1--18},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41597-020-00638-4}
 }
 
 @article{himanen:2018,
@@ -89,7 +97,8 @@
   number={1},
   pages={1--10},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41524-018-0107-6}
 }
 
 @article{larsen:2017,
@@ -100,7 +109,8 @@
   number={27},
   pages={273002},
   year={2017},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1361-648X/aa680e}  
 }
 
 @article{konnecke:2015,
@@ -111,7 +121,8 @@
   number={1},
   pages={301--305},
   year={2015},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
+  doi={10.1107/S1600576714027575}
 }
 
 @article{ong:2013,
@@ -121,7 +132,8 @@
   volume={68},
   pages={314--319},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2012.10.028}
 }
 
 @article{draxl:2018,
@@ -132,14 +144,20 @@
   number={9},
   pages={676--682},
   year={2018},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
[…]
