### 2023-04-03 #5324 @markus1978
> @markus1978 I am the AEiC on this track and here to help with the initial steps. For the moment, can you please work on the following:
>
> * [x]  Can you address the those potentially missing DOIs? ☝️ You can call `@editorialbot check references` here to check them, and use `@editorialbot generate pdf` to update the paper if needed.
> * [x]  Can you extend all affiliations to also have the city/town. Also you may want to include the department/discipline for each since some have it and some don't.

@Kevin-Mattheus-Moerman thank you vera much for your comment and taking care of our contribution as a AEiC.

I added the DOIs and your bot seems to like my changes.

I also added the cities to all affiliations. Please note that some cities might appear twice: once because they belong to the name of the university, second as the city itself. I added the departments consistently in english where appropriate. Sebastian works as a data steward for the whole IKZ (affiliation 2) does not belong to a single department. The universities/institutes have their names according to their policies of using their names in english publications. I hope this is compatible with JOSS policies.
