### 2023-04-04 1c897ff
Added support for exported `.eln` file format from ElabFTW
ElabFTW exports the experiments in ELNFileFormat. The ElabFTW parser extracts, picks up relevant information from the metadata and populate the NOMAD ELN.

Changelog: Added
