+++ nomad/parsing/parsers.py
@@ -29,6 +29,7 @@ from nomad.datamodel.context import Context, ClientContext
 from .parser import MissingParser, BrokenParser, Parser, ArchiveParser, MatchingParserInterface
 from .artificial import EmptyParser, GenerateRandomParser, TemplateParser, ChaosParser
 from .tabular import TabularDataParser
+from nomad.datamodel.metainfo.eln.elabftw_parser import ELabFTWParser
 
 
 try:
@@ -662,6 +663,7 @@ parsers = [
         supported_compressions=['gz', 'bz2', 'xz']),
 
     TabularDataParser(),
+    ELabFTWParser(),
     ArchiveParser()
 ]
 
