+++ nomad/datamodel/metainfo/eln/elabftw_parser.py
@@ -1,0 +1,334 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import os
+from typing import Union, Iterable
+import json
+import re
+
+from nomad.datamodel import EntryArchive, EntryData
+from nomad.datamodel.data import ElnIntegrationCategory
+from nomad.metainfo import Package, Quantity, JSON, MSection, Datetime, SubSection, Section
+from nomad import utils
+from nomad.metainfo.util import camel_case_to_snake_case
+from nomad.parsing.parser import MatchingParser
+
+m_package = Package(name='ElabFTWParser')
+
+
+def _remove_at_sign_from_keys(obj):
+    for k, v in list(obj.items()):
+        if k.startswith('@'):
+            obj[k.lstrip('@')] = v
+            del obj[k]
+            k = k.lstrip('@')
+        if isinstance(v, dict):
+            obj[k] = _remove_at_sign_from_keys(v)
+        if isinstance(v, list):
+            for i, item in enumerate(v):
+                if isinstance(item, dict):
+                    obj[k][i] = _remove_at_sign_from_keys(item)
+    return obj
+
+
+def _create_file_section(file, graph, parent_folder_raw_path, logger=None):
+    try:
+        section = _element_type_section_mapping[file['type']]()
+    except Exception:
+        logger.error(f"Could not find type fo the file {file['id']}")
+        raise ELabFTWParserError(f"Could not find type fo the file {file['id']}")
+    section.m_update_from_dict(graph[file['id']])
+    try:
+        file_name = file['id'].split('./')[1]
+        full_path = os.path.join(parent_folder_raw_path, file_name)
+        section.post_process(file_name=full_path)
+    except Exception:
+        logger.error(f"Could not set the file path for file {file['id']}")
+    return section
+
+
+class ELabFTWParserError(Exception):
+    ''' eln parser related errors. '''
+    pass
+
+
+class ELabFTWExperimentLink(MSection):
+    '''
+    This class contains information from other experiments that are linked to this specific experiment.
+    the external link can be accessed using the query parameter #id:
+    https://demo.elabftw.net/experiments.php?mode=view&id={itemid}
+    '''
+    itemid = Quantity(type=str, description='id of the external experiment linked to this experiemnt')
+    title = Quantity(type=str, description='title of the external experiment')
+    elabid = Quantity(type=str, description='hashed id')
+    category = Quantity(type=str, description='Category/status of the external experiment link')
+
+
+class ELabFTWItemLink(ELabFTWExperimentLink):
+    '''
+    This class holds information of the items related to this specific experiment.
+    The external link can be accessed via setting the query parameter #related:
+    https://demo.elabftw.net/database.php?mode=show&related={itemid}
+    '''
+    bookable = Quantity(type=bool)
+
+
+class ELabFTWSteps(MSection):
+    '''
+    Steps recorded for the current experiment
+    '''
+    id = Quantity(type=str, description='id of the current step')
+    item_id = Quantity(type=str, description='item_id of the current experiemnt')
+    body = Quantity(type=str, description='title of the step')
+    ordering = Quantity(type=str, description='location of the current step in the overall order')
[…]
