### 2023-04-08 nomad-coe/nomad#63 @Andrew-S-Rosen
JOSS Review: Several packages missing from installation with `pip install nomad-lab[parsing]`
In the documentation, it states to use `pip install nomad-lab[parsing]` for parsing-related dependencies. However, doing so does not install all the pre-requisite packages for parsing. For instance, the following are not included but are needed: `mongoengine`, `jwt`, and `unidecode`. These are installed with the `pip install nomad-lab[infrastructure]` extras, but if the user is primarily interested in parsing, they may not install these extras (or know to install them).
