### 2023-04-08 #5324 @arosen93
I'm happy to review.

Coincidentally, learning the ins and outs of the code (particularly the parsing infrastructure) was on my to-do list for this week anyway.
