### 2023-04-08 nomad-coe/nomad#63 @TLCFEM
Not sure why `mongoengine` is involved if one only parses files. Any bt available?
