### 2023-04-08 nomad-coe/nomad#63 @Andrew-S-Rosen
I'll report back with a backtrace, but I ran `normalize_all()` as well per the [docs](https://nomad-lab.eu/prod/v1/docs/local_parsers.html), so perhaps that's where the source was.
