### 2023-04-11 nomad-coe/nomad#63 @arosen93
Details: I'm on Ubuntu and clean conda environment with Python 3.9. I ran `pip install nomad-lab[parsing]` but specifically installed `nomad-lab` 1.0.8. The reason for this is that I can't get any newer versions nor the `main` branch to parse files locally (discussed [here](https://github.com/nomad-coe/nomad/issues/44)). I also manually installed `python-magic` per the discussion here: https://github.com/nomad-coe/nomad/issues/62. Apologies; it looks like it was some modules but not the ones I listed above.

Here is my input:

```python
import sys
from nomad.client import parse, normalize_all

archive = parse('vasprun.xml')
normalize_all(archive)
section_run = archive.run[0]
python_dict = section_run.m_to_dict()
```

Note: The [docs](https://nomad-lab.eu/prod/v1/docs/local_parsers.html) say to use `archive = parse(sys.argv[1])` but that doesn't work for me, so I manually put in the filepath.

With that said, several packages are still missing and here are the tracebacks:

netCDF4:
```python
Cell In[1], line 5
      2 from nomad.client import parse, normalize_all
      4 # match and run the parser
----> 5 archive = parse('vasprun.xml')
      6 # run all normalizers
      7 normalize_all(archive)

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/client/processing.py:70, in parse(mainfile_path, parser_name, backend_factory, strict, logger)
     68 except Exception as e:
     69     logger.error('parsing was not successful', exc_info=e)
---> 70     raise e
     71 finally:
     72     os.chdir(cwd)

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/client/processing.py:67, in parse(mainfile_path, parser_name, backend_factory, strict, logger)
     65     mainfile_path = os.path.abspath(mainfile_path)
     66     os.chdir(os.path.abspath(os.path.dirname(mainfile_path)))
---> 67     parser.parse(mainfile_path, entry_archive, logger=logger)
     68 except Exception as e:
     69     logger.error('parsing was not successful', exc_info=e)

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/parsing/parser.py:249, in MatchingParserInterface.parse(self, mainfile, archive, logger)
    248 def parse(self, mainfile: str, archive: EntryArchive, logger=None):
--> 249     self.mainfile_parser.parse(mainfile, archive, logger)

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/parsing/parser.py:245, in MatchingParserInterface.mainfile_parser(self)
    243         logger = utils.get_logger(__name__)
    244         logger.error('cannot instantiate parser.', exc_info=e)
--> 245         raise e
    246 return self._mainfile_parser

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/parsing/parser.py:240, in MatchingParserInterface.mainfile_parser(self)
    238 if self._mainfile_parser is None:
    239     try:
--> 240         Parser = self.import_parser_class()
    241         self._mainfile_parser = Parser()
    242     except Exception as e:

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/parsing/parser.py:259, in MatchingParserInterface.import_parser_class(self)
    257 logger = utils.get_logger(__name__)
    258 logger.error('cannot import parser', exc_info=e)
--> 259 raise e

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/parsing/parser.py:254, in MatchingParserInterface.import_parser_class(self)
    252 try:
    253     module_path, parser_class = self._parser_class_name.rsplit('.', 1)
--> 254     module = importlib.import_module(module_path)
    255     return getattr(module, parser_class)
    256 except Exception as e:

File ~/anaconda3/envs/joss/lib/python3.8/importlib/__init__.py:127, in import_module(name, package)
    125             break
    126         level += 1
--> 127 return _bootstrap._gcd_import(name[level:], package, level)

File <frozen importlib._bootstrap>:1014, in _gcd_import(name, package, level)

File <frozen importlib._bootstrap>:991, in _find_and_load(name, import_)

File <frozen importlib._bootstrap>:975, in _find_and_load_unlocked(name, import_)

File <frozen importlib._bootstrap>:671, in _load_unlocked(spec)

File <frozen importlib._bootstrap_external>:843, in exec_module(self, module)

File <frozen importlib._bootstrap>:219, in _call_with_frames_removed(f, *args, **kwds)

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/electronicparsers/__init__.py:52
     50 from .vasp.parser import VASPParser
     51 from .wien2k.parser import Wien2kParser
---> 52 from .yambo.parser import YamboParser

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/electronicparsers/yambo/__init__.py:19
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     17 # limitations under the License.
     18 #
---> 19 from .parser import YamboParser

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/electronicparsers/yambo/parser.py:23
     21 import logging
     22 from datetime import datetime
---> 23 from netCDF4 import Dataset  # pylint: disable=no-name-in-module
     24 from ase.data import chemical_symbols
     26 from nomad.parsing.file_parser import TextParser, Quantity, FileParser

ModuleNotFoundError: No module named 'netCDF4'
```

matid:
```python
Cell In[1], line 7
      5 archive = parse('vasprun.xml')
      6 # run all normalizers
----> 7 normalize_all(archive)
      9 # get the 'main section' section_run as a metainfo object
     10 section_run = archive.run[0]

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/client/processing.py:106, in normalize_all(entry_archive, logger)
    102 def normalize_all(entry_archive, logger=None):
    103     '''
    104     Parse the downloaded entry and run the whole normalizer chain.
    105     '''
--> 106     from nomad import normalizing
    108     for normalizer in normalizing.normalizers:
    109         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/normalizing/system.py:26
     24 import json
     25 import re
---> 26 from matid import SymmetryAnalyzer, Classifier
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D
     29 from nomad import atomutils, archive

ModuleNotFoundError: No module named 'matid'
```

bitarray:
```python
Cell In[1], line 7
      5 archive = parse('vasprun.xml')
      6 # run all normalizers
----> 7 normalize_all(archive)
      9 # get the 'main section' section_run as a metainfo object
     10 section_run = archive.run[0]

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/client/processing.py:106, in normalize_all(entry_archive, logger)
    102 def normalize_all(entry_archive, logger=None):
    103     '''
    104     Parse the downloaded entry and run the whole normalizer chain.
    105     '''
--> 106     from nomad import normalizing
    108     for normalizer in normalizing.normalizers:
    109         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/normalizing/__init__.py:43
     41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
---> 43 from .dos import DosNormalizer
     44 from .normalizer import Normalizer
     45 from .band_structure import BandStructureNormalizer

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad/normalizing/dos.py:23
     20 from nptyping import NDArray
     22 from nomad import config
---> 23 from nomad_dos_fingerprints import DOSFingerprint
     24 from nomad.datamodel.metainfo.simulation.calculation import (
     25     Dos, DosFingerprint, BandGap)
     26 from nomad.atomutils import get_volume

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad_dos_fingerprints/__init__.py:1
----> 1 from .DOSfingerprint import DOSFingerprint
      2 from .grid import Grid
      3 from .similarity import tanimoto_similarity

File ~/anaconda3/envs/joss/lib/python3.8/site-packages/nomad_dos_fingerprints/DOSfingerprint.py:2
      1 import numpy as np
----> 2 from bitarray import bitarray
      3 from functools import partial
      5 from .grid import Grid

ModuleNotFoundError: No module named 'bitarray'
```

As a disclaimer, I now happen to be one of the reviewers for the JOSS paper submission and have updated the titles accordingly.
