### 2023-04-11 nomad-coe/nomad#63 @Andrew-S-Rosen
# Overview

It looks like the issue arises because, in the [docs](https://nomad-lab.eu/prod/v1/docs/local_parsers.html), it suggests to use `normalize_all()` on the archive that is parsed, but the requirements for `normalize_all()` are not all installed with the `[parsing]` extras. You _can_ run `parse()` appropriately, so I guess it's a matter of if you want the user to be able to `normalize_all()` if they install the `[parsing]` extras only.

## Details of Build

System: Ubuntu via WSL

```bash
conda create --name joss_test python=3.9
conda activate joss_test
pip install -e .[parsing] # from https://github.com/arosen93/nomad/tree/develop
```

Note: I am using a fork of the repo because https://github.com/nomad-coe/nomad/issues/44 is preventing me from running the parsers unless I hard-code where the `tools.json` file is located.

## Code to Run

```python
from nomad.client import parse, normalize_all

archive = parse('vasprun.xml')
normalize_all(archive)
section_run = archive.run[0]
python_dict = section_run.m_to_dict()
```
[vasprun.xml.txt](https://github.com/nomad-coe/nomad/files/11205459/vasprun.xml.txt)

Note: The [docs](https://nomad-lab.eu/prod/v1/docs/local_parsers.html) say to use `archive = parse(sys.argv[1])` but that doesn't work for me:

```python
----> 1 archive = parse(sys.argv[1])

IndexError: list index out of range
```

So, I manually put in the filepath instead.

## Tracebacks

```python
Cell In[1], line 7
      5 archive = parse('vasprun.xml')
      6 # run all normalizers
----> 7 normalize_all(archive)
      9 # get the 'main section' section_run as a metainfo object
     10 section_run = archive.run[0]

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:26
     24 import json
     25 import re
---> 26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
     29 from nomad import atomutils, archive

File /mnt/c/Users/asros/github/nomad/dependencies/matid/matid/__init__.py:1
----> 1 from matid.classification.classifier import Classifier
      2 from matid.classification.periodicfinder import PeriodicFinder
      3 from matid.symmetry.symmetryanalyzer import SymmetryAnalyzer

File /mnt/c/Users/asros/github/nomad/dependencies/matid/matid/classification/classifier.py:17
      7 from matid.classifications import \
      8     Surface, \
      9     Atom, \
   (...)
     14     Class2D, \
     15     Class3D
     16     # Crystal
---> 17 import matid.geometry
     18 from matid.data import constants
     19 from matid.classification.periodicfinder import PeriodicFinder

File /mnt/c/Users/asros/github/nomad/dependencies/matid/matid/geometry.py:20
     17 from matid.core.linkedunits import Substitution
     18 import matid.geometry
---> 20 from sklearn.cluster import DBSCAN
     22 from scipy.spatial import Delaunay
     23 import spglib

ModuleNotFoundError: No module named 'sklearn'
```

After `pip install scikit-learn`:

```python
Cell In[1], line 7
      5 archive = parse('vasprun.xml')
      6 # run all normalizers
----> 7 normalize_all(archive)
      9 # get the 'main section' section_run as a metainfo object
     10 section_run = archive.run[0]

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:29
     26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
---> 29 from nomad import atomutils, archive
     30 from nomad.atomutils import Formula
     31 from nomad.units import ureg

File /mnt/c/Users/asros/github/nomad/nomad/archive/__init__.py:40
     36 from .storage import (
     37     serialise_container, write_archive, read_archive, ArchiveError, ArchiveReader, ArchiveWriter,
     38     ArchiveDict, ArchiveList, ArchiveItem)
     39 from .query import query_archive, filter_archive, ArchiveQueryError
---> 40 from .partial import (
     41     read_partial_archive_from_mongo, read_partial_archives_from_mongo,
     42     write_partial_archive_to_mongo, delete_partial_archives_from_mongo,
     43     create_partial_archive, compute_required_with_referenced)
     44 from .required import RequiredReader, RequiredValidationError

File /mnt/c/Users/asros/github/nomad/nomad/archive/partial.py:21
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     16 # limitations under the License.
     17 #
     19 from typing import Any, Tuple, Dict, Union, List
---> 21 from nomad import infrastructure, config
     22 from nomad.metainfo import MSection, Definition, Quantity, Reference, SubSection, Section
     23 from nomad.datamodel import EntryArchive

File /mnt/c/Users/asros/github/nomad/nomad/infrastructure.py:30
     28 import shutil
     29 from elasticsearch_dsl import connections
---> 30 from mongoengine import connect, disconnect
     31 from mongoengine.connection import ConnectionFailure
     32 import smtplib

ModuleNotFoundError: No module named 'mongoengine'
```

After `pip install mongoengine`:

```python
Cell In[1], line 5
      2 from nomad.client import parse, normalize_all
      4 archive = parse('vasprun.xml')
----> 5 normalize_all(archive)
      6 section_run = archive.run[0]
      7 python_dict = section_run.m_to_dict()

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:29
     26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
---> 29 from nomad import atomutils, archive
     30 from nomad.atomutils import Formula
     31 from nomad.units import ureg

File /mnt/c/Users/asros/github/nomad/nomad/archive/__init__.py:40
     36 from .storage import (
     37     serialise_container, write_archive, read_archive, ArchiveError, ArchiveReader, ArchiveWriter,
     38     ArchiveDict, ArchiveList, ArchiveItem)
     39 from .query import query_archive, filter_archive, ArchiveQueryError
---> 40 from .partial import (
     41     read_partial_archive_from_mongo, read_partial_archives_from_mongo,
     42     write_partial_archive_to_mongo, delete_partial_archives_from_mongo,
     43     create_partial_archive, compute_required_with_referenced)
     44 from .required import RequiredReader, RequiredValidationError

File /mnt/c/Users/asros/github/nomad/nomad/archive/partial.py:21
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     16 # limitations under the License.
     17 #
     19 from typing import Any, Tuple, Dict, Union, List
---> 21 from nomad import infrastructure, config
     22 from nomad.metainfo import MSection, Definition, Quantity, Reference, SubSection, Section
     23 from nomad.datamodel import EntryArchive

File /mnt/c/Users/asros/github/nomad/nomad/infrastructure.py:37
     35 from keycloak.exceptions import KeycloakAuthenticationError, KeycloakGetError
     36 import json
---> 37 import jwt
     38 from datetime import datetime
     39 import re

ModuleNotFoundError: No module named 'jwt'
```

After `pip install jwt`:

```python
Cell In[1], line 5
      2 from nomad.client import parse, normalize_all
      4 archive = parse('vasprun.xml')
----> 5 normalize_all(archive)
      6 section_run = archive.run[0]
      7 python_dict = section_run.m_to_dict()

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:29
     26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
---> 29 from nomad import atomutils, archive
     30 from nomad.atomutils import Formula
     31 from nomad.units import ureg

File /mnt/c/Users/asros/github/nomad/nomad/archive/__init__.py:40
     36 from .storage import (
     37     serialise_container, write_archive, read_archive, ArchiveError, ArchiveReader, ArchiveWriter,
     38     ArchiveDict, ArchiveList, ArchiveItem)
     39 from .query import query_archive, filter_archive, ArchiveQueryError
---> 40 from .partial import (
     41     read_partial_archive_from_mongo, read_partial_archives_from_mongo,
     42     write_partial_archive_to_mongo, delete_partial_archives_from_mongo,
     43     create_partial_archive, compute_required_with_referenced)
     44 from .required import RequiredReader, RequiredValidationError

File /mnt/c/Users/asros/github/nomad/nomad/archive/partial.py:21
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     16 # limitations under the License.
     17 #
     19 from typing import Any, Tuple, Dict, Union, List
---> 21 from nomad import infrastructure, config
     22 from nomad.metainfo import MSection, Definition, Quantity, Reference, SubSection, Section
     23 from nomad.datamodel import EntryArchive

File /mnt/c/Users/asros/github/nomad/nomad/infrastructure.py:40
     38 from datetime import datetime
     39 import re
---> 40 import unidecode
     42 from nomad import config, utils
     43 from nomad.utils.structlogging import get_logger

ModuleNotFoundError: No module named 'unidecode'
```

AFter `pip install unidecode`:

```python
Cell In[1], line 5
      2 from nomad.client import parse, normalize_all
      4 archive = parse('vasprun.xml')
----> 5 normalize_all(archive)
      6 section_run = archive.run[0]
      7 python_dict = section_run.m_to_dict()

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:29
     26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
---> 29 from nomad import atomutils, archive
     30 from nomad.atomutils import Formula
     31 from nomad.units import ureg

File /mnt/c/Users/asros/github/nomad/nomad/archive/__init__.py:40
     36 from .storage import (
     37     serialise_container, write_archive, read_archive, ArchiveError, ArchiveReader, ArchiveWriter,
     38     ArchiveDict, ArchiveList, ArchiveItem)
     39 from .query import query_archive, filter_archive, ArchiveQueryError
---> 40 from .partial import (
     41     read_partial_archive_from_mongo, read_partial_archives_from_mongo,
     42     write_partial_archive_to_mongo, delete_partial_archives_from_mongo,
     43     create_partial_archive, compute_required_with_referenced)
     44 from .required import RequiredReader, RequiredValidationError

File /mnt/c/Users/asros/github/nomad/nomad/archive/partial.py:21
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     16 # limitations under the License.
     17 #
     19 from typing import Any, Tuple, Dict, Union, List
---> 21 from nomad import infrastructure, config
     22 from nomad.metainfo import MSection, Definition, Quantity, Reference, SubSection, Section
     23 from nomad.datamodel import EntryArchive

File /mnt/c/Users/asros/github/nomad/nomad/infrastructure.py:43
     40 import unidecode
     42 from nomad import config, utils
---> 43 from nomad.utils.structlogging import get_logger
     45 # The metainfo is defined and used during imports. This is problematic.
     46 # We import all parsers very early in the infrastructure setup. This will populate
     47 # the metainfo with parser specific definitions, before the metainfo might be used.
     48 from nomad.parsing import parsers  # pylint: disable=unused-import

File /mnt/c/Users/asros/github/nomad/nomad/utils/structlogging.py:42
     40 from typing import cast, Any
     41 import logging
---> 42 import structlog
     43 from structlog.processors import StackInfoRenderer, format_exc_info, TimeStamper, JSONRenderer
     44 from structlog.stdlib import LoggerFactory

ModuleNotFoundError: No module named 'structlog'

After `pip install structlog`:

```python
Cell In[1], line 5
      2 from nomad.client import parse, normalize_all
      4 archive = parse('vasprun.xml')
----> 5 normalize_all(archive)
      6 section_run = archive.run[0]
      7 python_dict = section_run.m_to_dict()

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:94, in normalize_all(entry_archive, logger)
     90 def normalize_all(entry_archive, logger=None):
     91     '''
     92     Parse the downloaded entry and run the whole normalizer chain.
     93     '''
---> 94     from nomad import normalizing
     96     for normalizer in normalizing.normalizers:
     97         if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/__init__.py:41
     19 '''
     20 After parsing entries have to be normalized with a set of *normalizers*.
     21 In NOMAD-coe those were programmed in python (we'll reuse) and scala (we'll rewrite).
   (...)
     36     :members:
     37 '''
     39 from typing import List, Any, Iterable, Type
---> 41 from .system import SystemNormalizer
     42 from .optimade import OptimadeNormalizer
     43 from .dos import DosNormalizer

File /mnt/c/Users/asros/github/nomad/nomad/normalizing/system.py:29
     26 from matid import SymmetryAnalyzer, Classifier  # pylint: disable=import-error
     27 from matid.classifications import Class0D, Atom, Class1D, Material2D, Surface, Class3D  # pylint: disable=import-error
---> 29 from nomad import atomutils, archive
     30 from nomad.atomutils import Formula
     31 from nomad.units import ureg

File /mnt/c/Users/asros/github/nomad/nomad/archive/__init__.py:40
     36 from .storage import (
     37     serialise_container, write_archive, read_archive, ArchiveError, ArchiveReader, ArchiveWriter,
     38     ArchiveDict, ArchiveList, ArchiveItem)
     39 from .query import query_archive, filter_archive, ArchiveQueryError
---> 40 from .partial import (
     41     read_partial_archive_from_mongo, read_partial_archives_from_mongo,
     42     write_partial_archive_to_mongo, delete_partial_archives_from_mongo,
     43     create_partial_archive, compute_required_with_referenced)
     44 from .required import RequiredReader, RequiredValidationError

File /mnt/c/Users/asros/github/nomad/nomad/archive/partial.py:21
      1 #
      2 # Copyright The NOMAD Authors.
      3 #
   (...)
     16 # limitations under the License.
     17 #
     19 from typing import Any, Tuple, Dict, Union, List
---> 21 from nomad import infrastructure, config
     22 from nomad.metainfo import MSection, Definition, Quantity, Reference, SubSection, Section
     23 from nomad.datamodel import EntryArchive

File /mnt/c/Users/asros/github/nomad/nomad/infrastructure.py:43
     40 import unidecode
     42 from nomad import config, utils
---> 43 from nomad.utils.structlogging import get_logger
     45 # The metainfo is defined and used during imports. This is problematic.
     46 # We import all parsers very early in the infrastructure setup. This will populate
     47 # the metainfo with parser specific definitions, before the metainfo might be used.
     48 from nomad.parsing import parsers  # pylint: disable=unused-import

[…]
