### 2023-04-11 nomad-coe/nomad#63 @TLCFEM
Looks like normalisers introduce too many dependencies, which needs to be addressed.

> AttributeError: 'list' object has no attribute 'metadata'

For this one, `parse` returns a list, but not sure why the doc does not pick this up. Just change `section_run = archive.run[0]` to `section_run = archive[0].run[0]`, shall work for the file provided.
