### 2023-04-11 nomad-coe/nomad#63 @Andrew-S-Rosen
Thanks!

I don't think that solves it because the error happens in `normalize_all` before the `section_run`. Doing `archive.metadata` returns the same error. Here is the `archive` list: `[EntryArchive(run, workflow, workflow2, metadata)]`.

```python
Cell In[66], line 4
      1 from nomad.client import parse, normalize_all
      3 archive = parse('vasprun.xml')
----> 4 normalize_all(archive)
      5 section_run = archive[0].run[0]
      6 python_dict = section_run.m_to_dict()

File /mnt/c/Users/asros/github/nomad/nomad/client/processing.py:97, in normalize_all(entry_archive, logger)
     94 from nomad import normalizing
     96 for normalizer in normalizing.normalizers:
---> 97     if normalizer.domain is None or normalizer.domain == entry_archive.metadata.domain:
     98         normalize(normalizer, entry_archive, logger=logger)

AttributeError: 'list' object has no attribute 'metadata'
```
