### 2023-04-11 nomad-coe/nomad#63 @TLCFEM
You are very correct, I meant `archive = parse('vasprun.xml')` to `archive = parse('vasprun.xml')[0]`.
