### 2023-04-11 nomad-coe/nomad#63 @Andrew-S-Rosen
Beautiful! Yes, that does indeed make more sense. I got output! Thanks for the help and apologies for some of my own confusion to start off.
