### 2023-04-11 nomad-coe/nomad#63 @TLCFEM
The doc somehow just cannot keep up with the changes, I'll update that for the moment.

Dependency issue linked: https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR/-/issues/1423
