### 2023-04-12 0523495
Resolve "Make "available_properties" into a derived value."
