### 2023-04-13 nomad-coe/nomad#63 @lauri-codes
Thanks @arosen93 for the feedback. Indeed we need to update the docs and sort out the dependencies for local parsing.
