### 2023-04-14 578166a
Improved automatic topology detection performance.
Changelog: Changed
