+++ tests/normalizing/test_topology.py
@@ -20,23 +20,27 @@ import numpy as np
 from collections import defaultdict
 import pytest
 
-from tests.normalizing.conftest import (
+from nomad.units import ureg
+from tests.normalizing.conftest import (  # pylint: disable=unused-import
     get_template_for_structure,
     get_template_topology,
-    single_Cu_surface_atoms,
-    single_Cu_surface_topology,
-    single_Cr_surface_atoms,
-    single_Cr_surface_topology,
-    stacked_Cu_Ni_surface,
-    single_2D_graphene_layer_atoms,
-    single_2D_graphene_layer_topology,
-    single_2D_BN_layer_atoms,
-    single_2D_BN_layer_topology,
-    single_2D_MoS2_layer_atoms,
-    single_2D_MoS2_layer_topology,
-    stacked_C_BN_2D_layers
+    conv_bcc,
+    conv_fcc,
+    rattle,
+    stack,
+    surf,
+    single_cu_surface_topology,
+    single_cr_surface_topology,
+    stacked_cu_ni_surface_topology,
+    graphene,
+    graphene_topology,
+    boron_nitride,
+    boron_nitride_topology,
+    mos2,
+    mos2_topology,
+    stacked_graphene_boron_nitride_topology,
+    projection
 )
-from tests.normalizing.conftest import projection  # pylint: disable=unused-import
 
 
 def assert_topology(topology):
@@ -81,7 +85,6 @@ def test_topology_calculation(pbc):
 
     # Test the original structure
     original = topology[0]
-    assert original.structural_type == 'unavailable'
     assert original.atoms_ref.positions.shape == (6, 3)
     assert len(original.atoms_ref.labels) == 6
     assert original.atoms_ref.lattice_vectors.shape == (3, 3)
@@ -165,26 +168,61 @@ def test_no_topology(fixture, request):
 
 @pytest.mark.skip
 @pytest.mark.parametrize('surface, ref_topologies', [
-    pytest.param(single_Cu_surface_atoms()[0], single_Cu_surface_topology(),
+    pytest.param(rattle(surf(conv_fcc('Cu'), [1, 0, 0])), single_cu_surface_topology(),
                  id='single surface Cu FCC 100'),
-    pytest.param(single_Cu_surface_atoms()[1], single_Cu_surface_topology(),
+    pytest.param(rattle(surf(conv_fcc('Cu'), [1, 1, 0])), single_cu_surface_topology(),
                  id='single surface Cu FCC 110'),
-    pytest.param(single_Cr_surface_atoms()[0], single_Cr_surface_topology(),
+    pytest.param(rattle(surf(conv_bcc('Cr'), [1, 0, 0])), single_cr_surface_topology(),
                  id='single surface Cr BCC 100'),
-    pytest.param(single_Cr_surface_atoms()[1], single_Cr_surface_topology(),
+    pytest.param(rattle(surf(conv_bcc('Cr'), [1, 1, 0])), single_cr_surface_topology(),
                  id='single surface Cr BCC 110'),
-    pytest.param(stacked_Cu_Ni_surface()[0], stacked_Cu_Ni_surface()[1],
-                 id='stacked surfaces of Cu and Ni'),
-    pytest.param(single_2D_graphene_layer_atoms(), single_2D_graphene_layer_topology(),
+    pytest.param(rattle(
+        stack(
+            surf(conv_fcc('Cu'), [1, 0, 0], vacuum=0),
+            surf(conv_fcc('Ni'), [1, 0, 0], vacuum=0)
+        )),
+        stacked_cu_ni_surface_topology(),
+        id='stacked surfaces of Cu and Ni'),
+    pytest.param(rattle(graphene()), graphene_topology(),
[…]
