+++ tests/normalizing/conftest.py
@@ -37,6 +37,7 @@ from nomad.datamodel.results import (
     System as ResultSystem
 )
 from nomad.datamodel.optimade import Species
+from nomad.normalizing.common import cell_from_ase_atoms, nomad_atoms_from_ase_atoms
 from nomad.datamodel.metainfo.simulation.run import Run, Program
 from nomad.datamodel.metainfo.simulation.method import (
     Method, BasisSet, Electronic, DFT, XCFunctional, Functional,
@@ -1215,16 +1216,33 @@ def crystal_structure_properties(crystal_structure):
     return crystal_structure_property
 
 
-def single_Cu_surface_atoms():
-    # create simple Cu surface
-    Cu_fcc_100 = ase.build.fcc100('Cu', (3, 5, 5), vacuum=10, periodic=True)
-    Cu_fcc_100.rattle(stdev=0.001, seed=None, rng=None)
-    Cu_fcc_110 = ase.build.fcc110('Cu', (3, 5, 5), vacuum=10, periodic=True)
-    Cu_fcc_110.rattle(stdev=0.001, seed=None, rng=None)
-    return Cu_fcc_100, Cu_fcc_110
+def conv_fcc(symbols):
+    return ase.build.bulk(symbols, crystalstructure='fcc', a=3.61, cubic=True)
 
 
-def single_Cu_surface_prototype():
+def conv_bcc(symbols):
+    return ase.build.bulk(symbols, crystalstructure='bcc', cubic=True)
+
+
+def surf(conv_cell, indices, layers=[3, 3, 2], vacuum=10):
+    surface = ase.build.surface(conv_cell, indices, layers[2], vacuum=vacuum, periodic=True)
+    surface *= [layers[0], layers[1], 1]
+    return surface
+
+
+def stack(a, b):
+    stacked = ase.build.stack(a, b, axis=2, distance=3, maxstrain=6.7)
+    stacked.set_pbc([True, True, False])
+    ase.build.add_vacuum(stacked, 10)
+    return stacked
+
+
+def rattle(atoms):
+    atoms.rattle(stdev=0.001, seed=7, rng=None)
+    return atoms
+
+
+def single_cu_surface_prototype():
     prototype_Cu_fcc = Prototype()
     prototype_Cu_fcc.aflow_id = "A_cF4_225_a"
     prototype_Cu_fcc.assignment_method = "normalized-wyckoff"
@@ -1233,30 +1251,24 @@ def single_Cu_surface_prototype():
     return prototype_Cu_fcc
 
 
-def single_Cu_surface_topology() -> List[ResultSystem]:
+def single_cu_surface_topology() -> List[ResultSystem]:
     '''Copper surface topology'''
-    prototype_Cu_fcc = single_Cu_surface_prototype()
-    cell = Cell(
-        a=3.610000000000001 * ureg.angstrom,
-        b=3.610000000000001 * ureg.angstrom,
-        c=3.610000000000001 * ureg.angstrom,
-        alpha=1.5707963267948966 * ureg.rad,
-        beta=1.5707963267948966 * ureg.rad,
-        gamma=1.5707963267948966 * ureg.rad
-    )
+    prototype_Cu_fcc = single_cu_surface_prototype()
+    conv_cell = conv_fcc('Cu')
+    surface = surf(conv_cell, (1, 0, 0))
 
-    # create Cu topology system
     label_Cu = 'subsystem'
+    n_atoms = len(surface)
     structural_type_Cu = 'surface'
     elements_Cu = ['Cu']
-    formula_hill_Cu = 'Cu75'
-    formula_reduced_Cu = 'Cu75'
[…]

[…]
