### 2023-04-18 d3b1369
Add `Mesh` section for other physical meshes to inherit
