+++ nomad/datamodel/metainfo/simulation/method.py
@@ -30,87 +30,193 @@ from ..common import FastAccess
 m_package = Package()
 
 
-class KMesh(MSection):
+class Mesh(MSection):
     '''
-    Contains the settings for a uniformly spaced k-point grid (if employed).
+    Contains the settings for a sampling mesh.
+    Supports uniformly-spaced meshes and symmetry-reduced representations.
     '''
 
     m_def = Section(validate=False)
+
+    dimensionality = Quantity(
+        type=np.int32,
+        shape=[],
+        description='''
+        Dimensionality of the mesh.
+        ''')
+
+    sampling_method = Quantity(
+        type=MEnum(
+            'Gamma-centered', 'Monkhorst-Pack', 'Gamma-offcenter', 'Line-path',
+            'Equidistant', 'Logarithmic', 'Gauss-Legendre', 'Gauss-Laguerre'
+            'Clenshaw-Curtis', 'Newton-Cotes', 'Gauss-Hermite'
+        ),
+        shape=[],
+        description='''
+        Method used to generate the mesh:
+
+        | Name      | Description                      | Reference             |
+
+        | --------- | -------------------------------- | --------------------- |
+
+        | `'Gamma-centered'` | Regular mesh is centered around Gamma. No offset. |
+
+        | `'Monkhorst-Pack'` | Regular mesh with an offset of half the reciprocal lattice vector. |
+
+        | `'Gamma-offcenter'` | Regular mesh with an offset that is neither `'Gamma-centered'`, nor `'Monkhorst-Pack'`. |
+
+        | `'Line-path'` | Line path along high-symmetry points. Typically employed for simualting band structures. |
+
+        | `'Equidistant'`  | Equidistant 1D grid (also known as 'Newton-Cotes')                      |
+
+        | `'Logarithmic'`  | log distance 1D grid               |
+
+        | `'Gauss-Legendre'` | Quadrature rule for integration using Legendre polynomials |
+
+        | `'Gauss-Laguerre'` | Quadrature rule for integration using Laguerre polynomials |
+
+        | `'Clenshaw-Curtis'`  | Quadrature rule for integration using Chebyshev polynomials using discrete cosine transformations |
+
+        | `'Gauss-Hermite'`  | Quadrature rule for integration using Hermite polynomials |
+        ''')
 
     n_points = Quantity(
         type=np.int32,
         shape=[],
         description='''
-        Number of k points in the mesh (i.e. the k points used to evaluate energy_total).
+        Total number of points in the mesh, accounting for the multiplicities.
         ''')
 
     grid = Quantity(
         type=np.int32,
-        shape=[3],
+        shape=['dimensionality'],
         description='''
-        k-point grid used in notation [nx, ny, nz] in fractional coordinates.
-        ''')
-
-    generation_method = Quantity(
-        type=str,
-        shape=[],
-        description='''
-        Method used to generate the k points.
+        Amount of mesh point sampling along each axis, i.e. [nx, ny, nz].
         ''')
 
     points = Quantity(
-        type=np.float64,
-        shape=['n_points', 3],
+        type=np.complex128,
+        shape=['*', 'dimensionality'],
         description='''
-        List of all the k points in the $k$-point mesh. These are the k point used to
-        evaluate energy_total, and are in fractional coordinates (in the basis of the
-        reciprocal-lattice vectors).
+        List of all the points in the mesh.
+        ''')
+
+    multiplicities = Quantity(
+        type=np.float64,
+        shape=['*'],
+        description='''
+        The amount of times the same point reappears. These are accounted for in `n_points`.
+        A value larger than 1, typically indicates a symmtery operation that was applied to the mesh.
         ''')
 
[…]
