### 2023-04-19 c1c7c04
Fixes metainfo generation from nexus file
