+++ gui/workbox.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   InjectManifest: options => {
     // We override the default workbox cache max size.
-    options.maximumFileSizeToCacheInBytes = 20 * 1024 * 1024
+    options.maximumFileSizeToCacheInBytes = 25 * 1024 * 1024
     return options
   }
 }
