+++ nomad/datamodel/metainfo/eln/ikz_hall/hall_instrument.py
@@ -27,6 +27,9 @@ class MeasurementStateMachine(MSection):
     system_model = Quantity(
         type=str,
         description='What is the Model number of your used Hall Measurement system? 0=[75XX-LVWR(-HS)] 1=[75XX-LVWR-SWT(-HS), 77XX-LVWR] 2=[75XX-HVWR(-HS), 77XX-HVWR] 3=[76XX] 4=[95XX-LVWR(-HS)] 5=[95XX-LVWR-SWT(-HS), 97XX-LVWR] 6=[95XX-HVWR(-HS), 97XX-HVWR] 7=[77XXA] 8=[97XXA]')
+    # switch_matrix = Quantity(
+    #    type=str,
+    #    description='What is the Model number of your used Hall matrix? 0=[Lake Shore 775] 1=[Lake Shore 776]')
     wiring = Quantity(
         type=str,
         description='What is the wiring setup,either 0=[van der Pauw] or 1=[Hall Bar]?')
@@ -87,7 +90,7 @@ class Keithley6485(MSection):
         type=str,
         description='If the digital filter is turned on, what is its mode? 0=[Repeat] 1=[Moving]')
     digital_filter_count = Quantity(
-        type=np.dtype(np.float64),
+        type=int,
         description='If the digital filter is turned on, what is its count? Value must be between 1 and 100.')
     advanced_filter = Quantity(
         type=bool,
@@ -100,94 +103,95 @@ class Keithley6485(MSection):
 class Keithley220(MSection):
     '''Representation of an instrument'''
     compliance_voltage = Quantity(
-        type=int,
-        description='FILL THE DESCRIPTION')
+        type=np.dtype(np.float64),
+        description='What is the voltage compliance for your current source? in V. (default = 100 V)',
+        unit="volt")
 
 
 class Keithley2000(MSection):
     '''Representation of an instrument'''
     reading_rate = Quantity(
-        type=int,
-        description='FILL THE DESCRIPTION')
+        type=str,
+        description='What is the set reading rate for the Keithley2000? 0=[Slow] 1=[Medium(default)] 2=[Fast]')
     digital_filter = Quantity(
         type=bool,
-        description='FILL THE DESCRIPTION')
+        description='Is a digitial filter turned on? True=[on (default)]')
     digital_filter_type = Quantity(
-        type=int,
-        description='FILL THE DESCRIPTION')
+        type=str,
+        description='If the digital filter is turned on, what is its mode? 0=[Repeat] 1=[Moving]')
     digital_filter_count = Quantity(
         type=int,
-        description='FILL THE DESCRIPTION')
+        description='If the digital filter is turned on, what is its count? Value must be between 1 and 100.')
 
 
 class Keithley2182(MSection):
     '''Representation of an instrument'''
     line_synchronization = Quantity(
         type=bool,
-        description='FILL THE DESCRIPTION')
+        description='Is line synchronization turned on? True=[on (default)]')
     reading_rate = Quantity(
-        type=int,
-        description='FILL THE DESCRIPTION')
+        type=str,
+        description='What is the set reading rate for the Keithley2182? 0=[Slow(default)] 1=[Medium] 2=[Fast]')
     analog_filter = Quantity(
         type=bool,
-        description='FILL THE DESCRIPTION')
+        description='Is the analog filter turned on? False=[off (default)]')
     digital_filter = Quantity(
         type=bool,
-        description='FILL THE DESCRIPTION')
+        description='Is a digitial filter turned on? True=[on (default)]')
     digital_filter_type = Quantity(
-        type=int,
-        description='FILL THE DESCRIPTION')
+        type=str,
+        description='If the digital filter is turned on, what is its mode? 0=[Repeat] 1=[Moving]')
     digital_filter_window = Quantity(
-        type=int,
[…]
