### 2023-04-19 9622d91
Added plugin support for NOMAD Oasis.
Changelog: Added
