+++ docs/plugins.md
@@ -1,5 +1,9 @@
 Plugins allow you to add Python-based functionality to NOMAD without a custom build
-NOMAD image or package. We support different kinds of plugins:
+NOMAD image or release. Plugins can be installed at NOMAD start-up time. Therefore, you can
+configure each NOMAD (Oasis) with a different custom set of plugins or disable unnecessary
+plugins.
+
+We support different kinds of plugins:
 
 - Python **schema**, read also [Python schema documentation](schema/python.md).
 - **parser**, read also [parser development documentation](develop/parser.md).
@@ -10,7 +14,9 @@ NOMAD image or package. We support different kinds of plugins:
 
 We provide template projects on GitHub. You can fork these projects and follow the
 instructions in their `README.md`. These instructions will give you everything you
-need to run your plugin as a plugin developer:
+need to run and test your plugin as a plugin developer.
+The following sections here contain more background information and explain how to
+add plugins to a NOMAD installation.
 
 - [schema plugin](https://github.com/nomad-coe/nomad-schema-plugin-example)
 - [parser plugin](https://github.com/nomad-coe/nomad-parser-plugin-example)
@@ -53,7 +59,7 @@ for more details:
 
 ### Plugin metadata
 
-The file `nomad_plugin.yaml` contains the installation independent plugin metadata:
+The file `nomad_plugin.yaml` contains the installation independent *plugin metadata*:
 
 ```yaml
 {{ file_contents('examples/plugins/schema/nomadschemaexample/nomad_plugin.yaml') }}
@@ -86,20 +92,76 @@ each `options` entry with `python_package` will be merged with the data in the p
 Please note that `python_package` is the name of a Python package and not a path to the
 code. This also means that the package has to be in your `PYTHONPATH` (see below).
 
+
+{{pydantic_model('nomad.config.plugins.Schema', heading='### Parser plugin metadata')}}
+{{pydantic_model('nomad.config.plugins.Parser', heading='### Schema plugin metadata')}}
+
 Now follow the instructions for one of our examples and try for yourself:
 
 - [schema plugin](https://github.com/nomad-coe/nomad-schema-plugin-example)
 - [parser plugin](https://github.com/nomad-coe/nomad-parser-plugin-example)
 
-## Add a plugin to your NOMAD (Oasis)
+## Add a plugin to your NOMAD
 
-To add a plugin, you need to add the *plugin metadata* to `nomad.yaml` and you need
+To add a plugin, you need to add the *plugin metadata* to `nomad.yaml` (see above) and you need
 to add the *plugin code* to the `PYTHONPATH` of your NOMAD. The `nomad.yaml` needs to be
 edited manually in the usual way. There are several ways to
-add *plugin code* to a NOMAD installation:
+add *plugin code* to a NOMAD installation.
 
-- adding it to the `PYTHONPATH`, if you run nomad directly, e.g. as a developer
-- mounting plugin sources into NOMAD (Oasis) containers, e.g. as an Oasis admin (coming soon...)
+### Development setup of NOMAD
+
+Simply add the plugin directory to the `PYTHONPATH` environment variable. When you start
+the application (e.g. `nomad admin run appworker`), Python will find your code when NOMAD
+imports the `python_package` given in the `plugins.options` of your `nomad.yaml`.
+
+### NOMAD Oasis
+
+The NOMAD docker image adds the folder `/app/plugins` to the `PYTHONPATH`. You simply have
+to add the *plugin metadata* to your Oasis' `nomad.yaml` and mount your code into the `/app/plugins`
[…]
