### 2023-04-19 fa1de1d
Changed the behavior of the row mode in tabular parser
Changelog: Fixed
