### 2023-04-20 3b33241
Resolve "adding coauthors via nomad.yaml files does not work in certain Oases"
