### 2023-04-20 #5388 @editorialbot
[REVIEW]: NOMAD: A distributed web-based platform for managing materials science research data
**Submitting author:** @markus1978 (<a href="http://orcid.org/0000-0002-8038-2277">Markus Scheidgen</a>)
**Repository:** https://github.com/nomad-coe/nomad
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.1
