### 2023-04-20 c0fef89
Resolve "Extend text parser"
Implement functionality to write file parser results to archive. 

Changelog: Added
