### 2023-04-22 1cb4761
Fixed file table in parser about dialog.
The table was not correctly rendered due to a missing
markdown plugin.

Changelog: Fixed
