### 2023-04-22 0e61c9d
Updated plugin documentation and installation.
