### 2023-04-22 94b9e18
Resolve "Move reference and schema related client responsibilities into the API"
