### 2023-04-24 ed38069
Fix phonopy and lobster parser
Fix band path generation in phonopy
Add support for compressed files in lobster

Changelog: Fixed
