### 2023-04-25 80fcc8b
Added new metainfo section for band gaps Changelog: Added
