+++ nomad/normalizing/results.py
@@ -39,7 +39,8 @@ from nomad.normalizing.common import (
     structures_2d
 )
 from nomad.datamodel.results import (
-    BandGapElectronic,
+    BandGap,
+    BandGapDeprecated,
     RadialDistributionFunction,
     RadiusOfGyration,
     MeanSquaredDisplacement,
@@ -49,7 +50,7 @@ from nomad.datamodel.results import (
     GeometryOptimization,
     Trajectory,
     MolecularDynamics,
-    Methodology,
+    MDProvenance,
     TemperatureDynamic,
     VolumeDynamic,
     PressureDynamic,
@@ -223,6 +224,26 @@ class ResultsNormalizer(Normalizer):
                     i_species.chemical_symbols = [symbol]
                 i_species.concentration = [1.0]
 
+    def resolve_band_gap(self, path) -> Union[List[BandGap], None]:
+        """
+        Extract all band gaps from the given path
+        and return them in a list along with their provenance.
+        """
+        band_gaps = traverse_reversed(self.entry_archive, path)
+        if not band_gaps:
+            return None
+        bg_root: List[BandGap] = []
+        for bg in band_gaps:
+            bg_new = BandGap()
+            bg_new.index = bg.index
+            bg_new.value = bg.value
+            bg_new.type = bg.type
+            bg_new.energy_highest_occupied = bg.energy_highest_occupied
+            bg_new.energy_lowest_unoccupied = bg.energy_lowest_unoccupied
+            bg_new.provenance = bg.provenance
+            bg_root.insert(0, bg_new)
+        return bg_root
+
     def band_structure_electronic(self) -> Union[List[BandStructureElectronic], None]:
         """Returns a new section containing an electronic band structure. In
         the case of multiple valid band structures, only the latest one is
@@ -250,13 +271,10 @@ class ResultsNormalizer(Normalizer):
                     bs_new.segment = bs.segment
                     bs_new.spin_polarized = bs_new.segment[0].energies.shape[0] > 1
                     bs_new.energy_fermi = bs.energy_fermi
+
                     for info in bs.band_gap:
-                        info_new = bs_new.m_create(BandGapElectronic)
-                        info_new.index = info.index
-                        info_new.value = info.value
-                        info_new.type = info.type
-                        info_new.energy_highest_occupied = info.energy_highest_occupied
-                        info_new.energy_lowest_unoccupied = info.energy_lowest_unoccupied
+                        info_new = BandGapDeprecated().m_from_dict(info.m_to_dict())
+                        bs_new.m_add_sub_section(BandStructureElectronic.band_gap, info_new)
                     return bs_new
             return None
 
@@ -303,10 +321,8 @@ class ResultsNormalizer(Normalizer):
                     dos_new.spin_polarized = n_channels > 1
                     dos_new.energy_fermi = dos.energy_fermi
                     for info in dos.band_gap:
-                        info_new = dos_new.m_create(BandGapElectronic)
-                        info_new.index = info.index
-                        info_new.energy_highest_occupied = info.energy_highest_occupied
-                        info_new.energy_lowest_unoccupied = info.energy_lowest_unoccupied
+                        info_new = BandGapDeprecated().m_from_dict(info.m_to_dict())
+                        dos_new.m_add_sub_section(DOSElectronic.band_gap, info_new)
                     return dos_new
             return None
 
@@ -483,8 +499,8 @@ class ResultsNormalizer(Normalizer):
 
         return None
 
-    def get_md_methodology(self, workflow: Workflow) -> Optional[MolecularDynamics]:
-        """Retrieves the MD methodology from the given workflow.
[…]
