+++ gui/src/components/visualization/GreensFunctions.spec.js
@@ -25,8 +25,8 @@ test.each([
   ['loading', VisualizationState.Loading, undefined, undefined, undefined, 'greens-functions-regtau', 'greens-functions-imsiw'],
   ['error: invalid data layout', VisualizationState.Error, {invalid: 'data'}, undefined, undefined, undefined, undefined],
   ['valid', VisualizationState.Success, {tau: [0, 1, 2], regtau: [[[[1, 0, 1]], [[1, 0, 1]]]], iw: [-1, 0, 1], imsiw: [[[[1, 0, 1]], [[1, 0, 1]]]]}, {magnetic_state: 'paramagnetic'}, undefined, 'greens-functions-regtau', 'greens-functions-imsiw']
-])('greens functions: %s', async (id, state, data, methodology, classes, regtauTestID, imsiwTestID) => {
-  render(<GreensFunctions data={data} methodology={methodology} classes={classes} />)
+])('greens functions: %s', async (id, state, data, provenance, classes, regtauTestID, imsiwTestID) => {
+  render(<GreensFunctions data={data} provenance={provenance} classes={classes} />)
   await expectPlot(state, regtauTestID, gfError)
   await expectPlot(state, imsiwTestID, gfError)
 })
