### 2023-04-27 95349c3
GW and BSE method normalizer
