+++ nomad/normalizing/method.py
@@ -22,6 +22,8 @@ import re
 import numpy as np
 from typing import List, Tuple, Union
 
+from nomad.datamodel import EntryArchive
+from nomad.metainfo import MSection
 from nomad.units import ureg
 from nomad.metainfo import Section
 from nomad.metainfo.util import MTypes
@@ -29,56 +31,45 @@ from nomad.utils import RestrictedDict
 from nomad import config
 from nomad.datamodel.metainfo.simulation.method import KMesh
 from nomad.datamodel.results import (
-    Method,
-    Electronic,
-    Simulation,
-    HubbardKanamoriModel,
-    DFT,
-    Projection,
-    GW,
-    DMFT,
-    Precision,
-    xc_treatments,
-)
+    Method, Electronic, Simulation, HubbardKanamoriModel, DFT, Projection, GW, BSE, DMFT,
+    Precision, xc_treatments, Material)
 
 
 class MethodNormalizer():
-    def __init__(self, entry_archive, repr_system, material, logger):
+    def __init__(self, entry_archive: EntryArchive, repr_system: MSection, material: Material, logger):
         self.entry_archive = entry_archive
         self.repr_system = repr_system
         self.repr_method = None
         self.material = material
-        self.method_name = None
+        self.method_name = config.services.unavailable_value
         self.run = self.entry_archive.run[0]
         self.logger = logger
 
     def method(self) -> Method:
         '''Returns the populated results.method section.'''
         method = Method()
-        simulation = Simulation()
         repr_method = None
         method_name = config.services.unavailable_value
-        workflow_name = None
         methods = self.run.method
         n_methods = len(methods)
+        functional_long_name = ''
+        settings_basis_set = RestrictedDict(mandatory_keys=[None], optional_keys=[None], forbidden_values=[None])
+        gs_method = None
+        xs_method = None
+        method_def = {value.sub_section.name: value for _, value in Simulation.m_def.all_sub_sections.items()}
 
         def get_method_name(section_method):
             method_name = config.services.unavailable_value
-            if section_method.m_xpath('electronic') and section_method.electronic.m_xpath('method'):
+            if section_method.electronic and section_method.electronic.method:
                 method_name = section_method.electronic.method
-            elif section_method.m_xpath('gw') is not None:
-                method_name = 'GW'
-            elif section_method.m_xpath('projection') is not None:
-                method_name = 'Projection'
-            elif section_method.m_xpath('dmft') is not None:
-                method_name = 'DMFT'
-            elif section_method.m_xpath('core_hole') is not None:
-                method_name = 'CoreHole'
-            elif section_method.m_xpath('bse') is not None:
-                method_name = 'BSE'
+            else:
+                for method in ['gw', 'projection', 'dmft', 'core_hole', 'bse']:
+                    if section_method.m_xpath(method):
+                        method_name = getattr(section_method, method).m_def.name
+                        break
             return method_name
 
-        def functional_long_name_from_method(methods):
+        def functional_long_name_from_method():
             '''
             Long form of exchange-correlation functional, list of components and parameters
             as a string: see https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-meta-info/wikis/metainfo/XC-functional
@@ -140,26 +131,30 @@ class MethodNormalizer():
             return settings.to_dict()
 
         # workflow_name
-        if self.entry_archive.workflow:
-            workflow_name = list(filter(lambda x: x, map(lambda x: x.type, self.entry_archive.workflow)))[0]
-        method.workflow_name = workflow_name
-
-        # repr_method and method_name
-        #   If workflow_name is 'GW', set repr_method as the corresponding method_ref
-        if method.workflow_name and method.workflow_name == 'GW':
+        method.workflow_name = self.entry_archive.workflow2.name
[…]
