### 2023-04-28 08d4615
Fixed javascript error on north page.
The staging mechanism introduced a regression that caused
the render of north tools to fail.

Changelog: Fixed
