### 2023-04-28 0f65158
Added new metainfo for storing the level of the XC functional.
Changelog: Added
