### 2023-04-28 94d8412
Added config option to fallback to a different archive version.
This allows to configure multiple archive version for an installation
allowing to share an archive from another installation. All
write operations go to one version. Read operations go through
a list of versions to see what is available.

Changelog: Added
