+++ nomad/files.py
@@ -44,7 +44,7 @@ original mainfile, and vice versa.
 '''
 
 from abc import ABCMeta
-from typing import IO, Set, Dict, Iterable, Iterator, List, Tuple, Any, NamedTuple
+from typing import IO, Set, Dict, Iterable, Iterator, List, Tuple, Any, NamedTuple, Callable
 from functools import lru_cache
 from pydantic import BaseModel
 from datetime import datetime
@@ -531,6 +531,37 @@ def create_zipstream(
     return iter(zip_stream)
 
 
+def _versioned_archive_file_object(
+    target_dir: DirectoryObject, file_name: Callable[[str], str], fallback: bool
+) -> PathObject:
+    '''
+    Creates a file object for an archive file depending on the directory it is or
+    will be created in, the recipe to construct the name from a version suffix, and
+    a bool that denotes if alternative version suffixes should be considered.
+    '''
+    version_suffixes = config.fs.archive_version_suffix
+
+    if not isinstance(version_suffixes, list):
+        version_suffixes = [version_suffixes]
+
+    if len(version_suffixes) <= 1:
+        version_suffix = version_suffixes[0]
+        if not version_suffix or version_suffix == '':
+            return target_dir.join_file(file_name(''))
+        else:
+            return target_dir.join_file(file_name(f'-{version_suffix}'))
+
+    if not fallback:
+        return target_dir.join_file(file_name(f'-{version_suffixes[0]}'))
+
+    for version_suffix in version_suffixes:
+        current_file = target_dir.join_file(file_name(f'-{version_suffix}'))
+        if os.path.exists(current_file.os_path):
+            return current_file
+
+    return target_dir.join_file(file_name(f'-{version_suffixes[0]}'))
+
+
 class UploadFiles(DirectoryObject, metaclass=ABCMeta):
     ''' Abstract base class for upload files. '''
     def __init__(self, upload_id: str, create: bool = False):
@@ -794,7 +825,7 @@ class StagingUploadFiles(UploadFiles):
 
     def write_archive(self, entry_id: str, data: Any) -> int:
         ''' Writes the data as archive file and returns the archive file size. '''
-        archive_file_object = self.archive_file_object(entry_id)
+        archive_file_object = self._archive_file_object(entry_id)
         try:
             write_archive(archive_file_object.os_path, 1, data=[(entry_id, data)])
         except Exception as e:
@@ -804,18 +835,20 @@ class StagingUploadFiles(UploadFiles):
 
             raise e
 
-        return self.archive_file_object(entry_id).size
+        return archive_file_object.size
 
     def read_archive(self, entry_id: str, use_blocked_toc: bool = True) -> ArchiveReader:
         try:
-            return read_archive(self.archive_file_object(entry_id).os_path, use_blocked_toc=use_blocked_toc)
+            return read_archive(self._archive_file_object(entry_id, True).os_path, use_blocked_toc=use_blocked_toc)
 
         except FileNotFoundError:
             raise KeyError(entry_id)
 
-    def archive_file_object(self, entry_id: str) -> PathObject:
-        version_suffix = '-' + config.fs.archive_version_suffix if config.fs.archive_version_suffix else ''
-        return self._archive_dir.join_file(f'{entry_id}{version_suffix}.msg')
+    def _archive_file_object(self, entry_id: str, fallback: bool = False) -> PathObject:
+        def versioned_file_name(version_suffix):
+            return f'{entry_id}{version_suffix}.msg'
+
+        return _versioned_archive_file_object(self._archive_dir, versioned_file_name, fallback=fallback)
 
     def add_rawfiles(
             self, path: str, target_dir: str = '', cleanup_source_file_and_dir: bool = False,
@@ -1072,7 +1105,7 @@ class StagingUploadFiles(UploadFiles):
 
[…]
