### 2023-04-29 d0dfdd0
Use modern stmp impl
