+++ pyproject.toml
@@ -129,6 +129,7 @@ dev = [
     'mkdocs-material==8.1.1',
     'mkdocs-material-extensions==1.0.3',
     'mkdocs-macros-plugin==0.6.3',
+    'aiosmtpd'
 ]
 
 [project.scripts]
