### 2023-05-02 08ef593
Added automatic subsystem detection.
Changelog: Added
