+++ nomad/normalizing/topology.py
@@ -18,6 +18,8 @@
 
 from typing import Dict, List, Optional
 from collections import defaultdict
+import pathlib
+import json
 
 from ase import Atoms
 from ase.data import chemical_symbols
@@ -25,6 +27,8 @@ import numpy as np
 from matid.clustering import Clusterer, Cluster, Classification  # pylint: disable=import-error
 from matid.symmetry.symmetryanalyzer import SymmetryAnalyzer  # pylint: disable=import-error
 
+from nomad import utils
+from nomad import config
 from nomad import atomutils
 from nomad.atomutils import Formula
 from nomad.datamodel.results import Symmetry, Material, System, Relation, Prototype
@@ -39,6 +43,8 @@ from nomad.normalizing.common import (
 )
 
 SYMBOLS = np.array(chemical_symbols)
+with open(pathlib.Path(__file__).parent / 'data/top_50k_material_ids.json', "r") as fin:
+    TOP_50K_MATERIAL_IDS = json.load(fin)
 
 
 def get_topology_id(index: int) -> str:
@@ -143,7 +149,8 @@ class TopologyNormalizer():
         topology_calc = self.topology_calculation(material)
         if topology_calc:
             return topology_calc
-        topology_matid = self.topology_matid(material)
+        with utils.timer(self.logger, 'calculating topology with matid'):
+            topology_matid = self.topology_matid(material)
         if topology_matid:
             return topology_matid
 
@@ -225,14 +232,14 @@ class TopologyNormalizer():
         return list(topology.values())
 
     def topology_matid(self, material: Material) -> Optional[List[System]]:
-        # Topology is currently created only for systems that are classified as 2D
-        # or surface. TODO: Also other entries should get a topology: all of
-        # the symmetry analysis and additional system partitioning through MatID
-        # should be stored in the topology.
         '''
         Returns a list of the identified systems with topological relations and
-        classification of subsystems
+        classification of subsystems.
         '''
+        # TODO: Currently we only try to analyze 2D, surface and unavailable.
+        # This makes sure that e.g. most bulk DFT calculations are untouched.
+        # Later we should process them similarly once we know what kind of
+        # subsystems they get and we fully replace results.material with a list.
         if material.structural_type not in {'2D', 'surface', 'unavailable'}:
             return None
         try:
@@ -240,15 +247,13 @@ class TopologyNormalizer():
             atoms = ase_atoms_from_nomad_atoms(nomad_atoms)
         except Exception:
             return None
-        # TODO: MatID does not currently support non-periodic structures or
-        # structures with zero-sized cell
+        # TODO: Ignore systems with zero-sized cell
         cell = atoms.get_cell()
-        if not any(atoms.pbc) or cell is None or cell.volume == 0:
+        if cell.volume == 0:
             return None
-        # TODO: The matid processing is currently completely skipped. Should be
-        # enabled after more testing.
+        # In order to limit the processing time, a maximum system size is checked.
         n_atoms = len(atoms)
-        if n_atoms > 0:
+        if n_atoms > config.normalize.clustering_size_limit:
             return None
 
         topology: Dict[str, System] = {}
@@ -257,22 +262,42 @@ class TopologyNormalizer():
         add_system_info(original, topology)
 
         # Add all meaningful clusters to the topology
-        clusters = self._perform_matid_clustering(atoms)
-        filtered_clusters = self._filter_clusters(clusters)
-        for cluster in filtered_clusters:
+        clusterer = Clusterer()
+        clusters = clusterer.get_clusters(atoms, pos_tol=0.8)
+        for cluster in clusters:
             subsystem = self._create_subsystem(cluster)
             structural_type = subsystem.structural_type
-            # Currently only 2D and surface subsystems are included
-            if structural_type in {'surface', '2D'}:
[…]
