### 2023-05-03 3b97832
Added config switch for generating new timestamps.
Changelog: Added
