### 2023-05-03 fadb038
Fix RichText editor style in non editable mode
Changelog: Fixed
