+++ gui/src/components/entry/properties/WorkflowCard.js
@@ -17,948 +17,1217 @@
  */
 import React, { useEffect, useMemo, useRef, useState } from 'react'
 import PropTypes from 'prop-types'
-import { makeStyles, Tooltip, IconButton } from '@material-ui/core'
+import * as d3 from 'd3'
+import { makeStyles, Tooltip, IconButton, TextField, FormControl } from '@material-ui/core'
 import grey from '@material-ui/core/colors/grey'
+import blueGrey from '@material-ui/core/colors/blueGrey'
 import red from '@material-ui/core/colors/red'
+import { Replay, Undo, Label, LabelOff, PlayArrowSharp, StopSharp, Clear } from '@material-ui/icons'
+import { useHistory } from 'react-router-dom'
+import { isPlainObject } from 'lodash'
 import { PropertyCard, PropertyGrid } from './PropertyCard'
 import { resolveNomadUrl, resolveInternalRef, createEntryUrl } from '../../../utils'
 import { useApi } from '../../api'
-import { useHistory } from 'react-router-dom'
-import * as d3 from 'd3'
 import { getUrl } from '../../nav/Routes'
-import { nomadFontFamily, nomadPrimaryColor, nomadSecondaryColor, apiBase } from '../../../config'
-import { Replay, Undo, Redo } from '@material-ui/icons'
+import { nomadFontFamily, apiBase } from '../../../config'
+import { useAsyncError } from '../../../hooks'
 
 const useWorkflowGraphStyles = makeStyles(theme => ({
   root: {
     width: '100%',
     minWidth: 1000,
     '& .link': {
-      strokeWidth: 3,
       fill: 'none',
-      stroke: grey[500],
-      strokeOpacity: 0.5
-    },
-    '& .crosslink': {
-      fill: 'none',
-      stroke: red[500],
       strokeOpacity: 0.5,
-      strokeWidth: 3
+      fillOpacity: 0.5,
+      strokeWidth: 2.5
     },
     '& .text': {
       fontFamily: nomadFontFamily,
       fontSize: 12,
-      fontWeight: 'bold',
       dy: '0.35em',
-      fill: grey[800]
+      textAnchor: 'middle'
     },
-    '& .circle': {
-      stroke: grey[500],
+    '& .icon': {
       strokeWidth: 2,
       strokeOpacity: 0.5
     }
   }
 }))
 
-function addMarkers(svg, nodes, offset) {
-  const markerWidth = 5
+let archives = {}
 
-  const markerIds = []
+const resolveSection = async (source, query) => {
+  if (isPlainObject(source.section)) {
+    // already resolved
+    return source
+  }
 
-  const addIds = (node) => {
-    const nodeLinks = node.data.crossLinks || []
-    nodeLinks.forEach(link => {
-      const nodeTarget = nodes.filter(node => node.data.key === link[0])
-      nodeTarget.forEach(target => {
-        markerIds.push(`${node.id}-${target.id}`)
-      })
-    })
-    if (node._children) {
-      node._children.forEach(child => addIds(child))
+  let path = source.path
+  if (typeof path !== 'string') path = source.section || '/'
+  const pathSegments = path.split('#').filter(p => p)
+
+  let archive = source.archive
+  let baseUrl = createEntryUrl(apiBase, archive?.metadata?.upload_id, archive?.metadata?.entry_id)
+  let section
+  if (pathSegments.length === 1) {
+    // internal reference
+    path = pathSegments[0]
+    try {
+      section = source.section ? source.path : resolveInternalRef(path, archive)
+    } catch (error) {
+      console.error(`Cannot resolve section ${path}`)
+      return
+    }
+  } else {
+    // external reference
+    const url = resolveNomadUrl(path, baseUrl)
[…]
