### 2023-05-04 398f5ab
Lowered the worker replicas to free resources for reprocessing.
