### 2023-05-05 c59ceb8
Include new BandGap in GW
