### 2023-05-05 dbef478
Resolve "Issue in method normalizer for CoreHole"
