### 2023-05-05 2c65189
Removed invalid structural_type option.
