### 2023-05-08 45e9512
Added new metainfo for dimensionality and building_block
Changelog: Added
