+++ gui/tests/artifacts.js
@@ -1002,6 +1002,37 @@ window.nomadArtifacts = {
       "aggregatable": true,
       "suggestion": true
     },
+    "results.material.dimensionality": {
+      "name": "dimensionality",
+      "description": "Dimensionality of the system. For atomistic systems this is\nautomatically evaluated by using the topology-scaling algorithm:\nhttps://doi.org/10.1103/PhysRevLett.118.106101.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'0D'` | Not connected periodically |\n| `'1D'` | Periodically connected in one dimension |\n| `'2D'` | Periodically connected in two dimensions |\n| `'3D'` | Periodically connected in three dimensions |",
+      "type": {
+        "type_kind": "Enum",
+        "type_data": [
+          "0D",
+          "1D",
+          "2D",
+          "3D"
+        ]
+      },
+      "aggregatable": true,
+      "suggestion": true
+    },
+    "results.material.building_block": {
+      "name": "building_block",
+      "description": "More exact classification for this system, i.e. the type of \"building\nblock\" it represents.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'surface'` | Structure built from a unit cell that repeats periodically in two directions and at least twice, but not infinitely in a third direction. |\n| `'2D material'` | Structure built from a unit cell that repeats periodically in two directions and only once in a third direction. |\n| `'molecule'` | Molecule defined in the force-field topology |\n| `'monomer'` | Monomer defined in the force-field topology |\n| `'group'` | Generic group |",
+      "type": {
+        "type_kind": "Enum",
+        "type_data": [
+          "surface",
+          "2D material",
+          "molecule",
+          "monomer",
+          "group"
+        ]
+      },
+      "aggregatable": true,
+      "suggestion": true
+    },
     "results.material.functional_type": {
       "name": "functional_type",
       "description": "Classification based on the functional properties.",
@@ -1412,7 +1443,8 @@ window.nomadArtifacts = {
         "type_kind": "python",
         "type_data": "str"
       },
-      "aggregatable": true
+      "aggregatable": true,
+      "suggestion": true
     },
     "results.material.topology.method": {
       "name": "method",
@@ -1425,7 +1457,8 @@ window.nomadArtifacts = {
           "matid"
         ]
       },
-      "aggregatable": true
+      "aggregatable": true,
+      "suggestion": true
     },
     "results.material.topology.description": {
       "name": "description",
@@ -1472,6 +1505,37 @@ window.nomadArtifacts = {
           "group",
           "molecule",
           "monomer"
+        ]
+      },
+      "aggregatable": true,
+      "suggestion": true
+    },
+    "results.material.topology.dimensionality": {
+      "name": "dimensionality",
+      "description": "Dimensionality of the system. For atomistic systems this is\nautomatically evaluated by using the topology-scaling algorithm:\nhttps://doi.org/10.1103/PhysRevLett.118.106101.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'0D'` | Not connected periodically |\n| `'1D'` | Periodically connected in one dimension |\n| `'2D'` | Periodically connected in two dimensions |\n| `'3D'` | Periodically connected in three dimensions |",
+      "type": {
+        "type_kind": "Enum",
+        "type_data": [
+          "0D",
+          "1D",
+          "2D",
+          "3D"
+        ]
+      },
+      "aggregatable": true,
+      "suggestion": true
+    },
+    "results.material.topology.building_block": {
+      "name": "building_block",
[…]

[…]
