### 2023-05-08 22f5dc1
Added ELN base section for reading and storing structure files.
Changelog: Added
