### 2023-05-09 065afe4
Fix cp2k, abinit and phonopy parsers
Implemented various parser fixes as a result of reprocessing.

Changelog: Fixed
