+++ gui/tests/artifacts.js
@@ -1019,15 +1019,14 @@ window.nomadArtifacts = {
     },
     "results.material.building_block": {
       "name": "building_block",
-      "description": "More exact classification for this system, i.e. the type of \"building\nblock\" it represents.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'surface'` | Structure built from a unit cell that repeats periodically in two directions and at least twice, but not infinitely in a third direction. |\n| `'2D material'` | Structure built from a unit cell that repeats periodically in two directions and only once in a third direction. |\n| `'molecule'` | Molecule defined in the force-field topology |\n| `'monomer'` | Monomer defined in the force-field topology |\n| `'group'` | Generic group |",
+      "description": "More exact classification for this system, i.e. the type of \"building\nblock\" it represents.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'surface'` | Structure built from a unit cell that repeats periodically in two directions and at least twice, but not infinitely in a third direction. |\n| `'2D material'` | Structure built from a unit cell that repeats periodically in two directions and only once in a third direction. |\n| `'molecule'` | Molecule defined in the force-field topology |\n| `'monomer'` | Monomer defined in the force-field topology |",
       "type": {
         "type_kind": "Enum",
         "type_data": [
           "surface",
           "2D material",
           "molecule",
-          "monomer",
-          "group"
+          "monomer"
         ]
       },
       "aggregatable": true,
@@ -1527,15 +1526,14 @@ window.nomadArtifacts = {
     },
     "results.material.topology.building_block": {
       "name": "building_block",
-      "description": "More exact classification for this system, i.e. the type of \"building\nblock\" it represents.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'surface'` | Structure built from a unit cell that repeats periodically in two directions and at least twice, but not infinitely in a third direction. |\n| `'2D material'` | Structure built from a unit cell that repeats periodically in two directions and only once in a third direction. |\n| `'molecule'` | Molecule defined in the force-field topology |\n| `'monomer'` | Monomer defined in the force-field topology |\n| `'group'` | Generic group |",
+      "description": "More exact classification for this system, i.e. the type of \"building\nblock\" it represents.\n\n| Value | Description |\n| --------- | ----------------------- |\n| `'surface'` | Structure built from a unit cell that repeats periodically in two directions and at least twice, but not infinitely in a third direction. |\n| `'2D material'` | Structure built from a unit cell that repeats periodically in two directions and only once in a third direction. |\n| `'molecule'` | Molecule defined in the force-field topology |\n| `'monomer'` | Monomer defined in the force-field topology |",
       "type": {
         "type_kind": "Enum",
         "type_data": [
           "surface",
           "2D material",
           "molecule",
-          "monomer",
-          "group"
+          "monomer"
         ]
       },
       "aggregatable": true,
@@ -1801,6 +1799,24 @@ window.nomadArtifacts = {
       ],
       "aggregatable": true
     },
+    "results.material.topology.atomic_fraction": {
+      "name": "atomic_fraction",
+      "description": "The atomic fraction of this system in the full structure it is contained in.\nPer definition a positive value less than or equal to 1.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "aggregatable": false
+    },
+    "results.material.topology.mass_fraction": {
+      "name": "mass_fraction",
+      "description": "The mass fraction of this system in the full structure it is contained within.\nPer definition a positive value less than or equal to 1.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "aggregatable": false
+    },
     "results.material.topology.n_atoms": {
       "name": "n_atoms",
       "description": "The total number of species (atoms, particles) in the system.",
[…]
