### 2023-05-09 15189b2
Wrapped formula creation in try/except in order to not completely halt parsing...
