+++ docs/learn/oasis.md
@@ -1,0 +1,19 @@
+The software that runs NOMAD is Open-Source and can be used independently of the NOMAD
+*central installation* at [http://nomad-lab.eu](http://nomad-lab.eu).
+We call any NOMAD installation that is not the *central* one a NOMAD Oasis.
+
+<figure markdown>
+  ![oasis use-cases](oasis-use-cases.png){ width=700 }
+  <figcaption>NOMAD Oasis use-cases</figcaption>
+</figure>
+
+There are several use-cases how the NOMAD software could be used. Of course other
+uses and hybrids are imaginable:
+
+- Academia: Use the Oasis for local management of unpublished research data
+- Mirror: Use the Oasis as a mirror that hosts a copy of all published NOMAD data
+- Industry: Use of Oasis to manage private data and full internal use of published data in compliance with strict privacy policies
+- FAIRmat: Use Oasis to form a network of repositories to build a federated data infrastructure
+for materials science.
+This is what we do in the [FAIRmat project](https://www.fair-di.eu/fairmat/consortium).
+
