+++ docs/reference/cli.md
@@ -1,0 +1,13 @@
+---
+toc_depth: 6
+---
+
+# Command Line Interface (CLI)
+
+::: mkdocs-click
+    :module: nomad.cli.cli
+    :command: cli
+    :prog_name: nomad
+    :depth: 1
+    :list_subcommands: True
+

[…]
