+++ nomad/archive/query_reader.py
@@ -21,14 +21,16 @@ import copy
 import dataclasses
 import functools
 import re
-from typing import List, Any, Optional, Set, Dict, Tuple, Callable, Type
+from typing import Any, Callable, Type
 
 import orjson
 from fastapi import HTTPException
 from mongoengine import Q
 
 from nomad import utils
-from nomad.app.v1.models import Direction, WithQuery, MetadataPagination, Pagination, PaginationResponse
+from nomad.app.v1.models import (
+    Direction, WithQuery, MetadataPagination, Pagination, PaginationResponse, MetadataRequired
+)
 from nomad.app.v1.routers.datasets import DatasetPagination
 from nomad.app.v1.routers.entries import perform_search
 from nomad.app.v1.routers.uploads import (
@@ -99,13 +101,13 @@ class ConfigError(Exception):
 class ArchiveNode:
     upload_id: str  # the upload id of the current node
     entry_id: str  # the entry id of the current node
-    current_path: List[str]  # the path in the result container
+    current_path: list[str]  # the path in the result container
     result_root: dict  # the root of the result container
     ref_result_root: dict  # collection of referenced archives that should never change
     archive: Any  # current node in the archive
-    archive_root: Optional[ArchiveDict | dict]  # the root of the archive
+    archive_root: None | ArchiveDict | dict  # the root of the archive
     definition: Any  # definition of the current node
-    visited_path: Set[str]  # visited paths, for tracking circular references
+    visited_path: set[str]  # visited paths, for tracking circular references
     current_depth: int  # current depth, for tracking depth limit
     reader: Any  # the reader used to read the archive # pylint: disable=E0601
 
@@ -144,8 +146,7 @@ class ArchiveNode:
 
         path_stack: list = [v for v in reference_copy.split('/') if v]
 
-        reference_url: str = self.generate_reference(path_stack)
-        if reference_url in self.visited_path:
+        if (reference_url := self.generate_reference(path_stack)) in self.visited_path:
             raise ArchiveError(f'Circular reference detected: {reference_url}.')
 
         try:
@@ -163,8 +164,7 @@ class ArchiveNode:
         Go to a remote archive, which can be either in the same server or another installation.
         '''
         # this is a global reference, get the target archive first
-        parse_result = parse_path(reference, self.upload_id)
-        if parse_result is None:
+        if (parse_result := parse_path(reference, self.upload_id)) is None:
             raise ArchiveError(f'Invalid reference: {reference}.')
 
         installation, other_upload_id, id_or_file, kind, path = parse_result
@@ -189,8 +189,7 @@ class ArchiveNode:
         path_stack: list = [v for v in path.split('/') if v]
 
         other_prefix: str = f'../uploads/{other_upload_id}/archive/{other_entry_id}'
-        reference_url: str = f'{other_prefix}#/{"/".join(str(v) for v in path_stack)}'
-        if reference_url in self.visited_path:
+        if (reference_url := f'{other_prefix}#/{"/".join(str(v) for v in path_stack)}') in self.visited_path:
             raise ArchiveError(f'Circular reference detected: {reference_url}.')
 
         # get the archive
@@ -334,7 +333,7 @@ def _populate_result(container_root: dict, path: list, value):
 
     path_stack: list = list(reversed(path))
     target_container: dict | list = container_root
-    key_or_index: Optional[str | int] = None
+    key_or_index: None | str | int = None
 
     while len(path_stack) > 0:
         if key_or_index is not None:
@@ -371,7 +370,7 @@ def _populate_result(container_root: dict, path: list, value):
 
 
 @functools.lru_cache(maxsize=1024)
-def _parse_key(key: Optional[str]) -> tuple:
[…]
