+++ requirements.txt
@@ -167,7 +167,7 @@ orjson==3.6.0             # via h5grove, nomad-lab (pyproject.toml)
 packaging==23.0           # via dask, docker, hyperspy, ipykernel, jupyter-server, jupyterlab, jupyterlab-server, matplotlib, mdanalysis, mongomock, nbconvert, pint, scikit-image, sphinx
 palettable==3.3.0         # via pymatgen
 pamela==1.0.0             # via jupyterhub
-pandas==1.3.5             # via eelsdbconverter (dependencies/parsers/eelsdb/pyproject.toml), ifes-apt-tc-data-modeling, nexusutils (dependencies/parsers/nexus/pyproject.toml), panedr, pymatgen, xarray
+pandas==1.3.5             # via eelsdbconverter (dependencies/parsers/eelsdb/pyproject.toml), ifes-apt-tc-data-modeling, nexusutils (dependencies/parsers/nexus/pyproject.toml), nomad-lab (pyproject.toml), panedr, pymatgen, xarray
 pandocfilters==1.5.0      # via nbconvert
 panedr==0.2               # via atomisticparsers (dependencies/parsers/atomistic/pyproject.toml), nomad-lab (pyproject.toml)
 parmed==3.0.0             # via nomad-lab (pyproject.toml)
