### 2023-05-11 f42b7ca
Added basic information, such as mass fraction and atomic fraction, of each element in results.
Changelog: Added
