+++ nomad/datamodel/metainfo/eln/__init__.py
@@ -25,6 +25,7 @@ from nomad.units import ureg
 from nomad.datamodel.data import EntryData, ArchiveSection, author_reference, BasicElnCategory
 from nomad.metainfo.metainfo import MSection, MProxy, MEnum, Category, MCategory
 from nomad.datamodel.results import ELN, Results, Material
+from nomad.datamodel.results import ElementalComposition as ResultsElementalComposition
 from nomad.metainfo import Package, Quantity, Datetime, Reference, Section, SubSection
 from ase.data import chemical_symbols, atomic_numbers, atomic_masses
 from nomad.datamodel.metainfo.eln.perovskite_solar_cell_database import (
@@ -185,7 +186,7 @@ class Activity(ElnBaseSection):
             archive.results.eln.methods.append(self.m_def.name)
 
 
-class ElementalComposition(MSection):
+class ElementalComposition(ArchiveSection):
     '''A section for describing the elemental composition of a system, i.e. the element
     and its atomic fraction.
     '''
@@ -195,7 +196,7 @@ class ElementalComposition(MSection):
         description='''
         The symbol of the element, e.g. 'Pb'.
         ''',
-        a_eln=dict(component='EnumEditQuantity'))
+        a_eln=dict(component='AutocompleteEditQuantity'))
     atomic_fraction = Quantity(
         type=np.float64,
         description='''
@@ -204,6 +205,59 @@ class ElementalComposition(MSection):
         ''',
         a_eln=dict(component='NumberEditQuantity')
     )
+    mass_fraction = Quantity(
+        type=np.float64,
+        description='''
+        The mass fraction of the element in the system it is contained within.
+        Per definition a positive value less than or equal to 1.
+        ''',
+        a_eln=dict(component='NumberEditQuantity')
+    )
+
+    def normalize(self, archive, logger: Any) -> None:
+        '''
+        The normalizer for the `ElementalComposition` class. Will add a
+        results.material subsection if none exists. Will append the element to the
+        elements property of that subsection and a
+        nomad.datamodel.results.ElementalComposition instances to the
+        elemental_composition property  using the element and atomic fraction from this
+        section.
+
+        Args:
+            archive (EntryArchive): The archive containing the section that is being
+            normalized.
+            logger (Any): A structlog logger.
+        '''
+        super(ElementalComposition, self).normalize(archive, logger)
+
+        if self.element:
+            if not archive.results:
+                archive.results = Results()
+            if not archive.results.material:
+                archive.results.material = Material()
+
+            if self.element not in chemical_symbols:
+                logger.warn(
+                    f"'{self.element}' is not a valid element symbol and this "
+                    "elemental_composition section will be ignored.")
+            elif self.element not in archive.results.material.elements:
+                archive.results.material.elements += [self.element]
+            if self.atomic_fraction or self.mass_fraction:
+                comp_result_section = archive.results.material.elemental_composition
+                result_composition = ResultsElementalComposition(
+                    element=self.element,
+                    atomic_fraction=self.atomic_fraction,
+                    mass_fraction=self.mass_fraction,
+                    mass=atomic_masses[atomic_numbers[self.element]] * ureg.amu,
+                )
+                existing_elements = [comp.element for comp in comp_result_section]
+                if self.element in existing_elements:
+                    index = existing_elements.index(self.element)
+                    comp_result_section[index].atomic_fraction = self.atomic_fraction
+                    comp_result_section[index].mass_fraction = self.mass_fraction
+                    comp_result_section[index].mass = atomic_masses[atomic_numbers[self.element]] * ureg.amu
+                else:
+                    comp_result_section.append(result_composition)
 
 
 class System(Entity):
@@ -217,10 +271,68 @@ class System(Entity):
         section_def=ElementalComposition,
         repeats=True)
 
[…]
