### 2023-05-11 13a65f9
Added warning when deleting uploads or entries that are referenced
Changelog: Added
