### 2023-05-12 cb17002
remove Hall schema and modify settings.json
