+++ nomad/datamodel/metainfo/__init__.py
@@ -28,7 +28,6 @@ from .eln.material_library import m_package
 from .eln.perovskite_solar_cell_database import m_package
 from .downloads import m_package
 from .eln.labfolder import m_package
-from .eln.ikz_hall import m_package
 
 for plugin in config.plugins.filtered():
     if isinstance(plugin, Schema):
