+++ nomad/datamodel/metainfo/eln/ikz_hall/hall_instrument.py
@@ -1,419 +1,0 @@
-#
-# Copyright The NOMAD Authors.
-#
-# This file is part of NOMAD. See https://nomad-lab.eu for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-import numpy as np
-from nomad.metainfo import MSection, Quantity, SubSection
-
-
-class MeasurementStateMachine(MSection):
-    '''Representation of an instrument'''
-    use_instruments = Quantity(
-        type=bool,
-        description='Do you have instruments connected to the computer? True for measurement setups and false for office computers.')
-    system_model = Quantity(
-        type=str,
-        description='What is the Model number of your used Hall Measurement system? 0=[75XX-LVWR(-HS)] 1=[75XX-LVWR-SWT(-HS), 77XX-LVWR] 2=[75XX-HVWR(-HS), 77XX-HVWR] 3=[76XX] 4=[95XX-LVWR(-HS)] 5=[95XX-LVWR-SWT(-HS), 97XX-LVWR] 6=[95XX-HVWR(-HS), 97XX-HVWR] 7=[77XXA] 8=[97XXA]')
-    # switch_matrix = Quantity(
-    #    type=str,
-    #    description='What is the Model number of your used Hall matrix? 0=[Lake Shore 775] 1=[Lake Shore 776]')
-    wiring = Quantity(
-        type=str,
-        description='What is the wiring setup,either 0=[van der Pauw] or 1=[Hall Bar]?')
-    reconfirm_ammeter_removal = Quantity(
-        type=bool,
-        description='Is the safety question on for Amperemeter removal above 2mA?')
-    electrometer = Quantity(
-        type=str,
-        description='If there is a control electrometer installed, which one is it? 0=[Keithley 6512] 1=[Keithley 6514]')
-    control_electrometer = Quantity(
-        type=bool,
-        description='Is a control electrometer installed in the setup?')
-    numberofsamples = Quantity(
-        type=int,
-        description='How many samples are measured at the same time? Only applicable for certain Systems. Options: 1, 2 or 4')
-    voltmeter = Quantity(
-        type=str,
-        description='Which voltmeter is used in your setup? 0=[Keithley 2000] 1=[Keithley 2182] 2=[Keithley 182]')
-    currentmeter = Quantity(
-        type=str,
-        description='Which amperemeter is used in your setup? 0=[Keithley 6485] 1=[Keithley 485/6/7]')
-    current_source = Quantity(
-        type=str,
-        description='Which current source is used in your setup? 0=[Keithley 220] 1=[Keithley 6220]')
-    ac_hall = Quantity(
-        type=bool,
-        description='Does your setup use AC Hall measurements.')
-    ac_hall_type = Quantity(
-        type=str,
-        description='If your setup uses AC Hall measurements, does it use 0=[AC Current only] or 1=[AC Field and Current]?')
-    contact_blasting = Quantity(
-        type=bool,
-        description='Is the contact blasting option turned on? Only for Keithley 2400 at IEEE address 24.')
-
-
-class Keithley7001(MSection):
-    '''Representation of an instrument'''
-    resting_state = Quantity(
-        type=str,
[…]
