+++ nomad/datamodel/metainfo/eln/ikz_hall/nexus_to_msection.py
@@ -1,353 +1,0 @@
-#
-# Copyright The NOMAD Authors.
-#
-# This file is part of NOMAD. See https://nomad-lab.eu for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from typing import Generator, Tuple, Optional
-import re
-
-from .measurement import (
-    Measurement,
-    VariableTemperatureMeasurement,
-    VariableTemperatureData,
-    VariableFieldMeasurement,
-    VariableFieldData,
-    IVCurveMeasurement,
-    IVData
-)
-
-from .hall_instrument import (
-    Instrument,
-    Keithley7001,
-    Keithley6485,
-    Keithley182,
-    Keithley2000,
-    Keithley2182,
-    Keithley220,
-    Keithley2400,
-    Keithley2700,
-    FieldController,
-    TemperatureController,
-    TemperatureDomain
-)
-
-import nexusutils.dataconverter.readers.hall.helpers as hall_helpers  # pylint: disable=import-error
-from nomad.units import ureg
-
-
-def get_measurement_object(measurement_type: str) -> Measurement:
-    '''
-    Gets a measurement MSection object from the given measurement type.
-
-    Args:
-        measurement_type (str): The measurement type.
-
-    Returns:
-        Measurement: A MSection representing a Hall measurement.
-    '''
-    if measurement_type == 'Variable Temperature Measurement':
-        return VariableTemperatureMeasurement()
-    if measurement_type == 'Variable Field Measurement':
-        return VariableFieldMeasurement()
-    if measurement_type == 'IV Curve Measurement':
-        return IVCurveMeasurement()
-    return Measurement()
-
-
-def get_data_object(measurement_type: str):
-    '''
-    Gets a measurement data MSection object from the given measurement type.
-
-    Args:
-        measurement_type (str): The measurement type.
-
-    Returns:
-        A MSection representing a Hall measurement data object.
-    '''
-    if measurement_type == 'Variable Temperature Measurement':
-        return VariableTemperatureData()
-    if measurement_type == 'Variable Field Measurement':
-        return VariableFieldData()
-    if measurement_type == 'IV Curve Measurement':
-        return IVData()
-    return None
-
-
-def clean_key_str(key: str) -> str:
-    '''
-    Cleans a key, i.e.e replaces spaces and `-` with `_` and converts to lower case
-
-    Args:
-        key (str): The `dirty` key
-
-    Returns:
-        str: The clean key
-    '''
-    return re.sub(r'[\s_\-]+', '_', key).lower()
-
-
-def split_value_unit(expr: str) -> Tuple[str, Optional[str]]:
-    '''
-    Searches for a value unit pair and returns the values for a combination
-    `value [unit]`.
-
-    Args:
-        expr (str): The expression to search for a value unit pair.
-
-    Returns:
-        Tuple[str, Optional[str]]:
-            A tuple of value and unit.
-            Returns the expr, where spaces are replaced with `_` and None when no
-            value unit expression is found.
-    '''
-    is_value_unit = re.search(r'([^\[]+)\s\[(.*)\]', expr)
-    if is_value_unit:
-        value = clean_key_str(is_value_unit.group(1))
[…]
