### 2023-05-12 baba3d4
Resolve "Issue with DOS normalization factor errors"
