### 2023-05-14 6d010e0
Restructure metainfo for clearer semantic connection between `BasisSet` and purpose. Introduce `electrons_representation` as a container for basis sets and pseudo-potentials.
