+++ tests/normalizing/test_method.py
@@ -184,6 +184,61 @@ def test_method_eels(eels):
     assert method.method_name == "EELS"
 
 
+def test_basis_set(normalized_example):
+    ref_basis_sets = {
+        'exciting': ['APW', 'LAPW', 'APW+lo', 'LAPW+lo', '(L)APW', '(L)APW+lo'],
+        'fleur': ['APW', 'LAPW', 'APW+lo', 'LAPW+lo', '(L)APW', '(L)APW+lo'],  # TODO: does not match the metainfo name
+        'WIEN2k': ['APW', 'LAPW', 'APW+lo', 'LAPW+lo', '(L)APW', '(L)APW+lo'],
+        'Elk': ['APW', 'LAPW', 'APW+lo', 'LAPW+lo', '(L)APW', '(L)APW+lo'],
+        'ABINIT': ['plane waves'],
+        'AFlow': ['plane waves'],
+        'atomate': ['plane waves'],
+        'CASTEP': ['plane waves'],
+        'CPMD': ['plane waves'],
+        'MaterialsProject': ['plane waves'],
+        'Quantum Espresso': ['plane waves'],
+        'VASP': ['plane waves'],
+        'AMS': ['atom-centered orbitals'],
+        'Crystal': ['atom-centered orbitals'],
+        'DMol3': ['atom-centered orbitals'],
+        'FHI-aims': ['atom-centered orbitals'],
+        'FHI-vibes': ['atom-centered orbitals'],
+        'GAMESS': ['atom-centered orbitals'],
+        'Gaussian': ['atom-centered orbitals'],
+        'LOBSTER': ['atom-centered orbitals'],
+        'NWChem': ['atom-centered orbitals'],
+        'ORCA': ['atom-centered orbitals'],
+        'Psi4': ['atom-centered orbitals'],
+        'turbomole': ['atom-centered orbitals'],
+        'CP2K': ['gaussians + plane waves', 'plane waves'],
+        'BigDFT': ['real-space grid'],
+        'Octopus': ['real-space grid'],
+        'GPAW': ['real-space grid', 'plane waves', 'atom-centered orbitals'],
+        'Siesta': ['real-space grid', 'atom-centered orbitals'],
+        'ONETEP': ['support functions'],
+        'DL_POLY_4': [],
+        'libAtoms': [],
+        'Phonopy': [],
+        'ATK': [],
+        'gulp': [],
+        'elastic': [],
+        'MOLCAS': [],
+        'qbox': [],
+        'ASR': [],
+        'YAMBO': [],
+    }
+    try:
+        base = normalized_example.results.method.simulation
+    except AttributeError:
+        return
+    if base is not None:
+        program_name = base.program_name
+        if program_name != 'not processed':
+            reference = ref_basis_sets[program_name]
+            if basis_set := base.precision.basis_set:
+                assert basis_set in reference + ['unavailable']
+
+
 @pytest.mark.parametrize('entry, method_identified', [
     ('hash_exciting', True),
     ('hash_vasp', False)
