+++ gui/src/components/search/FilterRegistry.js
@@ -376,6 +376,10 @@ registerFilter('results.method.method_name', idMethod, {...termQuantity, scale: 
 registerFilter('results.method.workflow_name', idMethod, {...termQuantity, scale: '1/4'})
 registerFilter('results.method.simulation.program_name', idMethod, {...termQuantity, scale: '1/4'})
 registerFilter('results.method.simulation.program_version', idMethod, termQuantity)
+registerFilter('results.method.simulation.precision.k_line_density', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
+registerFilter('results.method.simulation.precision.basis_set', idPrecision, {...termQuantity, scale: '1/4'})
+registerFilter('results.method.simulation.precision.planewave_cutoff', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
+registerFilter('results.method.simulation.precision.apw_cutoff', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
 registerFilter('results.method.simulation.dft.basis_set_type', idDFT, {...termQuantity, scale: '1/4'})
 registerFilter('results.method.simulation.dft.core_electron_treatment', idDFT, termQuantity)
 registerFilter('results.method.simulation.dft.jacobs_ladder', idDFT, {...termQuantity, scale: '1/2'})
