### 2023-05-15 a7ffc56
Processing fixes
Fixed lammps, gaussian and bigdft parser errors.

Changelog: Fixed
