### 2023-05-16 #5388 @zhubonan
Hi @sgbaird @berquist and  @arosen93 how is going reviewing the package?
I can see @arosen93 opened this issue here: https://github.com/nomad-coe/nomad/issues/63.

BTW @sgbaird @berquist if you have comments on the package please also open issues on their GitHub page and mention this issues ticket, so they are linked together.
