### 2023-05-16 #5388 @berquist
I've read the paper and believe that NOMAD belongs in JOSS.  I'm still distilling my notes on the (website) documentation and haven't tried installing/testing/running other than building the Docker image.

My current concerns are about the author list.  How do I raise those concerns?
