### 2023-05-17 5ae550d
Resolve "small typo in the jupyterhub config script"
