### 2023-05-19 7b7e6b0
Added basic topologies under results.material.topology for all simulation entries.
Changelog: Added
