### 2023-05-19 5c097f9
Resolve "Processing issues with topology and formulas"
