### 2023-05-22 3feb63c
Resolve "Only the main_author should be allowed to remove an upload"
