### 2023-05-23 #5388 @zhubonan
Just type
Then press the "Comment" button.

This should be able to trigger the editorial bot to generate one for you.
