### 2023-05-23 26bb996
Improved resolution of system references in the visualizer, removed additional comma.
