### 2023-05-23 ec6a8c1
Added the functionality to optionally show the section label instead of full path of the reference
Changelog: Added
