### 2023-05-24 eace52d
Resolve "Documentation for third party integration"
Added documentation for `third-party integration`.

Changelog: Added
