### 2023-05-24 bca5b7d
Resolve "Consider MEnum as key quantity"
