### 2023-05-25 #5388 @arosen93
@markus1978 and @zhubonan:

# Overview

I have finished my review of the code and paper. Overall, it is clear that the NOMAD code is well suited for JOSS and will be of significant interest to the computational materials science community at large. I am personally very excited about the code and am looking forward to use it in my work.

I have no comments about the paper, which is clearly written.

In terms of the checklist items, I have checked off most of them (see below for further details). However, I still think there are a few things that are worth improving on prior to acceptance of this work in JOSS. In general, there is a _ton_ of code, and from a user or contributor standpoint, it's not immediately clear where to look for what in many cases. That accessibility for new users/contributors is the area that I think can be improved the most. It would be a shame if this enormous amount of work didn't end up being adopted outside the inner NOMAD sphere of users simply because of its scale.

# Documentation Comments

## Navigating the Codebase

Overall, I find the codebase to be a bit challenging to navigate. There is the main `nomad` repository (which is mirrored on GitLab and GitHub) and then there's the many NOMAD-hosted packages that `nomad` relies on. For instance, as a user wanting to use the parsers, it's not clear to me based on the GitHub repo where to begin. Do I use the `nomad` repo? The `atomistic-parsers`? The `electronic-parsers`? Some conglomeration thereof? I ultimately figured it out but not without a lot of digging. There should be a clearer map of the various repos in the NOMAD-CoE workspace to help contributors and users.

One small thing that will help (but is not sufficient on its own) is updating documentation links on the `nomad` GitHub repo to be https://nomad-lab.eu/prod/v1/docs/ instead of https://nomad-lab.eu/nomad-lab/. The former leads directly to the docs, whereas the later leads to the broader NOMAD site.

## API Documentation

The documentation for NOMAD itself is very clear, but I was left with many questions about how to use the `nomad` Python tools. There is a [Python section](https://nomad-lab.eu/prod/v1/docs/pythonlib.html) in the documentation, but it's somewhat minimal given how large the codebase is as a whole. For instance, ideally I would like to see more detailed information about the various APIs, such as:
- Auto-generated documentation based on the docstrings highlighting the various functions and classes as well as their inputs and outputs.
- The expected schemas for each code. It's nice to know what to expect before running the code itself.
- What input/output files are parsed for each computational chemistry/materials package.
- What versions of each computational chemistry/materials are supported?

In general, I feel like there is a ton of code here but only a small fraction of it is made accessible to the user (especially prospective developers or code contributors) unless I am missing a section of the documentation.

# Installation and Dependencies

## Splitting up the Dependencies

From a usability standpoint, it's very difficult to incorporate the NOMAD suite into existing codebases due to the large number of dependencies, several of which are not necessarily actively maintained (and therefore will likely become increasingly difficult to incorporat  with newer versions of Python and secondary dependencies). In particular, I can see many use cases where someone might want to use the parsing features of NOMAD but aren't necessarily looking to parse every plausible code supported by NOMAD. In this case, it would be nice to be able to install only the `nomad-lab[parsing]` dependencies that are necessary for the specific code(s) of interest. For instance, `asr` is only needed for Atomic Simulation Recipes, which most people aren't going to need. It would be very helpful and increase the usability of the code if additional "extras" could be provided that would break the `parsing` dependencies down into only those needed for either specific codes or types of codes.

## Installation Woes

On the topic of dependencies, although I checked of "Does installation proceed as outlined," that is only strictly true once the issue I raised (https://github.com/nomad-coe/nomad/issues/63) is resolved.

# Unit Tests

## For Developers

There are unit tests that are run on the GitLab mirror, but they aren't run on the GitHub repo where users can contribute PRs. So, from a contributor perspective, it's not clear from the README or documentation how to run the unit tests locally to test out code. Having to rely on monitoring the GitLab repo for test failures is a bit cumbersome for the user. Since I see a `tests` folder on GitHub, is it as simple as just doing `pytest` in that folder? If so, that should be added to the docs.
