### 2023-05-25 4732c7a
Resolve "When uploading a file with the FileEditQuantity, no upload progress is shown"
