### 2023-05-26 #5388 @sgbaird
Just started tonight. I should have mine within a couple days. I already thought highly of NOMAD, and I'm more impressed as I learn more.
