### 2023-05-27 nomad-coe/nomad#67 @sgbaird
JOSS Review: file 'oasis.md' contains a link to 'assets/nomad-oasis.zip' which is not found in the  documentation files
Following the directions at https://nomad-lab.eu/prod/v1/docs/develop/setup.html

```python
./scripts/setup_dev_env.sh
```
```python
...
Successfully installed nomad-lab-1.2.0rc1.dev124+gaa26e28b7
INFO     -  [macros] - Macros arguments: {'module_name': 'nomad/mkdocs', 'modules': [], 'include_dir': '',
            'include_yaml': [], 'j2_block_start_string': '', 'j2_block_end_string': '', 'j2_variable_start_string': '',
            'j2_variable_end_string': '', 'verbose': False}
INFO     -  [macros] - Found local Python module 'nomad/mkdocs' in:
            /mnt/c/Users/sterg/Documents/GitHub/sparks-baird/nomad
INFO     -  [macros] - Found external Python module 'nomad/mkdocs' in:
            /mnt/c/Users/sterg/Documents/GitHub/sparks-baird/nomad
INFO     -  [macros] - Extra variables (config file): ['generator', 'homepage']
INFO     -  [macros] - Extra filters (module): ['pretty']
INFO     -  Cleaning site directory
INFO     -  Building documentation to directory: /mnt/c/Users/sterg/Documents/GitHub/sparks-baird/nomad/site
INFO     -  The following pages exist in the docs directory, but are not included in the "nav" configuration:
              - aitoolkit.md
WARNING  -  Documentation file 'oasis.md' contains a link to 'assets/nomad-oasis.zip' which is not found in the
            documentation files.
WARNING  -  Documentation file 'oasis.md' contains a link to 'assets/nomad-oasis.zip' which is not found in the
            documentation files.
WARNING  -  Documentation file 'plugins.md' contains a link to 'assets/nomad-oasis-with-plugins.zip' which is not found
            in the documentation files.
INFO     -  Documentation built in 14.97 seconds
```
