### 2023-05-27 nomad-coe/nomad#68 @sgbaird
JOSS Review: importing `ArchiveQuery` gives `FileNotFoundError`
Running on:
`Ubuntu 20.04.6 LTS (GNU/Linux 5.10.16.3-microsoft-standard-WSL2 x86_64)`

```python
(nomad-pypi) sgbaird@Dell-G7:~$ python
Python 3.9.16 (main, Mar  8 2023, 14:00:05)
[GCC 11.2.0] :: Anaconda, Inc. on linux
>>> from nomad.client.archive import ArchiveQuery
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/client/__init__.py", line 19, in <module>
    from .archive import ArchiveQuery
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/client/archive.py", line 28, in <module>
    from nomad import config, metainfo as mi
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/__init__.py", line 95, in <module>
    north = NORTH()
  File "pydantic/main.py", line 340, in pydantic.main.BaseModel.__init__
  File "pydantic/main.py", line 1077, in pydantic.main.validate_model
  File "pydantic/fields.py", line 867, in pydantic.fields.ModelField.validate
  File "pydantic/fields.py", line 1148, in pydantic.fields.ModelField._apply_validators
  File "pydantic/class_validators.py", line 285, in pydantic.class_validators._generic_validator_cls.lambda4
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/models.py", line 272, in load_tools
    with open(path, 'rt') as f:
FileNotFoundError: [Errno 2] No such file or directory: '/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/../../dependencies/nomad-remote-tools-hub/tools.json'
```

<details><summary>Details</summary>

(base) sgbaird@Dell-G7:~$ conda create -n nomad-pypi python==3.9.*
Collecting package metadata (current_repodata.json): done
Solving environment: done

## Package Plan ##

  environment location: /home/sgbaird/miniconda3/envs/nomad-pypi

  added / updated specs:
    - python=3.9

The following NEW packages will be INSTALLED:

  _libgcc_mutex      pkgs/main/linux-64::_libgcc_mutex-0.1-main
  _openmp_mutex      pkgs/main/linux-64::_openmp_mutex-5.1-1_gnu
  ca-certificates    pkgs/main/linux-64::ca-certificates-2023.01.10-h06a4308_0
  ld_impl_linux-64   pkgs/main/linux-64::ld_impl_linux-64-2.38-h1181459_1
  libffi             pkgs/main/linux-64::libffi-3.4.4-h6a678d5_0
  libgcc-ng          pkgs/main/linux-64::libgcc-ng-11.2.0-h1234567_1
  libgomp            pkgs/main/linux-64::libgomp-11.2.0-h1234567_1
  libstdcxx-ng       pkgs/main/linux-64::libstdcxx-ng-11.2.0-h1234567_1
  ncurses            pkgs/main/linux-64::ncurses-6.4-h6a678d5_0
  openssl            pkgs/main/linux-64::openssl-1.1.1t-h7f8727e_0
  pip                pkgs/main/linux-64::pip-23.0.1-py39h06a4308_0
  python             pkgs/main/linux-64::python-3.9.16-h7a1cb2a_2
  readline           pkgs/main/linux-64::readline-8.2-h5eee18b_0
  setuptools         pkgs/main/linux-64::setuptools-66.0.0-py39h06a4308_0
  sqlite             pkgs/main/linux-64::sqlite-3.41.2-h5eee18b_0
  tk                 pkgs/main/linux-64::tk-8.6.12-h1ccaba5_0
  tzdata             pkgs/main/noarch::tzdata-2023c-h04d1e81_0
  wheel              pkgs/main/linux-64::wheel-0.38.4-py39h06a4308_0
  xz                 pkgs/main/linux-64::xz-5.4.2-h5eee18b_0
  zlib               pkgs/main/linux-64::zlib-1.2.13-h5eee18b_0

Proceed ([y]/n)? y

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate nomad-pypi
#
# To deactivate an active environment, use
#
#     $ conda deactivate

(base) sgbaird@Dell-G7:~$ conda activate nomad-pypi
(nomad-pypi) sgbaird@Dell-G7:~$ pip install nomad-lab
Collecting nomad-lab
  Downloading nomad-lab-1.1.10.tar.gz (20.8 MB)
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
[SKIPPING MOST INSTALLATION INFO DUE TO CHARACTER LENGTH LIMIT]

Successfully installed Markdown-3.4.3 SecretStorage-3.3.3 astroid-2.11.7 asttokens-2.2.1 atomicwrites-1.4.1 bleach-6.0.0 build-0.10.0 colorama-0.4.6 coverage-7.2.6 devtools-0.8.0 dill-0.3.6 essential-generators-1.0 executing-0.10.0 ghp-import-2.1.0 isort-5.12.0 jaraco.classes-3.2.3 jeepney-0.8.0 keyring-23.13.1 lazy-object-proxy-1.9.0 mccabe-0.7.0 mergedeep-1.3.4 mkdocs-1.2.3 mkdocs-macros-plugin-0.6.3 mkdocs-material-8.1.1 mkdocs-material-extensions-1.0.3 more-itertools-9.1.0 mypy-1.3.0 mypy-extensions-1.0.0 names-0.3.0 pip-tools-6.13.0 pkginfo-1.9.6 platformdirs-3.5.1 pluggy-1.0.0 py-1.11.0 pycodestyle-2.8.0 pylint-2.13.9 pylint-mongoengine-0.4.0 pylint-plugin-utils-0.7 pymdown-extensions-10.0.1 pyproject_hooks-1.0.0 pytest-3.10.0 pytest-cov-2.7.1 pytest-timeout-1.4.2 python-gitlab-2.10.1 pyyaml-env-tag-0.1 readme-renderer-37.3 requests-toolbelt-1.0.0 rope-0.21.0 termcolor-2.3.0 tomli-2.0.1 twine-3.4.2 typed-ast-1.5.4 watchdog-3.0.0
(nomad-pypi) sgbaird@Dell-G7:~$ python
Python 3.9.16 (main, Mar  8 2023, 14:00:05)
[GCC 11.2.0] :: Anaconda, Inc. on linux
>>> from nomad.client.archive import ArchiveQuery
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/client/__init__.py", line 19, in <module>
    from .archive import ArchiveQuery
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/client/archive.py", line 28, in <module>
    from nomad import config, metainfo as mi
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/__init__.py", line 95, in <module>
    north = NORTH()
  File "pydantic/main.py", line 340, in pydantic.main.BaseModel.__init__
  File "pydantic/main.py", line 1077, in pydantic.main.validate_model
  File "pydantic/fields.py", line 867, in pydantic.fields.ModelField.validate
  File "pydantic/fields.py", line 1148, in pydantic.fields.ModelField._apply_validators
  File "pydantic/class_validators.py", line 285, in pydantic.class_validators._generic_validator_cls.lambda4
  File "/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/models.py", line 272, in load_tools
    with open(path, 'rt') as f:
FileNotFoundError: [Errno 2] No such file or directory: '/home/sgbaird/miniconda3/envs/nomad-pypi/lib/python3.9/site-packages/nomad/config/../../dependencies/nomad-remote-tools-hub/tools.json'
