### 2023-05-28 #5388 @sgbaird
## Overview

NOMAD is an excellent platform with a convenient web app user face, extensive documentation and tutorials, a wide range of features, and a Python package. While NOMAD hosts its own large database of materials, it also provides setup instructions for an isolated solution when someone wishes to implement their own private/secure database. In the past, I had a good experience using NOMAD to create [a dataset of chemical formulas](https://figshare.com/articles/dataset/NOMAD_Chemical_Formulas_and_Calculation_IDs/19319783), larger than what is offered by a database like Materials Project, yet more targeted than a random generation of formulas. Our group used NOMAD recently to [screen for novel, high-performing superconductors](https://doi.org/10.26434/chemrxiv-2023-8t8kt). I think the name NOvely MAterials Discovery (NOMAD) is certainly fitting. Using NOMAD through the API was a positive experience, and the developers were timely in offering support. It also seems like one of the best venues to submit candidate crystal structures from generative modeling efforts. These are some specific examples, and there are many more use-cases. NOMAD is a large benefit to a broad audience.

As an aside, I wanted to point out the following statement that stuck out to me:
> Materials scientists require effective solutions for managing their research data, but they cannot and should not develop their own individual solutions.

NOMAD is a tool that materials scientists can use to effectively manage their research data, and it lessens the gap to becoming a materials data scientist.

#### Installation
  - https://github.com/nomad-coe/nomad/issues/68

#### Functionality
Ideally, I'd like to test out the functionality for ["publish"](https://nomad-lab.eu/prod/v1/gui/search/entries/user/uploads), ["explore"](https://nomad-lab.eu/prod/v1/gui/search/entries/search/entries), and ["analyze"](https://nomad-lab.eu/aitoolkit) (as mentioned at https://nomad-lab.eu/nomad-lab/).
![image](https://github.com/openjournals/joss-reviews/assets/45469701/411dabfc-5975-4b7e-b118-0c874ebbc46a). I will try to work through [the tutorials](https://nomad-lab.eu/nomad-lab/tutorials.html) to ensure each of these work, though I'd like to get the installation issue mentioned above figured out prior to this.

### Documentation

Echoing @ arosen93, the requirements are lengthy and strict in terms of pinned dependencies.

It's great that the installation instructions are well documented, the package is available to download via PyPI, and the web app is easy to use.

The community guidelines would benefit from additional details. For ideas, here's an [example of a PyScaffold default template for CONTRIBUTING.md](https://github.com/pyscaffold/pyscaffold-demo/blob/master/CONTRIBUTING.rst), and here's [an example of how I modified it for another project](https://github.com/sparks-baird/xtal2png/blob/main/CONTRIBUTING.md) (incidentally, a published JOSS manuscript).

Like @arosen93 mentioned, auto-generated API docs for the Python package would be nice.

Agreed with @arosen93. I can see unit tests at, e.g., https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR/-/pipelines/168086, but users should be able to run unit tests locally. It should also be explained why unit tests are run on Gitlab and not on GitHub (e.g., better/faster resources on GitLab through the sponsoring organization), preferably in the same location as the instructions for the unit tests.

### Software paper

> This public NOMAD service contains over 12 million individual materials science simulations and an increasing number of entries describing materials science experiments. NOMAD is publicly available since 2014 and includes data from over 500 international authors.

If available, it would be nice to see an order of magnitude estimate of the number of experimental entries (e.g., over NUM entries describing materials science experiments and counting).

I learned about the NOMAD OASIS feature when reading the manuscript. Great feature!

> NOMAD OASIS can be freely used as per our OSI license ...

... our Open Source Initiative (OSI) license ...

> NOMAD implements OPTIMADE and is an active part of the OPTIMADE consortium.

Could this be confused to mean that NOMAD is the "parent organization" of OPTIMADE? Perhaps "NOMAD conforms to OPTIMADE standards and is accessible via their API. Additionally, NOMAD is an active participant of the OPTIMADE consortium."

> Other materials science databases (and the respective software) focus on only publishing data that were produced by the group behind the database itself. Typical examples are databases of high-throughput simulations that try to systematically explore theoretical materials. Three of the larger databases of this kind are the Materials Project (Jain et al., 2013), ...

Two examples that come to mind with an interface for contributing user data (not internally generated by the group). The following is copied from https://github.com/sparks-baird/self-driving-lab-demo/discussions/117#discussioncomment-6019686:
- [Foundry by Globus](https://labs.globus.org/projects/foundry.html)
  - > Here, we introduce Foundry, a simple way to publish and discover datasets for machine learning and to link these datasets to predictive models. Foundry is a synthesis of service capabilities from the Materials Data Facility and DLHub, layered with Python tooling, standardized metadata, and a file structure specification to meet the needs of the machine learning community.
- [MPContribs](https://materialsproject.org/contribs)
  - > MPContribs provides a platform and advanced programming interface (API) to contribute computational as well as experimental data to Materials Project. Data on MPContribs is collectively maintained as annotations to existing MP materials (or formulas and chemical systems), and automatically exposed to over 200,000 MP users.

Is it also worth mentioning non-materials-science-specific options, like MongoDB? Even if not incorporated into the manuscript, I'm curious to hear your opinion.

While not a database that supports user contributions, a notable database is [MPDS](https://mpds.io), which arguably contains the largest collection of experimentally determined crystal structures and materials properties extracted from literature. However, the data is kept private through a subscription paywall. I think it would be worth mentioning MPDS.

Please consider adding bulleted lists of key features from the three categories in https://nomad-lab.eu/nomad-lab/ (see image above) as some type of list in the manuscript (concatenated list, three separate lists, table, etc.).
