### 2023-05-30 62d3a60
Moved ELN Base Sections
Moved the base sections from nomad.datamodel.metainfo.eln to nomad.datamodel.metainfo.basesections

Reimported the base sections into nomad.datamodel.metainfo.eln to retain compatibility

Changelog: Changed
