+++ nomad/datamodel/metainfo/basesections.py
@@ -1,0 +1,1175 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import datetime
+import re
+from typing import (
+    Dict,
+    List,
+)
+
+import numpy as np
+from ase.data import (
+    chemical_symbols,
+    atomic_numbers,
+    atomic_masses,
+)
+from structlog.stdlib import (
+    BoundLogger,
+)
+
+from nomad import (
+    utils,
+)
+from nomad.units import (
+    ureg,
+)
+from nomad.metainfo import (
+    Quantity,
+    Datetime,
+    Reference,
+    Section,
+    SubSection,
+)
+from nomad.metainfo.util import (
+    MEnum,
+)
+from nomad.datamodel.data import (
+    ArchiveSection,
+    EntryData,
+)
+from nomad.datamodel.results import (
+    Results,
+    ELN,
+    ElementalComposition as ResultsElementalComposition,
+    Material,
+)
+
+
+class BaseSection(ArchiveSection):
+    '''
+    A generic abstract base section that provides a few commonly used properties.
+
+    If you inherit from this section, but do not need some quantities, list those
+    quantities in the `eln.hide` annotation of your inheriting section definition.
+
+    Besides predefining some quantities, these base sections will add some metadata
+    to NOMAD's search. A particular example are `tags`, if you define a string
+    or enum quantity in your sections named `tags`, its values will be searchable.
+    '''
+    m_def = Section(
+        links=['http://purl.obolibrary.org/obo/BFO_0000001'],
+    )
+    name = Quantity(
+        type=str,
+        description='A short human readable and descriptive name.',
+        a_eln=dict(component='StringEditQuantity', label='Short name'),
+    )
+    datetime = Quantity(
+        type=Datetime,
+        description='The date and time associated with this section.',
+        a_eln=dict(component='DateTimeEditQuantity'),
+    )
+    lab_id = Quantity(
+        type=str,
+        description='''An ID string that is unique at least for the lab that produced this
+            data.''',
+        a_eln=dict(component='StringEditQuantity', label="ID"),
+    )
+    description = Quantity(
+        type=str,
+        description='Any information that cannot be captured in the other fields.',
+        a_eln=dict(component='RichTextEditQuantity'),
+    )
+
+    def normalize(self, archive, logger: BoundLogger) -> None:
+        '''
+        The normalizer for the `BaseSection` class.
+
+        Args:
+            archive (EntryArchive): The archive containing the section that is being
+            normalized.
+            logger (BoundLogger): A structlog logger.
+        '''
+        super(BaseSection, self).normalize(archive, logger)
+
+        if isinstance(self, EntryData):
+            if archive.data == self and self.name:
+                archive.metadata.entry_name = self.name
+            elif self.name is None:
+                self.name = archive.metadata.entry_name.split('.')[0].replace('_', ' ')
+            EntryData.normalize(self, archive, logger)
+
+        if not archive.results:
+            archive.results = Results(eln=ELN())
+        if not archive.results.eln:
+            archive.results.eln = ELN()
+
[…]
