+++ gui/tests/artifacts.js
@@ -7449,7 +7449,7 @@ window.nomadArtifacts = {
                 "name": "starting_method_ref",
                 "description": "Links the current section method to a section method containing the starting\nparameters.",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "type": {
                   "type_kind": "reference",
@@ -7464,7 +7464,7 @@ window.nomadArtifacts = {
                 "name": "core_method_ref",
                 "description": "Links the current section method to a section method containing the core settings.",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "type": {
                   "type_kind": "reference",
@@ -7491,7 +7491,7 @@ window.nomadArtifacts = {
                 "name": "methods_ref",
                 "description": "Links the section method to other method sections. For instance, one calculation\nis a perturbation performed using a self-consistent field (SCF) calculation as\nstarting point, or a simulated system is partitioned in regions with different but\nconnected Hamiltonians (e.g., QM/MM, or a region treated via Kohn-Sham DFT\nembedded into a region treated via orbital-free DFT).",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "type": {
                   "type_kind": "reference",
@@ -8434,7 +8434,7 @@ window.nomadArtifacts = {
                 "name": "sub_system_ref",
                 "description": "Links the current section system to a sub system.",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "type": {
                   "type_kind": "reference",
@@ -8449,7 +8449,7 @@ window.nomadArtifacts = {
                 "name": "systems_ref",
                 "description": "Links the current section system to other section systems. Such a link is\nnecessary for example between the supercell and the reference unit cell in a phonon\ncalculation. The relationship should be described by kind and the referred section\nsystem is given by value. An external url can also be provided in place of value.",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "type": {
                   "type_kind": "reference",
@@ -8515,7 +8515,7 @@ window.nomadArtifacts = {
                 "m_parent_sub_section": "sub_sections",
                 "name": "atoms",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "sub_section": "/packages/1/section_definitions/1"
               },
@@ -8541,7 +8541,7 @@ window.nomadArtifacts = {
                 "m_parent_sub_section": "sub_sections",
                 "name": "prototype",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "sub_section": "/packages/1/section_definitions/3",
                 "repeats": true
@@ -8552,7 +8552,7 @@ window.nomadArtifacts = {
                 "m_parent_sub_section": "sub_sections",
                 "name": "springer_material",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "sub_section": "/packages/1/section_definitions/4",
                 "repeats": true
@@ -8563,7 +8563,7 @@ window.nomadArtifacts = {
                 "m_parent_sub_section": "sub_sections",
                 "name": "symmetry",
                 "categories": [
-                  "/packages/16/category_definitions/0"
+                  "/packages/17/category_definitions/0"
                 ],
                 "sub_section": "/packages/1/section_definitions/2",
[…]
