### 2023-05-30 39891c4
Resolve "Replace workflow with workflow2"
Remove old workflow section: definitions, parsing, normalization.

Changelog: Added
