+++ tests/archive/test_query_reader.py
@@ -61,6 +61,8 @@ def assert_dict(d1, d2):
         del d1[GeneralReader.__CACHE__]
     if 'pagination' in d1:
         del d1['pagination']
+    if 'm_def' in d1:
+        del d1['m_def']
     assert set(d1.keys()) == set(d2.keys())
     for k, v in d1.items():
         if isinstance(v, dict):
@@ -553,17 +555,17 @@ def test_remote_reference(json_dict, example_data_with_reference, test_user):
         'upload_files_server_path': 'id_published_with_ref',
         'm_raw': {
             'mainfile_for_id_02': {
-                'path': 'mainfile_for_id_02', 'is_file': True, 'size': 3237, 'access': 'unpublished'},
+                'path': 'mainfile_for_id_02', 'is_file': True, 'size': 3231, 'access': 'unpublished'},
             'mainfile_for_id_03': {
-                'path': 'mainfile_for_id_03', 'is_file': True, 'size': 3237, 'access': 'unpublished'},
+                'path': 'mainfile_for_id_03', 'is_file': True, 'size': 3231, 'access': 'unpublished'},
             'mainfile_for_id_04': {
-                'path': 'mainfile_for_id_04', 'is_file': True, 'size': 3237, 'access': 'unpublished'},
+                'path': 'mainfile_for_id_04', 'is_file': True, 'size': 3231, 'access': 'unpublished'},
             'mainfile_for_id_05': {
-                'path': 'mainfile_for_id_05', 'is_file': True, 'size': 3237, 'access': 'unpublished'},
+                'path': 'mainfile_for_id_05', 'is_file': True, 'size': 3231, 'access': 'unpublished'},
             'mainfile_for_id_06': {
-                'path': 'mainfile_for_id_06', 'is_file': True, 'size': 3237, 'access': 'unpublished'},
+                'path': 'mainfile_for_id_06', 'is_file': True, 'size': 3231, 'access': 'unpublished'},
             'mainfile_for_id_01': {
-                'path': 'mainfile_for_id_01', 'is_file': True, 'size': 3237, 'access': 'unpublished',
+                'path': 'mainfile_for_id_01', 'is_file': True, 'size': 3231, 'access': 'unpublished',
                 'm_entries': {
                     'process_running': False, 'current_process': None,
                     'process_status': 'SUCCESS', 'last_status_message': None,
@@ -734,12 +736,12 @@ def test_remote_reference(json_dict, example_data_with_reference, test_user):
             'm_request': {
                 'directive': 'resolved',
                 'resolve_inplace': True,
-                'include': ['workflow']
+                'include': ['workflow2']
             },
         }
     }, result={'m_uploads': {'id_published_with_ref': {'m_entries': {'id_01': {'m_archive': {
-        'workflow': [{
-            'calculation_result_ref': 'm_uploads/id_published_with_ref/m_entries/id_01/m_archive/run/0/calculation/1'}],
+        'workflow2': {
+            'results': {'calculation_result_ref': 'm_uploads/id_published_with_ref/m_entries/id_01/m_archive/run/0/calculation/1'}},
         'run': [{
             'calculation': [
                 None, {
@@ -763,20 +765,20 @@ def test_remote_reference(json_dict, example_data_with_reference, test_user):
                         'H']},
                 'symmetry': [{
                     'space_group_number': 221}]}]}]}}}}},
-        'm_archive': {'workflow': [
-            {'workflows_ref': ['m_uploads/id_published_with_ref/m_entries/id_01/m_archive/workflow/0']}]}})
[…]
