+++ nomad/normalizing/results.py
@@ -32,6 +32,7 @@ from nomad.normalizing.method import MethodNormalizer
 from nomad.normalizing.material import MaterialNormalizer
 from nomad.datamodel.metainfo.workflow import Workflow
 from nomad.datamodel.metainfo.simulation.system import System, Symmetry as SystemSymmetry
+from nomad.datamodel.metainfo.simulation.workflow import ThermodynamicsResults
 from nomad.normalizing.common import structures_2d
 from nomad.datamodel.results import (
     BandGap,
@@ -188,6 +189,34 @@ class ResultsNormalizer(Normalizer):
         ).material()
 
         results.method = MethodNormalizer(self.entry_archive, repr_system, results.material, logger).method()
+
+        # set entry type based on method and material
+        workflow = self.entry_archive.workflow2
+        if workflow is not None:
+
+            workflow_name = workflow.name if workflow.name else workflow.m_def.name
+
+            tag = ''
+            if results.method.simulation:
+                tag = 'simulation'
+
+            try:
+                method_name = results.method.method_name
+                program_name = results.method.simulation.program_name
+                if workflow_name == 'SinglePoint' and method_name:
+                    self.entry_archive.metadata.entry_type = f'{program_name} {method_name} {workflow_name}'
+                else:
+                    self.entry_archive.metadata.entry_type = f'{program_name} {workflow_name}'
+            except Exception:
+                self.entry_archive.metadata.entry_type = workflow_name
+            type_tag = f'{self.entry_archive.metadata.entry_type} {tag}'
+
+            # Populate entry_name
+            material = results.material
+            if material and material.chemical_formula_descriptive:
+                self.entry_archive.metadata.entry_name = f'{material.chemical_formula_descriptive} {type_tag}'
+            else:
+                self.entry_archive.metadata.entry_name = f'{type_tag}'
 
     def resolve_band_gap(self, path: list[str]) -> Union[List[BandGap], None]:
         """Extract all band gaps from the given `path` and return them in a list along
@@ -387,7 +416,14 @@ class ResultsNormalizer(Normalizer):
           - There is a non-empty array of temperatures.
           - There is a non-empty array of energies.
         """
-        path = ["workflow", "thermodynamics"]
+        workflow = self.entry_archive.workflow2
+        if workflow is None or not hasattr(workflow, 'results'):
+            return None
+        if not isinstance(workflow.results, ThermodynamicsResults):
+            return None
+
+        path = ["workflow2", "results"]
+
         for thermo_prop in traverse_reversed(self.entry_archive, path):
             temperatures = thermo_prop.temperature
             energies = thermo_prop.vibrational_free_energy_at_constant_volume
@@ -407,7 +443,13 @@ class ResultsNormalizer(Normalizer):
           - There is a non-empty array of temperatures.
           - There is a non-empty array of energies.
         """
-        path = ["workflow", "thermodynamics"]
+        workflow = self.entry_archive.workflow2
+        if workflow is None or not hasattr(workflow, 'results'):
+            return None
+        if not isinstance(workflow.results, ThermodynamicsResults):
+            return None
+
+        path = ["workflow2", "results"]
         for thermo_prop in traverse_reversed(self.entry_archive, path):
             temperatures = thermo_prop.temperature
             heat_capacities = thermo_prop.heat_capacity_c_v
@@ -423,24 +465,23 @@ class ResultsNormalizer(Normalizer):
         """Populates both geometry optimization methodology and calculated
         properties based on the first found geometry optimization workflow.
         """
-        path = ["workflow"]
+        path = ["workflow2"]
         for workflow in traverse_reversed(self.entry_archive, path):
             # Check validity
-            if workflow.type == "geometry_optimization" and workflow.calculations_ref:
-
+            if workflow.m_def.name == 'GeometryOptimization':
[…]
