### 2023-05-30 nomad-coe/nomad#67 @markus1978
Yes there is an additional script creating those .zip files that should be called in `./scripts/setup_dev_env.sh`.
