### 2023-05-30 nomad-coe/nomad#68 @markus1978
We are about to remove this file entirely and make it part of the configuration. The problem will be solved, once the new config is merged.
