### 2023-05-30 nomad-coe/nomad#70 @markus1978
Let's treat this as a duplicate of #69 for now.
