### 2023-05-30 nomad-coe/nomad#70 @markus1978
When I run this with a newer Python version on Colab (Python 3.10.x), I not surprisingly run into #68.
