### 2023-05-30 nomad-coe/nomad#70 @markus1978
To get rid of the `ContextualVersionConflict: (wrapt 1.14.1`, please install `pip install wrapt==1.14.1` and restart the runtime.

This happens when the Python runtime already imported one version of a package and now after installing `nomad-lab` and thereby changing the version and now trying to import this other version. Forcing the install of `wrapt==1.14.1` after installing `nomad-lab` and getting a fresh Python runtime allows to work around this.
