### 2023-05-30 449ca9e
Fixed the staging deployment.
Instead of using the branch as a tag, we are pushing the current branch image ue under the staging tag and then using the staging tag.

Changelog: Fixed
