+++ .gitlab-ci.yml
@@ -280,6 +280,9 @@ deploy prod staging:
     - mkdir ~/.kube/
     - echo ${CI_K8S_PROD_CONFIG} | base64 -d > ~/.kube/config
   script:
+    - docker pull ${CI_REGISTRY_IMAGE}:${DOCKER_TAG}
+    - docker tag ${CI_REGISTRY_IMAGE}:staging
+    - docker push ${CI_REGISTRY_IMAGE}:staging
     - helm dependency update ops/kubernetes/nomad
     - helm upgrade nomad-staging-v1 ops/kubernetes/nomad
       --install
@@ -289,7 +292,7 @@ deploy prod staging:
       --timeout=15m
       --wait
     - docker pull ${CI_REGISTRY_IMAGE}:${DOCKER_TAG}
-    - docker run -t -e NOMAD_KEYCLOAK_REALM_NAME=fairdi_nomad_prod ${CI_REGISTRY_IMAGE}/dev_python:${DOCKER_TAG}
+    - docker run -t -e NOMAD_KEYCLOAK_REALM_NAME=fairdi_nomad_prod ${CI_REGISTRY_IMAGE}/dev_python:staging
       nomad client -n https://nomad-lab.eu/prod/v1/staging/api -u test -w $CI_NOMAD_TEST_PASSWORD integrationtests --skip-publish --skip-doi
   rules:
     - when: manual
