### 2023-05-30 f2d8e82
Fixed the name of the fhi parser.
The name was accidentally changed, presumably when adding the
plugins. It is now changed back. It is important that the
name does not change, because of existing entries.

Changelog: Fixed
