### 2023-05-30 daa4f98
Changed the used jupyterlab image in NORTH.
We are now using our custom image with nomad-lab pre-installed.

Changelog: Changed
