+++ gui/tests/artifacts.js
@@ -69652,7 +69652,7 @@ window.nomadArtifacts = {
   },
   "northTools": {
     "jupyter": {
-      "image": "jupyter/datascience-notebook:lab-3.6.2",
+      "image": "gitlab-registry.mpcdf.mpg.de/nomad-lab/nomad-remote-tools-hub/jupyterlab:latest",
       "description": "### **Jupyter Notebook**: The Classic Notebook Interface\n\nThe Jupyter Notebook is the original web application for creating and sharing computational documents. It offers a simple, streamlined, document-centric experience.",
       "short_description": "Basic jupyter run with an empty notebook or on given notebook file.",
       "cmd": null,
