### 2023-05-31 c1f4147
Added authentication for mongo and elastic use.
Changelog: Added
