+++ ops/kubernetes/deployments/prod-values.yaml
@@ -22,7 +22,7 @@ gui:
   matomoSiteId: 2
   encyclopediaBase: "https://nomad-lab.eu/prod/rae/encyclopedia/#"
   aitoolkitEnabled: false
-  config:     
+  config:
     entry:
       cards:
         exclude: ["relatedResources", "workflow"]
@@ -41,9 +41,11 @@ worker:
 elastic:
   host: elasticsearch.elasticsearch-7.svc.cluster.local
   port: 9200
+  secret: nomad-elastic
 
 mongo:
   host: mongodb://mongodb-0.mongo.mongodb.svc.cluster.local,mongodb-1.mongo.mongodb.svc.cluster.local,mongodb-2.mongo.mongodb.svc.cluster.local/?replicaSet=rs0
+  secret: nomad-mongo
 
 logstash:
   host: logstash.elk.svc.cluster.local
