+++ nomad/cli/aflow.py
@@ -564,7 +564,7 @@ def write_prototype_data_file(aflow_prototypes: dict, filepath) -> None:
     with io.open(filepath, "w", encoding="utf8") as f:
         json_dump = json.dumps(aflow_prototypes, ensure_ascii=False, indent=4, sort_keys=True, cls=NoIndentEncoder)
         json_dump = re.sub(r"\"(-?\d+(?:[\.,]\d+)?)\"", r'\1', json_dump)  # Removes quotes around numbers
-        f.write("# -*- coding: utf-8 -*-\naflow_prototypes = {}\n".format(json_dump))
+        f.write("aflow_prototypes = {}\n".format(json_dump))
 
 
 def update_prototypes(ctx, filepath, matches_only):
