+++ nomad/aflow_prototypes.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Copyright The NOMAD Authors.
 #
@@ -2950,7 +2949,7 @@ aflow_prototypes = {
                 "atom_labels": ["Cd", "Cd", "Al", "Al", "Al", "Al", "S", "S", "S", "S", "S", "S", "S", "S"],
                 "atom_positions": [[0.0, 0.0, 0.0], [2.775, 2.775, 5.15], [0.0, 0.0, 5.15], [2.775, 2.775, 0.0], [0.0, 2.775, 2.575], [2.775, 0.0, 7.7250000000000005], [1.443, 1.3875, 1.3390000000000002], [4.107, 4.1625, 1.3390000000000002], [1.3875, 4.107, 8.961], [4.1625, 1.443, 8.961], [4.218, 4.1625, 6.489000000000001], [1.3319999999999999, 1.3875, 6.489000000000001], [4.1625, 1.3319999999999999, 3.8110000000000004], [1.3875, 4.218, 3.8110000000000004]],
                 "lattice_vectors": [[5.55, 0.0, 0.0], [0.0, 5.55, 0.0], [0.0, 0.0, 10.3]],
-                "normalized_wyckoff_matid": {"a": {"X_1": 1}, "b": {"X_2": 1}, "d": {"X_1": 1}, "g": {"X_0": 4}}
+                "normalized_wyckoff_matid": {"a": {"X_1": 1}, "b": {"X_2": 1}, "c": {"X_1": 1}, "g": {"X_0": 4}}
             },
             {
                 "Notes": "BPO<sub>4</sub> (H0<sub>7</sub>) Structure",
@@ -2964,7 +2963,7 @@ aflow_prototypes = {
                 "atom_labels": ["P", "P", "B", "B", "O", "O", "O", "O", "O", "O", "O", "O"],
                 "atom_positions": [[0.0, 0.0, 0.0], [2.1702, 2.1702, 3.3251], [0.0, 2.1702, 1.66255], [2.1702, 0.0, 4.98765], [1.1111424, 1.11374664, 2.4752044399999997], [3.2292576, 3.2266533600000002, 2.4752044399999997], [1.11374664, 3.2292576, 4.17499556], [3.2266533600000002, 1.1111424, 4.17499556], [3.2813423999999998, 3.2839466399999995, 5.80030444], [1.0590575999999998, 1.05645336, 5.80030444], [3.2839466399999995, 1.0590575999999998, 0.8498955600000001], [1.05645336, 3.2813423999999998, 0.8498955600000001]],
                 "lattice_vectors": [[4.3404, 0.0, 0.0], [0.0, 4.3404, 0.0], [0.0, 0.0, 6.6502]],
-                "normalized_wyckoff_matid": {"a": {"X_1": 1}, "d": {"X_2": 1}, "g": {"X_0": 4}}
+                "normalized_wyckoff_matid": {"a": {"X_1": 1}, "c": {"X_2": 1}, "g": {"X_0": 4}}
             },
             {
                 "Notes": "Ni<sub>3</sub>P ($D0_{e}$) Structure",
@@ -3764,7 +3763,7 @@ aflow_prototypes = {
                 "atom_labels": ["Sb", "Sb", "Ga", "Ga"],
                 "atom_positions": [[0.0, 0.0, 0.0], [2.7395050752, 2.7395050752, 1.5300028345], [0.0, 2.7395050752, 0.76500141725], [2.7395050752, 0.0, 2.29500425175]],
[…]
