+++ examples/data/light_eln/schema.archive.yaml
@@ -1,0 +1,215 @@
+# Schemas can be defined as yaml files like this. The archive.yaml format will be
+# interpreted by nomad as a nomad archive. Therefore, all definitions have to be
+# put in a top-level section called "definitions"
+definitions:
+  # The "definitions" section is interpreted as a nomad schema package
+  # Schema packages can have a name:
+  name: 'Electronic Lab Notebook example schema'
+  # Schema packages contain section definitions. This is where the interesting schema
+  # information begins.
+  sections:
+    # Here we define a section called "Chemical":
+    Chemical:
+      # Section definition can have base_sections. Base sections are other schema
+      # definition and all properties of these will be inherited.
+      base_sections:
+        - 'nomad.datamodel.metainfo.eln.Chemical'  # Provides typical quantities like name, descriptions, chemical_formula and makes those available for search
+        - 'nomad.datamodel.data.EntryData'  # Declares this as a top-level entry section. This determines the types of entries you can create. With this we will be able to create a "Chemical" entry.
+      # All definitions, sections, sub_sections, quantities, can provide a description.
+      description: |
+        This is an example description for Chemical.
+        A description can contain **markdown** markup and TeX formulas, like $\sum\limits_{i=0}^{n}$.
+      # Sections define quantities. Quantities allow to manage actual data. Quantities
+      # can have various types, shapes, and units.
+      quantities:
+        # Here we define a quantity called "from"
+        form:
+          # This defines a Enum type with pre-defined possible values.
+          type:
+            type_kind: Enum
+            type_data:
+              - crystalline solid
+              - powder
+          # Annotations allow to provide additional information that is beyond just defining
+          # the possible data.
+          m_annotations:
+            # The eln annotation allows add the quantity to a ELN
+            eln:
+              component: EnumEditQuantity  # A form field component for EnumQuantities that uses a pull down menu.
+        cas_number:
+          type: str
+          m_annotations:
+            eln:
+              component: StringEditQuantity
+        ec_number:
+          type: str
+          m_annotations:
+            eln:
+              component: StringEditQuantity
+    Instrument:
+      base_sections:
+        - nomad.datamodel.metainfo.eln.Instrument
+        - nomad.datamodel.data.EntryData
+    Process:
+      base_section: nomad.datamodel.metainfo.eln.Process
+      quantities:
+        instrument:
+          type: Instrument
+          m_annotations:
+            eln:
+              component: ReferenceEditQuantity
+    Sample:
+      m_annotations:
+        # The template annotation allows to define what freshly created entries (instances of this schema) will look like.
+        # In this example we create a sample with an empty pvd_evaporation process.
+        template:
+          processes:
+            pvd_evaporation: {}
+      base_sections:
+        - 'nomad.datamodel.metainfo.eln.Sample'
+        - 'nomad.datamodel.data.EntryData'
+      quantities:
+        name:
+          type: str  # The simple string type
+          default: Default Sample Name
+          m_annotations:
+            eln:
+              component: StringEditQuantity  # A simple text edit form field
+        tags:
+          type:
+            type_kind: Enum
+            type_data:
+              - internal
+              - collaboration
+              - project
+              - other
+          shape: ['*']  # Shapes define non scalar values, like lists ['*'], vectors ['*', 3], etc.
+          m_annotations:
+            eln:
+              component: AutocompleteEditQuantity  # Allows to edit enums with an auto complete text form field
+        chemicals:
+          type: Chemical  # Types can also be other sections. This allows to reference a different section.
+          shape: ['*']
+          m_annotations:
+            eln:
+              component: ReferenceEditQuantity  # A editor component that allows to select from available "Chemical"s
+        substrate_type:
+          type:
[…]
