### 2023-06-08 9b9e694
Resolve "Referenced section is not working"
Fix resolution of ArchiveSection ref and negative index in archive browser

Changelog: Fixed
