### 2023-06-08 37195bd
Implement eos fit normalization #1439
Changelog: Added
