+++ docs/develop/setup.md
@@ -46,7 +46,7 @@ This branch can be pushed to the repo, and then later may be merged to the relev
 
 Nomad is based on python modules from the NOMAD-coe project.
 This includes parsers, python-common and the meta-info. These modules are maintained as
-their own GITLab/git repositories. To clone and initialize them run:
+their own GITLab/git repositories. To clone and initialize these modules, run:
 
 ```sh
 git submodule update --init
@@ -60,7 +60,7 @@ You should work in a Python virtual environment.
 
 #### pyenv
 The nomad code currently targets python 3.9. If your host machine has an older version installed,
-you can use [pyenv](https://github.com/pyenv/pyenv) to use python 3.9 in parallel to your
+you can use [pyenv](https://github.com/pyenv/pyenv) to use python 3.9 in parallel with your
 system's python.
 
 #### virtualenv
@@ -104,7 +104,7 @@ unix/linux systems. It can be installed on MacOS with homebrew:
 brew install libmagic
 ```
 
-If you are using an Mac with Apple Silicon, we recommend to use rosetta, homebrew
+If you are using a Mac with Apple Silicon, we recommend that you use rosetta, homebrew
 for Intel, and install and use an Intel based Python. The second answer in this
 [Stackoverflow post](https://stackoverflow.com/questions/64882584/how-to-run-the-homebrew-installer-under-rosetta-2-on-m1-macbook)
 describes how to use both the Apple and Intel homebrew simultaneously.
@@ -141,14 +141,14 @@ and nomad itself.
     ```
 
     Next we install the `nomad` package itself (including all extras). The `-e`
-    option will install the NOMAD with symbolic links allowing you to change
+    option will install NOMAD with symbolic links that allow you to change
     the code without having to reinstall after each change.
 
     ```sh
     pip install -e .[parsing,infrastructure,dev]
     ```
 
-    If pip tries to use and compile sources and this creates errors, it can be told to prefer binary version:
+   If pip tries to use and compile sources that create errors, it can be told to prefer the binary version:
 
     ```sh
     pip install -e .[parsing,infrastructure,dev] --prefer-binary
@@ -161,7 +161,7 @@ and nomad itself.
     ```
 
     This file includes some of the server details that are needed so that the
-    GUI can make the initial connection properly. If you e.g. change the server
+    GUI can make the initial connection properly. If for example you change the server
     address in your NOMAD configuration file, it will be necessary to regenerate
     this .env file. In production this file will be overridden.
 
@@ -184,7 +184,7 @@ To run NOMAD, some 3rd party services are needed
 - rabbitmq: a task queue used to distribute work in a cluster
 
 All 3rd party services should be run via *docker-compose* (see below).
-Keep in mind the *docker-compose* configures all services in a way that mirror
+Keep in mind that *docker-compose* configures all services in a way that mirrors
 the configuration of the python code in `nomad/config.py` and the gui config in
 `gui/.env.development`.
 
@@ -236,7 +236,7 @@ keycloak:
 ```
 
 You might also want to exclude some of the default plugins, or only include the plugins
-you'll need. Especially plugins that slower start-up and import times due to instantiation
+you'll need. Especially plugins with slower start-up and import times due to instantiation
 of large schemas (e.g. nexus create couple thousand definitions for 70+ applications) can
 often be excluded.
 
@@ -274,12 +274,7 @@ To run the worker directly with celery, do (from the root)
 celery -A nomad.processing worker -l info
 ```
 
-Before you can run the gui, make sure that generated the GUI artifacts:
-```sh
-./scripts/generate_gui_artifacts.sh
-```
-
-If you run the gui on its own (e.g. with react dev server below), you also have to have
+If you run the gui on its own (e.g. with react dev server below), you also need to start
[…]
