### 2023-06-09 5392781
Address performance issue in matching parser with child archives
Replace use of text parser in get_mainfile_keys with step-wise search

Changelog: Fixed
