+++ gui/src/components/search/FilterRegistry.js
@@ -378,12 +378,11 @@ registerFilter('results.method.method_name', idMethod, {...termQuantity, scale: 
 registerFilter('results.method.workflow_name', idMethod, {...termQuantity, scale: '1/4'})
 registerFilter('results.method.simulation.program_name', idMethod, {...termQuantity, scale: '1/4'})
 registerFilter('results.method.simulation.program_version', idMethod, termQuantity)
-registerFilter('results.method.simulation.precision.native_tier', idPrecision, {...termQuantityNonExclusive, scale: '1/2'})
+registerFilter('results.method.simulation.precision.native_tier', idPrecision, {...termQuantity, placeholder: "E.g. VASP - accurate", label: 'Code-specific Tier'})
 registerFilter('results.method.simulation.precision.k_line_density', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
 registerFilter('results.method.simulation.precision.basis_set', idPrecision, {...termQuantity, scale: '1/4'})
-registerFilter('results.method.simulation.precision.planewave_cutoff', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
-registerFilter('results.method.simulation.precision.apw_cutoff', idPrecision, {...numberHistogramQuantity, scale: '1/2'})
-registerFilter('results.method.simulation.dft.basis_set_type', idDFT, {...termQuantity, scale: '1/4'})
+registerFilter('results.method.simulation.precision.planewave_cutoff', idPrecision, {...numberHistogramQuantity, label: 'Plane-wave Cutoff', scale: '1/2'})
+registerFilter('results.method.simulation.precision.apw_cutoff', idPrecision, {...numberHistogramQuantity, label: 'APW Cutoff', scale: '1/2'})
 registerFilter('results.method.simulation.dft.core_electron_treatment', idDFT, termQuantity)
 registerFilter('results.method.simulation.dft.jacobs_ladder', idDFT, {...termQuantity, scale: '1/2'})
 registerFilter('results.method.simulation.dft.xc_functional_type', idDFT, {
@@ -440,7 +439,7 @@ registerFilter('results.method.simulation.dmft.magnetic_state', idDMFT, {...term
 registerFilter('results.method.simulation.dmft.inverse_temperature', idDMFT, {...numberHistogramQuantity, scale: '1/2'})
 registerFilter('results.method.simulation.dmft.u', idDMFT, {...numberHistogramQuantity, scale: '1/2'})
 registerFilter('results.method.simulation.dmft.hunds_hubbard_ratio', idDMFT, {...numberHistogramQuantity, label: `JH / U`, scale: '1/2'})
-registerFilter('results.method.simulation.precision.k_line_density', idPrecision, termQuantity)
+registerFilter('results.method.simulation.precision.k_line_density', idPrecision, {...termQuantity, label: 'k-line Density'})
 registerFilter('results.eln.sections', idELN, termQuantity)
 registerFilter('results.eln.tags', idELN, termQuantity)
 registerFilter('results.eln.methods', idELN, termQuantity)
