+++ nomad/config/models.py
@@ -987,13 +987,12 @@ class UI(StrictSettings):
                             'authors': {'label': 'Authors', 'align': 'left'},
                             'results.method.method_name': {'label': 'Method name'},
                             'results.method.simulation.program_name': {'label': 'Program name'},
-                            'results.method.simulation.dft.basis_set_name': {'label': 'Basis set name'},
                             'results.method.simulation.dft.xc_functional_type': {'label': 'XC Functional Type'},
-                            'results.method.simulation.precision.apw_cutoff': {'label': 'APW cutoff'},
-                            'results.method.simulation.precision.basis_set': {'label': 'Basis set'},
-                            'results.method.simulation.precision.k_line_density': {'label': 'K-line density'},
-                            'results.method.simulation.precision.native_tier': {'label': 'Native tier'},
-                            'results.method.simulation.precision.planewave_cutoff': {'label': 'Plane wave cutoff'},
+                            'results.method.simulation.precision.apw_cutoff': {'label': 'APW Cutoff'},
+                            'results.method.simulation.precision.basis_set': {'label': 'Basis Set'},
+                            'results.method.simulation.precision.k_line_density': {'label': 'k-line Density'},
+                            'results.method.simulation.precision.native_tier': {'label': 'Code-specific tier'},
+                            'results.method.simulation.precision.planewave_cutoff': {'label': 'Plane-wave cutoff'},
                             'results.material.structural_type': {'label': 'Dimensionality'},
                             'results.material.symmetry.crystal_system': {'label': 'Crystal system'},
                             'results.material.symmetry.space_group_symbol': {'label': 'Space group symbol'},
@@ -1076,13 +1075,12 @@ class UI(StrictSettings):
                             'results.method.method_name': {'label': 'Method name'},
                             'upload_create_time': {'label': 'Upload time', 'align': 'left'},
                             'authors': {'label': 'Authors', 'align': 'left'},
-                            'results.method.simulation.dft.basis_set_name': {'label': 'Basis set name'},
                             'results.method.simulation.dft.xc_functional_type': {'label': 'XC Functional Type'},
-                            'results.method.simulation.precision.apw_cutoff': {'label': 'APW cutoff'},
-                            'results.method.simulation.precision.basis_set': {'label': 'Basis set'},
-                            'results.method.simulation.precision.k_line_density': {'label': 'K-line density'},
-                            'results.method.simulation.precision.native_tier': {'label': 'Native tier'},
-                            'results.method.simulation.precision.planewave_cutoff': {'label': 'Plane wave cutoff'},
+                            'results.method.simulation.precision.apw_cutoff': {'label': 'APW Cutoff'},
+                            'results.method.simulation.precision.basis_set': {'label': 'Basis Set'},
+                            'results.method.simulation.precision.k_line_density': {'label': 'k-line Density'},
+                            'results.method.simulation.precision.native_tier': {'label': 'Code-specific tier'},
+                            'results.method.simulation.precision.planewave_cutoff': {'label': 'Plane-wave cutoff'},
                             'results.material.structural_type': {'label': 'Dimensionality'},
                             'results.material.symmetry.crystal_system': {'label': 'Crystal system'},
                             'results.material.symmetry.space_group_symbol': {'label': 'Space group symbol'},

[…]
