### 2023-06-12 09c0cf4
Adjust matching parser regex for new mopac version
Changelog: Fixed
