### 2023-06-12 553e4e1
Fixed nomad-lab install and ArchiveQuery use.
This moves most deps from nomad-lab[parsing] into nomad-lab.
Now almost all parsers and metainfo can be used from nomad-lab
alone.

This als fixes ArchiveQuery "required" for polymorph sub-sections.

It also reintroduces the old workflow schema to allow backwards compatible
queries for the old workflows.

Changlog: Fixed
