+++ gui/tests/artifacts.js
@@ -13319,7 +13319,7 @@ window.nomadArtifacts = {
                 "m_parent_index": 0,
                 "m_parent_sub_section": "sub_sections",
                 "name": "species",
-                "sub_section": "/packages/20/section_definitions/0",
+                "sub_section": "/packages/21/section_definitions/0",
                 "repeats": true
               },
               {
@@ -16912,7 +16912,7 @@ window.nomadArtifacts = {
                 "description": "Specific heat capacity values at constant volume.",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/5/quantities/5"
+                  "type_data": "/packages/19/section_definitions/5/quantities/5"
                 },
                 "shape": []
               },
@@ -16924,7 +16924,7 @@ window.nomadArtifacts = {
                 "description": "The temperatures at which heat capacities are calculated.",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/5/quantities/1"
+                  "type_data": "/packages/19/section_definitions/5/quantities/1"
                 }
               }
             ]
@@ -16944,7 +16944,7 @@ window.nomadArtifacts = {
                 "description": "The Helmholtz free energies per atom at constant volume.",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/5/quantities/7"
+                  "type_data": "/packages/19/section_definitions/5/quantities/7"
                 },
                 "shape": []
               },
@@ -16956,7 +16956,7 @@ window.nomadArtifacts = {
                 "description": "The temperatures at which Helmholtz free energies are calculated.",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/5/quantities/1"
+                  "type_data": "/packages/19/section_definitions/5/quantities/1"
                 }
               }
             ]
@@ -17043,7 +17043,7 @@ window.nomadArtifacts = {
                 "name": "volumes",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/50/quantities/1"
+                  "type_data": "/packages/19/section_definitions/50/quantities/1"
                 }
               },
               {
@@ -17053,7 +17053,7 @@ window.nomadArtifacts = {
                 "name": "energies_raw",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/50/quantities/2"
+                  "type_data": "/packages/19/section_definitions/50/quantities/2"
                 }
               },
               {
@@ -17063,7 +17063,7 @@ window.nomadArtifacts = {
                 "name": "energies_fit",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/49/quantities/1"
+                  "type_data": "/packages/19/section_definitions/49/quantities/1"
                 }
               }
             ]
@@ -17200,7 +17200,7 @@ window.nomadArtifacts = {
                 "description": "List of energy_total values gathered from the single configuration\ncalculations that are a part of the optimization trajectory.",
                 "type": {
                   "type_kind": "quantity_reference",
-                  "type_data": "/packages/18/section_definitions/12/quantities/1"
+                  "type_data": "/packages/19/section_definitions/12/quantities/1"
                 }
               },
               {
@@ -59798,7 +59798,7 @@ window.nomadArtifacts = {
                 "name": "nomad_data_schema",
                 "type": {
                   "type_kind": "reference",
-                  "type_data": "/packages/21/section_definitions/3"
+                  "type_data": "/packages/22/section_definitions/3"
                 }
               }
             ],
@@ -61630,7 +61630,7 @@ window.nomadArtifacts = {
[…]
