+++ nomad/datamodel/metainfo/simulation/legacy_workflows.py
@@ -1,0 +1,2243 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import numpy as np
+from nptyping import NDArray
+from nomad.metainfo import (
+    MSection, MEnum, Quantity, Section, SubSection, SectionProxy,
+    Reference, derived)
+from nomad.datamodel.metainfo.simulation.calculation import Calculation, Dos, BandStructure
+from nomad.datamodel.metainfo.simulation.run import Run
+from nomad.datamodel.metainfo.simulation.system import System, Atoms, AtomsGroup
+from nomad.datamodel.metainfo.common import FastAccess
+
+
+class Interface(MSection):
+    '''
+    Section containing results of an interface (stacking fault, gamma surface, etc.)
+    workflow.
+    '''
+
+    m_def = Section(validate=False)
+
+    energy_extrinsic_stacking_fault = Quantity(
+        type=np.dtype(np.float64),
+        shape=[],
+        unit='joule / m **2',
+        description='''
+        Value of the relaxed extrinsic stacking fault energy per unit area.
+        ''')
+
+    energy_intrinsic_stacking_fault = Quantity(
+        type=np.dtype(np.float64),
+        shape=[],
+        unit='joule / m **2',
+        description='''
+        Value of the relaxed intrinsic stacking fault energy per unit area.
+        ''')
+
+    dimensionality = Quantity(
+        type=np.dtype(np.int32),
+        shape=[],
+        description='''
+        Dimensionality of the property, i.e. 1 for stacking fault energy and 2 for gamma
+        surface.
+        ''')
+
+    shift_direction = Quantity(
+        type=str,
+        shape=['dimensionality'],
+        description='''
+        shift direction of the two crystal parts to calculate the fault energy.
+        ''')
+
+    n_displacements = Quantity(
+        type=np.dtype(np.int32),
+        shape=[],
+        description='''
+        Number of displacements in the shift to calculate the fault energy.
+        ''')
+
+    displacement_fraction = Quantity(
+        type=np.dtype(np.float64),
+        shape=['dimensionality', 'n_displacements'],
+        description='''
+        Relative displacements of the two crystal parts along the direction indicated by
+        shift_direction.
+        ''')
+
+    energy_fault_plane = Quantity(
+        type=np.dtype(np.float64),
+        shape=['n_displacements'],
+        unit='joule / m ** 2',
+        description='''
+        Value of the relaxed excess energy per unit area for each displacement.
+        ''')
+
+    gamma_surface = Quantity(
+        type=np.dtype(np.float64),
+        shape=['n_displacements', 'n_displacements'],
+        unit='joule / m ** 2',
+        description='''
+        Value of the gamma surface, i.e. the excess energy per unit area calculated for
+        each displacement.
+        ''')
+
+    slip_fraction = Quantity(
+        type=np.dtype(np.float64),
+        shape=[],
+        description='''
+        Relative displacement between two crystal parts where the energy is maximum.
+        ''')
+
+    energy_unstable_stacking_fault = Quantity(
+        type=np.dtype(np.float64),
+        shape=[],
+        unit='joule / m **2',
+        description='''
[…]
