+++ nomad/parsing/tabular.py
@@ -15,12 +15,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-from typing import Union, List, Iterable, Dict, Callable, Set, Any, Tuple, cast, Iterator
+import os
+from typing import List, Dict, Callable, Set, Any, Tuple, Iterator, Union, Iterable, cast
 
 import pandas as pd
 from memoization import cached
-import os.path
 import re
 import math
 import numpy as np
@@ -28,14 +27,16 @@ import json
 import yaml
 
 from nomad import utils
+from nomad.parsing import MatchingParser
 from nomad.units import ureg
-from nomad.datamodel import EntryArchive, EntryData, ArchiveSection
+from nomad.datamodel import EntryArchive, EntryMetadata
+from nomad.datamodel.data import ArchiveSection, EntryData
+from nomad.metainfo import Section, Quantity, Package, Reference, MSection, Property
+from nomad.metainfo.metainfo import MetainfoError, SubSection, MProxy, SectionProxy
 from nomad.datamodel.context import Context
-from nomad.metainfo import Section, Quantity, Package, Reference, SectionProxy, MSection, Property
-from nomad.metainfo.metainfo import MetainfoError, SubSection
-from nomad.parsing.parser import MatchingParser
-from nomad.datamodel.metainfo.annotations import TabularMode, TabularAnnotation, TabularParserAnnotation
+from nomad.datamodel.metainfo.annotations import TabularAnnotation, TabularParserAnnotation
 from nomad.metainfo.util import MSubSectionList
+from nomad.utils import generate_entry_id
 
 # We define a simple base schema for tabular data. The parser will then generate more
 # specialized sections based on the table headers. These specialized definitions will use
@@ -65,16 +66,6 @@ def traverse_to_target_data_file(section, path_list: List[str]):
             raise MetainfoError(f'The path {temp} in path_to_data_file does not exist')
 
 
-def extract_tabular_parser_annotation(archive):
-    for quantity_def in archive.m_def.all_quantities.values():
-        annotation = quantity_def.m_get_annotations('tabular_parser')
-        if not annotation:
-            pass
-        else:
-            return annotation, quantity_def
-    return None, None
-
-
 class TableRow(EntryData):
     ''' Represents the data in one row of a table. '''
     table_ref = Quantity(
@@ -89,8 +80,15 @@ class Table(EntryData):
         description='References that connect to each row. Each row is stored in it individual entry.')
 
 
-class TableData(ArchiveSection):
+class TabularParserError(Exception):
+    ''' Tabular-parser related errors. '''
+    pass
 
+
+class TableData(ArchiveSection):
+    '''
+
+    '''
     fill_archive_from_datafile = Quantity(
         type=bool,
         a_eln=dict(component='BoolEditQuantity'),
@@ -103,11 +101,11 @@ class TableData(ArchiveSection):
         super(TableData, self).normalize(archive, logger)
 
         if self.fill_archive_from_datafile:
-            annotation, quantity_def = extract_tabular_parser_annotation(self)
-            if annotation:
-                if isinstance(annotation, list):
-                    annotation = annotation[0]
-                self.tabular_parser(quantity_def, archive, logger, annotation)
+            for quantity_def in self.m_def.all_quantities.values():
+                annotation = quantity_def.m_get_annotations('tabular_parser')
+                annotation = annotation[0] if isinstance(annotation, list) else annotation
+                if annotation:
+                    self.tabular_parser(quantity_def, archive, logger, annotation)
 
     def tabular_parser(self, quantity_def: Quantity, archive, logger, annotation: TabularParserAnnotation):
         if logger is None:
[…]
