+++ nomad/datamodel/metainfo/annotations.py
@@ -359,21 +359,49 @@ class TabularParserAnnotation(AnnotationModel):
         holds the path to the `.csv` or excel file.
     ''')
     separator: str = Field(None, description='An alias for `sep`')
-    mode: TabularMode = Field(TabularMode.column, description='''
-        Either `column`, `row`, `entry` or `root`. With `column` the whole column is mapped into a quantity
-        (needs to be a list).
-        With `row` each row (and its cells) are mapped into instances of a repeating
-        sub section, where each section represents a row (quantities need to be scalars).
-        With `entry` new entry is created and populated from each row (and its cells) where
-        all quantities should remain to be scalars.
+    column_sections: List[str] = Field([], description='''
+        A `list` of paths to the sub-sections where the tabular quantities are to be filled from the
+        entire column of the excel/csv file (i.e. in the column mode). Each path is a
+        `/` separated list of nested sub-sections. The targeted sub-sections, will be
+        considered when mapping table columns to quantities.
+        Has to be used to annotate the quantity that
+        holds the path to the `.csv` or excel file.
+    ''')
+    row_sections: List[str] = Field([], description='''
+        A `list` of paths to the repeating sub-sections where the tabular quantities are to be filled from individual rows
+        of the excel/csv file (i.e. in the row mode). Each path is a
+        `/` separated list of nested sub-sections. The targeted sub-sections, will be
+        considered when mapping table rows to quantities.
+        Has to be used to annotate the quantity that
+        holds the path to the `.csv` or excel file.
+    ''')
+    entry_sections: List[str] = Field([], description='''
+        A `list` of paths to the (sub)sections where the tabular quantities are to be filled from individual rows
+        of the excel/csv file, to create distinct entries. Each path is a
+        `/` separated list of nested sub-sections. The targeted (sub)sections, will be
+        considered when mapping table rows to quantities. The schema of the resultant entry follows the
+        (sub)section's schema. In order to parse the entire schema using entry mode, then set the
+        first item in this list to `root`.
         Has to be used to annotate the quantity that
         holds the path to the `.csv` or excel file.
     ''')
     target_sub_section: List[str] = Field([], description='''
-        A lists of paths to sub-sections of the annotation quantity's section. Each path is a
+        this feature is deprecated and will be removed in future release. Use `row_sections` instead.
+        A `list` of paths to the repeating sub-sections where the tabular quantities are to be filled from individual rows
+        of the excel/csv file (i.e. in the row mode). Each path is a
         `/` separated list of nested sub-sections. The targeted sub-sections, will be
-        considered when mapping table columns to quantities.
+        considered when mapping table rows to quantities.
         Has to be used to annotate the quantity that
+        holds the path to the `.csv` or excel file.
+    ''')
+    mode: Optional[TabularMode] = Field(TabularMode.column, description='''
+        This is optional. It will be removed in future release.
+        Either `column`, `row`, or `entry`. With `column` the whole column is mapped into a quantity
+        (needs to be a list).
+        With `row` each row (and its cells) are mapped into instances of a repeating
+        sub section, where each section represents a row (quantities need to be scalars).
+        With `entry` new entry is created and populated from each row (and its cells) where
+        all quantities should remain to be scalars. Has to be used to annotate the quantity that
         holds the path to the `.csv` or excel file.
     ''')
 
