### 2023-06-15 1fdc3c1
Cleaning up the method how the version has been determined during the CI/CD build process
Changelog: Fixed
