### 2023-06-15 6b68dc8
Removed the old material_library ELN example build-in schema.
Changelog: Changed
