+++ nomad/datamodel/metainfo/eln/material_library/__init__.py
@@ -1,797 +1,0 @@
-#
-# Copyright The NOMAD Authors.
-#
-# This file is part of NOMAD. See https://nomad-lab.eu for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import numpy as np
-
-from nomad.units import ureg
-from nomad.metainfo import (
-    MSection, Package, Quantity, SubSection, MEnum, Reference, Datetime, Section)
-from nomad.datamodel.data import EntryData, ArchiveSection, ElnExampleCategory
-
-
-m_package = Package(name='material_library')
-
-
-class Chemical(EntryData):
-    '''A chemical available in the lab.'''
-    m_def = Section(label_quantity='formula', categories=[ElnExampleCategory])
-
-    chemical_name = Quantity(type=str, a_eln=dict(component='StringEditQuantity'))
-
-    formula = Quantity(
-        type=str,
-        description='Empirical formula of the chemical (Hill notation).',
-        a_eln=dict(component='StringEditQuantity'))
-
-    form = Quantity(
-        type=MEnum([
-            'solid',
-            'powder',
-            'liquid',
-            'gas']),
-        shape=[],
-        description='Physical state of the chemical.',
-        a_eln=dict(component='RadioEnumEditQuantity'))
-
-    supplier = Quantity(
-        type=str,
-        description='Supplier of the chemical.',
-        a_eln=dict(component='StringEditQuantity'))
-
-    sku_no = Quantity(
-        type=str,
-        description=' Stock keeping unit of the chemical. (e.g. 244619-50G)',
-        a_eln=dict(component='StringEditQuantity'))
-
-    opening_date = Quantity(
-        type=Datetime,
-        description='Opening date of the chemical.',
-        a_eln=dict(component='DateTimeEditQuantity'))
-
-    impurities = Quantity(
-        type=str,
-        description='''Descriptions of the impurities of the product.''',
-        a_eln=dict(component='StringEditQuantity'))
-
-    cas_number = Quantity(
-        type=str,
-        description='''The CAS number is a unique and unambiguous
-                    identifier for a specific substance.''',
-        a_eln=dict(component='StringEditQuantity'))
-
-    sds_link = Quantity(
-        type=str,
-        description='''The corresponding Safety Data Sheet (SDS) of the product.''',
-        a_eln=dict(component='FileEditQuantity'),
-        a_browser=dict(adaptor='RawFileAdaptor'))
-
-    comments = Quantity(
-        type=str,
-        description='''Remarks about the chemical beyond the
-                    typically collected information. Here you can collect information
-                    about your experience using this chemical, for example if it might
-                    have been contaminated during an experiment''',
-        a_eln=dict(component='RichTextEditQuantity'))
-
-    def normalize(self, archive, logger):
-        super(Chemical, self).normalize(archive, logger)
-
-        if self.formula:
-            archive.metadata.entry_name = self.formula
-            if self.chemical_name:
-                archive.metadata.entry_name += f' ({self.chemical_name})'
-
-
-class Maintenance(MSection):
-    m_def = Section(label_quantity='datetime')
-
-    datetime = Quantity(
-        type=Datetime,
-        description='The date and time of the maintenance.',
[…]
