+++ requirements.txt
@@ -65,6 +65,7 @@ email-validator==1.3.0    # via optimade
 emmet-core==0.48.4        # via mp-api
 entrypoints==0.4          # via ipyparallel, jupyter-client, jupyterhub, numcodecs
 escapism==1.0.1           # via dockerspawner
+et-xmlfile==1.1.0         # via openpyxl
 exceptiongroup==1.1.1     # via cattrs
 fastapi==0.92.0           # via nomad-lab (pyproject.toml)
 fasteners==0.18           # via mdanalysis, zarr
@@ -160,6 +161,7 @@ numexpr==2.8.4            # via hyperspy
 numpy==1.22.4             # via ase, biopython, cftime, dask, griddataformats, gsd, h5grove, h5py, hyperspy, ifes-apt-tc-data-modeling, imageio, matid (dependencies/matid/pyproject.toml), matplotlib, mdanalysis, mdtraj, mrcfile, netcdf4, nexusutils (dependencies/parsers/nexus/pyproject.toml), nomad-dos-fingerprints (dependencies/nomad-dos-fingerprints/pyproject.toml), nomad-lab (pyproject.toml), nptyping, numba, numcodecs, numexpr, pandas, phonopy, pymatgen, pyscf, pywavelets, radioactivedecay, scikit-image, scikit-learn, scipy, sparse, spglib, tifffile, xarray, zarr
 oauthenticator==15.1.0    # via nomad-lab (pyproject.toml)
 oauthlib==3.2.2           # via jupyterhub
+openpyxl==3.1.2           # via nomad-lab (pyproject.toml)
 optimade[mongo]==0.22.1   # via nomad-lab (pyproject.toml)
 orjson==3.6.0             # via h5grove, nomad-lab (pyproject.toml)
 packaging==23.0           # via dask, docker, hyperspy, ipykernel, jupyter-server, jupyterlab, jupyterlab-server, matplotlib, mdanalysis, mongomock, nbconvert, pint, scikit-image, sphinx
