### 2023-06-16 55eb608
Added new file formats for the API endpoint used for downloading atomistic structures, added button for downloading a file from the 3D structure visualizer.
Changelog: Added
