+++ nomad/app/v1/routers/systems.py
@@ -16,12 +16,13 @@
 # limitations under the License.
 #
 from typing import Union, Dict, List
-from io import StringIO
+from io import StringIO, BytesIO
 from collections import OrderedDict
 from enum import Enum
 
+import numpy as np
 from fastapi import APIRouter, Depends, Path, Query, status, HTTPException
-from fastapi.responses import StreamingResponse
+from fastapi.responses import Response
 import ase.io
 import ase.build
 from MDAnalysis.lib.util import NamedStream
@@ -29,9 +30,10 @@ from MDAnalysis.coordinates.PDB import PDBWriter
 
 from nomad.units import ureg
 from nomad.utils import strip, deep_get, query_list_to_dict
+from nomad.atomutils import Formula
 from nomad.normalizing.common import ase_atoms_from_nomad_atoms, mda_universe_from_nomad_atoms
 from nomad.datamodel.metainfo.simulation.system import Atoms as NOMADAtoms
-from .entries import answer_entry_archive_request, _bad_id_response
+from .entries import answer_entry_archive_request
 
 from .auth import create_user_dependency
 from ..utils import create_responses
@@ -39,20 +41,139 @@ from ..models import User, HTTPExceptionModel
 
 router = APIRouter()
 default_tag = 'systems'
-format_list: List[dict] = [
-    {
+
+
+def write_pdb(atoms: NOMADAtoms, entry_id: str = None, formula: str = None) -> str:
+    '''For writing a PDB file.'''
+
+    # Add custom title that contains the entry id.
+    lines = []
+    if entry_id is not None:
+        lines.append(f'TITLE     NOMAD ENTRY ID: {entry_id}\n')
+
+    # PDB files do not contain a field for the full lattice vectors and the PBC.
+    # To work around this, they are stored as REMARKs. The info on full lattice
+    # vectors and pbc are used by the the GUI visualizer.
+    lattice_vectors = atoms.lattice_vectors
+    pbc = atoms.periodic
+    if lattice_vectors is not None:
+        cell = lattice_vectors.to(ureg.angstrom).magnitude
+        lines.append('REMARK 285 LATTICE VECTORS\n')
+        lines.append(f'REMARK 285  A: {cell[0, 0]:.3f}, {cell[0, 1]:.3f}, {cell[0, 2]:.3f}\n')
+        lines.append(f'REMARK 285  B: {cell[1, 0]:.3f}, {cell[1, 1]:.3f}, {cell[1, 2]:.3f}\n')
+        lines.append(f'REMARK 285  C: {cell[2, 0]:.3f}, {cell[2, 1]:.3f}, {cell[2, 2]:.3f}\n')
+    else:
+        lines.append('REMARK 285 UNITARY VALUES FOR THE UNIT CELL SET BECAUSE UNIT CELL INFORMATION\n')
+        lines.append('REMARK 285 WAS MISSING. PROTEIN DATA BANK CONVENTIONS REQUIRE THAT CRYST1\n')
+        lines.append('REMARK 285 RECORD IS INCLUDED, BUT THE VALUES ON THIS RECORD ARE MEANINGLESS.\n')
+    if pbc is not None:
+        pbc = ['TRUE' if x else 'FALSE' for x in pbc]
+        lines.append(f'REMARK 285 PBC (A, B, C): {pbc[0]}, {pbc[1]}, {pbc[2]}\n')
+
+    mda_string_stream = StringIO()
+    mda_named_stream = NamedStream(mda_string_stream, f'temp.{format}', close=False, reset=False)
+    writer = PDBWriter(mda_named_stream, remarks='')
+    universe = mda_universe_from_nomad_atoms(atoms)
+    writer.write(universe)
+    writer.close()
+
[…]
