### 2023-06-16 6b98f08
Changed the versions of pyproject.toml dependencies.
We relaxed the version requirements for the basic
package dependenceis. This makes the installation more likely
at the expense of potential  runtime problems.

Changelog: Changed
