+++ pyproject.toml
@@ -13,23 +13,23 @@ dynamic = ["version"]
 license = { file = "LICENSE" }
 requires-python = ">=3.9"
 dependencies = [
-    'numpy==1.22.4',
-    'nptyping==1.4.4',
-    'pandas==1.3.5',
+    'numpy~=1.22.4',
+    'nptyping~=1.4.4',
+    'pandas~=1.3.5',
     'cachetools>=4.2.4',
-    'docstring-parser==0.12',
+    'docstring-parser>=0.12',
     'pint==0.17',
     'orjson==3.6.0',
     'click>=7.1.2',
     'requests>=2.28.2',
     'pytz>=2022.7.1',
     'aniso8601>=7.0.0',
-    'ase==3.19.0',
+    'ase~=3.19.0',
     'python-keycloak>=0.26.1',
     'elasticsearch-dsl==7.4.0',
     'pydantic>=1.10.8',
-    'jmespath==0.10.0',
-    'httpx==0.23.3',
+    'jmespath>=0.10.0',
+    'httpx>=0.23.3',
     'memoization>=0.4.0',
     'rfc3161ng>=2.1.3',
     'lxml~=4.6',
@@ -38,14 +38,14 @@ dependencies = [
     'python-magic==0.4.24',
     'hjson>=3.0.2',
     'scipy>=1.7.1',
-    'panedr==0.2',
-    'parmed==3.0.0',
+    'panedr>=0.2',
+    'parmed>=3.0.0',
     'mdanalysis==2.5.0',
-    'xarray==0.20.2',
-    'phonopy==2.11.0',
-    'pymatgen==2023.2.28',
+    'xarray~=0.20.2',
+    'phonopy~=2.11.0',
+    'pymatgen~=2023.2.28',
     'bitarray>=2.3.5',
-    'scikit-learn==1.0.2',
+    'scikit-learn~=1.0.2',
     'toposort',
     'openpyxl>=3.0.0'
 ]
