### 2023-06-18 0ae2971
Use bulk insert
