+++ gui/tests/artifacts.js
@@ -1618,7 +1618,8 @@ window.nomadArtifacts = {
         "type_data": [
           "parser",
           "user",
-          "matid"
+          "matid",
+          "porosity"
         ]
       },
       "aggregatable": true,
@@ -1990,6 +1991,95 @@ window.nomadArtifacts = {
         "type_data": "int"
       },
       "shape": [],
+      "aggregatable": false
+    },
+    "results.material.topology.sbu_type": {
+      "name": "sbu_type",
+      "description": "The topological representation of the metal secondary building units (sbus).\nThe shape of most metal sbus are well defined and form the basis of most\n popular MOFs. The most common example is the paddlewheel, rodlike mofs,\n irmofs, uio66",
+      "type": {
+        "type_kind": "python",
+        "type_data": "str"
+      },
+      "shape": [],
+      "aggregatable": true,
+      "suggestion": true
+    },
+    "results.material.topology.largest_cavity_diameter": {
+      "name": "largest_cavity_diameter",
+      "description": "The largest cavity diameter is the largest sphere that can be inserted in a porous\nsystem without overlapping with any of the atoms in the system.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter",
+      "aggregatable": false
+    },
+    "results.material.topology.pore_limiting_diameter": {
+      "name": "pore_limiting_diameter",
+      "description": "The pore limiting diameter is the largest sphere that can freely\ndiffuse through the porous network without overlapping with any of the\natoms in the system.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter",
+      "aggregatable": false
+    },
+    "results.material.topology.largest_included_sphere_along_free_sphere_path": {
+      "name": "largest_included_sphere_along_free_sphere_path",
+      "description": "The largest included sphere along free sphere path is\nlargest sphere that can be inserted in the pore.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter",
+      "aggregatable": false
+    },
+    "results.material.topology.accessible_surface_area": {
+      "name": "accessible_surface_area",
+      "description": "The surface area accessible is the area that is accessible to guest molecules\nin a porous system. It is generally considered to be the entire surface area\nthat can be spanned by a probe of a specific radius. In NOMAD, by default we use\na probe that has a radius of 1.86 Angstrom, which correspond to the\ncovalent radii of nitrogen gas. For biomolecular system, a radii of\n1.4 Angstrom can be used, which correspond to the covalent radii\nof water.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter ** 2",
+      "aggregatable": false
+    },
+    "results.material.topology.accessible_volume": {
+      "name": "accessible_volume",
+      "description": "Volume of unoccupied space in a system that can be accessible to\nguest molecules, like solvents.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter ** 3",
+      "aggregatable": false
+    },
+    "results.material.topology.void_fraction": {
+      "name": "void_fraction",
+      "description": "Ratio of the the volume of the unoccupied space in the system\nto the volume of the entire system. It is a good proxy to\ndetermine how porous a system is. Highly porous systems\noften have a larger void fraction, meanwhile compact or dense\nsystems have smaller void fractions.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "aggregatable": false
+    },
+    "results.material.topology.n_channels": {
+      "name": "n_channels",
[…]
