### 2023-06-20 #5388 @zhubonan
Hope everyone is well and not melted by the summer heat (yet) 🥵  Just want to give ping here!
@arosen93 @sgbaird Thanks for reviewing and posting the comments.

@berquist I can see there are still some checked items in the checklist? Do you still have concerns about the author list? Note that we do allow non code-contributor to be on the paper, which is not a problem.

@markus1978 how is it going addressing the comments?
