### 2023-06-20 #5388 @markus1978
@zhubonan Thanks for the pin and thanks for the reviews again.

I did not come up with the "big reply" yet, as I was under the impression that @berquist does still need sometime:
> I've read the paper and believe that NOMAD belongs in JOSS. I'm still distilling my notes on the (website) documentation and haven't tried installing/testing/running other than building the Docker image.

But be assured that we are already working on the raised issues. To keep it brief for now, these are main things we are working on:
- there is a refactor of the whole documentation (this was planned anyways)
- a new release for the nomad-lab pypi package and tests for different environments (python versions, jupyter, colab)

With these two things done, we should have addressed most of what @arosen93 and @sgbaird raised. I will prepare the formal big reply and address the issues in detail, once @berquist signaled that he finished his review.
