### 2023-06-21 24ce707
Resolve "Bug: Gui crashes when emptying quantity field"
