### 2023-06-21 58a213a
Resolve "Adding option to create one new entry for all rows in tabular parser"
