+++ docs/schema/elns.md
@@ -49,6 +49,11 @@ with each comes unique options for importing and interacting with your data. In 
 NOMAD tabular parser to import your data, follow three sections below. In each section you
 can find a commented sample schema with a step-by-step guide on how to import your tabular data.
 
+Tabular parser, implicitly, parse the data into the same NOMAD entry where the datafile is loaded. Also, explicitly, 
+this can be defined by putting the corresponding annotations under `current_entry` (check the examples below).
+In addition, tabular parser can be set to parse the data into new entry (or entries). For this, the proper annotations 
+should be appended to `new_entry` annotation in your schema file.
+
 Two main components of any tabular parser schema are:
 1) implementing the correct base-section(s), and
 2) providing a `data_file` `Quantity` with the correct `m_annotations`. 
@@ -84,8 +89,8 @@ Important notes:
 - `shape: ['*']` under `My_Quantity` is essential to parse the entire column of the data file.
 - The `data_file` `Quantity` can have any arbitrary name (e.g. `xlsx_file`)
 - `My_Quantity` can also be defined within another subsection (see next sample schema)
-- Use `column_sections` to specify which sub_section(s) is to be filled in this mode. `Leaving this field empty`
-causes the parser to parse the entire schema under column mode.
+- Use `current_entry` and append `column_to_sections` to specify which sub_section(s) is to be filled in 
+this mode. `Leaving this field empty` causes the parser to parse the entire schema under column mode.
 
 ```yaml
 --8<-- "examples/data/docs/tabular-parser-col-mode.archive.yaml"
@@ -112,11 +117,10 @@ Important notes:
 
 - This schema demonstrates how to import data within a subsection of another subsection, meaning the
 targeted quantity should not necessarily go into the main `quantites`.
-- Setting `row_sections` signals that for each row in the sheet_name (provided in `My_Quantity`),
+- Setting `row_to_sections` under `current_entry` signals that for each row in the sheet_name (provided in `My_Quantity`),
 one instance of the corresponding (sub-)section (in this example, `My_Subsection` sub-section as it has the `repeats`
 option set to true), will be appended. Please bear in mind that if this mode is selected, then all other quantities
 in this sub_section, should exist in the same sheet_name.
-- Leaving `row_sections` empty causes the parser to parse the entire schema using <b>column mode</b>!
 
 
 ```yaml
@@ -141,9 +145,9 @@ following schema by substituting `My_Quantity` with any appropriate name (e.g. `
 
 Important note:
 
-- To create new entries based on your entire schema, set `entry_sections` to `- root`. Otherwise, you can
+- To create new entries based on your entire schema, set `row_to_entries` to `- root`. Otherwise, you can
 provide the relative path of specific sub_section(s) in your schema to create new entries.
-- Leaving `entry_sections` empty causes the parser to parse the entire schema using <b>column mode</b>!
+- Leaving `row_to_entries` empty causes the parser to parse the entire schema using <b>column mode</b>!
 
 
 ```yaml
@@ -168,7 +172,7 @@ Each individual data_file quantity can now contain a list of sub_sections which 
 using one- or all of the modes mentioned above. Check the `MyOverallSchema` section in
 `Complex Schema` example below. It contains 2 data_file quantities that each one, contains separate instructions
 to populate different parts of the schema. `data_file_1` is responsible to fill `MyColSubsection` while `data_file_2`
-fills all sub_sections listed in `row_sections` and `entry_sections`.
+fills all sub_sections listed in `row_to_sections` and `entry_to_sections` under `new_entry`.
 
 - When using the entry mode, you can create a custom `Quantity` to hold a reference to each new entries
 generated by the parser. Check the `MyEntrySubsection` section in the `Complex Schema` example below.
