### 2023-06-21 #5388 @berquist
I apologize for sitting on what I've done so far and am working on tidying up now.  I probably won't be able to cover everything the way I'd want, but have focused mostly on the Docker things which no one else has brought up yet.

> Do you still have concerns about the author list? Note that we do allow non code-contributor to be on the paper, which is not a problem.

Yes; since non-code contributors are allowed, there are still a few names that seem to have contributed heavily that are not on the list at all.
