### 2023-06-22 374f8c7
Removed MOFStructures parser.
