### 2023-06-22 #5388 @markus1978
> Yes; since non-code contributors are allowed, there are still a few names that seem to have contributed heavily that are not on the list at all.

Of course I am open to address any concerns here. If this is more comfortable, I assume the review process would also allow a private emails thread in this case?
