### 2023-06-22 c77e2bf
1. Foldable list of edit quantities 2. Fixed rawFileAdaptor for array shape
Changelog: Added
