### 2023-06-22 4d00ff9
Fixed issue with repeating data being available for scatter plots, added class...
