+++ gui/src/components/search/Filter.js
@@ -1,0 +1,241 @@
+/*
+ * Copyright The NOMAD Authors.
+ *
+ * This file is part of NOMAD. See https://nomad-lab.eu for further info.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { isNil, isArray, isEmpty } from 'lodash'
+import { searchQuantities } from '../../config'
+import {
+  getDatatype,
+  getSerializer,
+  getDeserializer,
+  formatLabel,
+  DType
+} from '../../utils'
+import { Unit } from '../../units'
+
+/**
+ * Filter is a wrapper for metainfo (quantity or section) that can be searched.
+ * It contains information extracted from the metainfo together with additional
+ * configuration that can e.g. be extracted from the NOMAD configuration file.
+ */
+export class Filter {
+  options
+  aggs
+  value
+  placeholder
+  multiple
+  exclusive
+  unit
+  dtype
+  customSerialization
+  serializerExact
+  serializerPretty
+  dimension
+  deserializer
+  label
+  labelFull
+  nested
+  aggregatable
+  section
+  repeats
+  widget
+  parent
+  schema
+  description
+  scale
+  /**
+  * @param {obj} def Metainfo definition. Optional as filters may not
+  *   correspond to metainfo items.
+  * @param {obj} params Contains configuration for the filter.
+  *  - dtype: The data type for this filter. If no value is given and the
+  *      name corresponds to a metainfo name the data type is read directly from
+  *      the metainfo.
+  *  - description: Description of the filter shown e.g. in the tooltips. If no
+  *      value is given and the name corresponnds to a metainfo, the metainfo
+  *      description is used.
+  *  - unit: The unit for this filter. If no value is given and the name
+  *      corresponds to a metainfo name the data type is read directly from the
+  *      metainfo.
+  *  - label: Short name displayed for this filter. If no value is given and the
+  *      name corresponds to a metainfo name the description is read directly
+  *      from the metainfo.
+  *  - labelFull: Long name displayed for this filter.
+  *  - placeholder: Placeholder displayed for this filter in input fields.
+  *  - multiple: Whether the user can simultaneously provide multiple values for
+  *      this filter.
+  *  - exclusive: Whether this filter is exclusive: only one value is
+  *      associated with a single entry.
+  *  - queryMode: The default query mode (e.g. 'any', 'all) when multiple values
+  *      can specified for this filter. Defaults to 'any'.
+  *  - options: Object containing explicit options that this filter supports.
+  *  - default: A default value which is implicitly enforced in the API call.
+  *      This value will not be serialized in the search bar.
+  *  - scale: The default scaling to perform for filter statistics.
+  *  - value: Optional object containing a custom setter/getter for the filter value.
+  *      Used when transforming the value from GUI to API and vice versa. E.g.
+  *        value = {
+  *          get: (value) => value
+  *          set: (value) => value
+  *        }
+  *  - aggs: Object containing default values for specific aggregation types
+  *      that may be requested for this filter.
[…]
