+++ gui/src/components/search/FilterRegistry.js
@@ -15,18 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { isNil, isArray, isEmpty } from 'lodash'
-import {
-  setToArray,
-  getDatatype,
-  getSerializer,
-  getDeserializer,
-  formatLabel,
-  DType,
-  getSuggestions
-} from '../../utils'
+import { setToArray, DType, getSuggestions } from '../../utils'
 import { searchQuantities } from '../../config'
-import { Unit } from '../../units'
+import { Filter, getEnumOptions } from './Filter'
 import elementData from '../../elementData'
 
 // Containers for filter information
@@ -60,23 +51,6 @@ const idMetadata = 'metadata'
 const idOptimade = 'optimade'
 
 /**
- * Used to gather a list of fixed filter options from the metainfo.
- * @param {string} quantity Metainfo name
- * @returns Dictionary containing the available options and their labels.
- */
-function getEnumOptions(quantity, exclude = ['not processed']) {
-  const metainfoOptions = searchQuantities?.[quantity]?.type?.type_data
-  if (isArray(metainfoOptions) && metainfoOptions.length > 0) {
-    const opt = {}
-    for (const name of metainfoOptions) {
-      opt[name] = {label: name}
-    }
-    exclude.forEach(value => delete opt[value])
-    return opt
-  }
-}
-
-/**
  * This function is used to register a new filter within the SearchContext.
  * Filters are entities that can be searched through the filter panel and the
  * search bar, and can be encoded in the URL. Notice that a filter in this
@@ -95,98 +69,20 @@ function getEnumOptions(quantity, exclude = ['not processed']) {
  * @param {string} name Name of the filter.
  * @param {string} group The group into which the filter belongs to. Groups
  * are used to e.g. in showing FilterSummaries about a group of filters.
- * @param {obj} config Data object containing options for the filter. Can
- * include the following data:
-  *  - aggs: Object containing default values for specific aggregation types
-  *      that may be requested for this filter.
-  *      Also completely customized setter/getter methods are supported. E.g.
-  *        aggs = {
-  *          terms: {size: 5},
-  *          histogram: {buckets: 20},
-  *          min_max: {set: (config) => ({}), get: (agg) => ({})}
-  *        }
-  *  - value: Object containing a custom setter/getter for the filter value.
-  *  - multiple: Whether the user can simultaneously provide multiple values for
-  *      this filter.
-  *  - queryMode: The default query mode (e.g. 'any', 'all) when multiple values
-  *      can specified for this filter. Defaults to 'any'.
-  *  - exclusive: Whether this filter is exclusive: only one value is
-  *      associated with a single entry.
-  *  - widget: Object that determines the default widget for this filter.
-  *  - options: Object containing explicit options that this filter supports.
-  *  - unit: The unit for this filter. If no value is given and the name
-  *      corresponds to a metainfo name the data type is read directly from the
-  *      metainfo.
-  *  - dtype: The data type for this filter. If no value is given and the
-  *      name corresponds to a metainfo name the data type is read directly from
-  *      the metainfo.
-  *  - label: Name of the filter shown in the GUI. If no value is given and the
-  *      name corresponds to a metainfo name the description is read directly
-  *      from the metainfo.
-  *  - description: Description of the filter shown e.g. in the tooltips. If no
[…]
