### 2023-06-22 5d9af6e
The plot axis units should firstly follow the eln.defaultDisplayUnit annotation if exists
Changelog: Fixed
