### 2023-06-23 50d749a
Fix some visual problems
Changelog: Fixed
