+++ docs/reference/parsers.md
@@ -1,5 +1,12 @@
 # Supported parsers
 
-!!! attention
+!!! note
 
-    This part of the documentation is still work in progress.
+    You might also want to read:
+
+    - [How to run parsers locally](../apis/local_parsers.md)
+    - [How to develop a parser plugin](../plugins/parsers.md)
+
+This is a list of all available parsers and supported file formats:
+
+{{ parser_list() }}
