### 2023-06-29 1c53788
Resolve "File preview does not load beyond initially loaded file header"
