### 2023-06-30 5d61bcd
Fixed issue with extra_fields in extracted elabFTW `.eln` file
Changelog: Fixed
