+++ nomad/parsing/elabftw/elabftw.py
@@ -318,7 +318,7 @@ class ELabFTWParser(MatchingParser):
             try:
                 experiment_data.m_update_from_dict(export_data[0])
                 experiment_data.extra_fields = export_data[0]['metadata']['extra_fields']
-            except (IndexError, KeyError):
+            except (IndexError, KeyError, TypeError):
                 logger.warning(f'Couldn\'t read and parse the data from export-elabftw.json file')
             elabftw_experiment.experiment_data = experiment_data
 
