### 2023-07-03 70d1d5a
Changed version requirements on some dependencies.
We loosened some version requirements for key dependencies
to make it easier to install the nomad-lab pypi package.

Changelog: Changed
