+++ requirements.txt
@@ -2,19 +2,18 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --annotation-style=line --extra=infrastructure --extra=parsing --output-file=requirements.txt dependencies/matid/pyproject.toml dependencies/nomad-dos-fingerprints/pyproject.toml dependencies/parsers/atomistic/pyproject.toml dependencies/parsers/database/pyproject.toml dependencies/parsers/eelsdb/pyproject.toml dependencies/parsers/electronic/pyproject.toml dependencies/parsers/nexus/pyproject.toml dependencies/parsers/workflow/pyproject.toml pyproject.toml
+#    pip-compile --annotation-style=line --extra=infrastructure --extra=parsing --output-file=requirements.txt --resolver=backtracking dependencies/matid/pyproject.toml dependencies/nomad-dos-fingerprints/pyproject.toml dependencies/parsers/atomistic/pyproject.toml dependencies/parsers/database/pyproject.toml dependencies/parsers/eelsdb/pyproject.toml dependencies/parsers/electronic/pyproject.toml dependencies/parsers/nexus/pyproject.toml dependencies/parsers/workflow/pyproject.toml pyproject.toml
 #
 alabaster==0.7.12         # via sphinx
 alembic==1.9.1            # via jupyterhub
 amqp==5.1.1               # via kombu
 aniso8601==7.0.0          # via nomad-lab (pyproject.toml)
 anyio==3.6.2              # via httpcore, jupyter-server, starlette, watchfiles
-appnope==0.1.3            # via ipykernel, ipython
 argon2-cffi==21.3.0       # via jupyter-server, nbclassic, notebook
 argon2-cffi-bindings==21.2.0  # via argon2-cffi
 arrow==1.2.3              # via isoduration
 asciitree==0.3.3          # via zarr
-ase==3.19.0               # via asr, ifes-apt-tc-data-modeling, matid (dependencies/matid/pyproject.toml), nexusutils (dependencies/parsers/nexus/pyproject.toml), nomad-lab (pyproject.toml)
+ase==3.19.0               # via asr, ifes-apt-tc-data-modeling, matid (dependencies/matid/pyproject.toml), nomad-lab (pyproject.toml)
 asgiref==3.6.0            # via nomad-lab (pyproject.toml)
 asr==0.4.1                # via nomad-lab (pyproject.toml), workflowparsers (dependencies/parsers/workflow/pyproject.toml)
 astunparse==1.6.3         # via mdtraj
@@ -84,7 +83,7 @@ griddataformats==0.7.0    # via mdanalysis
 gsd==2.7.0                # via mdanalysis
 h11==0.14.0               # via httpcore, uvicorn
 h5grove==1.2.0            # via jupyterlab-h5web, nomad-lab (pyproject.toml)
-h5py==3.6.0               # via electronicparsers (dependencies/parsers/electronic/pyproject.toml), h5grove, hyperspy, ifes-apt-tc-data-modeling, jupyterlab-h5web, nexusutils (dependencies/parsers/nexus/pyproject.toml), nomad-lab (pyproject.toml), phonopy, pyscf, workflowparsers (dependencies/parsers/workflow/pyproject.toml)
+h5py==3.6.0               # via electronicparsers (dependencies/parsers/electronic/pyproject.toml), h5grove, hyperspy, ifes-apt-tc-data-modeling, jupyterlab-h5web, nomad-lab (pyproject.toml), phonopy
 hjson==3.0.2              # via nomad-lab (pyproject.toml)
[…]
