### 2023-07-03 b0d5905
Added more documentation on data, schemas, and plugins.
Changelog: Added
