+++ docs/reference/plugins.md
@@ -1,0 +1,9 @@
+# Built-in plugins
+
+!!! note
+
+    You might also want to read [the plugin how-tos](../plugins/plugins.md)
+
+This is a list of all built-in plugins:
+
+{{ plugin_list() }}
