### 2023-07-05 cb14c69
Add service "logtransfer" to submit logs generated at a NOMAD OASIS to the central unit (opt-in).
Changelog: New feature
