+++ nomad/app/v1/main.py
@@ -28,7 +28,7 @@ from nomad import config, utils
 from .common import root_path
 from .routers import (
     users, entries, materials, auth, info, datasets, uploads, suggestions, metainfo,
-    north, systems, graph
+    north, systems, federation, graph
 )
 
 logger = utils.get_logger(__name__)
@@ -90,6 +90,7 @@ async def unicorn_exception_handler(request: Request, e: Exception):
 
 app.include_router(info.router, prefix='/info')
 app.include_router(auth.router, prefix='/auth')
+app.include_router(federation.router, prefix='/federation')
 app.include_router(materials.router, prefix='/materials')
 app.include_router(entries.router, prefix='/entries')
 app.include_router(datasets.router, prefix='/datasets')
