+++ nomad/normalizing/soap.py
@@ -1,0 +1,159 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import numpy as np
+
+try:
+    from quippy import descriptors
+except ImportError:
+    descriptors = None
+
+from nomad.datamodel.metainfo.simulation.system import Descriptors, SOAP
+
+from .normalizer import SystemBasedNormalizer
+
+
+class SoapNormalizer(SystemBasedNormalizer):
+    def normalize_system(self, system, is_representative):
+        # Only store SOAP for representative system to start with
+        if not is_representative:
+            return True
+
+        # structures used to be stored in results, adding them back like this for now...
+        if system.atoms is None:
+            return False
+
+        if not descriptors:
+            self.logger.warning('SOAP normalizer runs, but quippy is not installed.')
+            return False
+
+        # TODO compute descriptors for primitive system, need to discuss how to store this stuff.
+        # repr_symmetry = system.symmetry[0]
+        # print(repr_symmetry)
+        # symmetry_analyzer = repr_symmetry.m_cache.get("symmetry_analyzer")
+        # prim_atoms = symmetry_analyzer.get_primitive_system()
+        # print("prim_atoms is", type(prim_atoms), prim_atoms)
+
+        soap = SOAP()
+        atoms = system.atoms.to_ase(raise_exp=True)
+
+        # setup params to be used by quippy
+        params, _ = dataset_to_params(atoms)
+        N, S, L = params["n_max"], params["n_Z"], params["l_max"]
+        soap.n_max = N
+        soap.l_max = L
+        soap.r_cut = np.float64(params["soap cutoff"])
+        soap.atom_sigma = np.float64(params["atom_sigma"])
+
+        # regular soap
+        quippy_str = params_to_quippy_str(params)
+        desc = descriptors.Descriptor(quippy_str)
+        output = desc.calc(atoms)
+        ps_array_list = [flat_to_array(p, N, L, S) for p in output["data"]]
+        soap.soap = np.array(
+            [np.reshape(d, (S, S, -1)) for d in ps_array_list]
+        )
+
+        # global soap
+        quippy_str += " average=T"
+        desc = descriptors.Descriptor(quippy_str)
+        output = desc.calc(atoms)
+        A = flat_to_array(output["data"][0], N, L, S)
+        soap.global_soap = np.reshape(A, (S, S, -1))
+
+        # tensor reduced soap
+        quippy_str = params_to_quippy_str(params)
+        quippy_str += "sym_mix=F Z_mix=T R_mix=T K=50 coupling=F"
+        desc = descriptors.Descriptor(quippy_str)
+        output = desc.calc(atoms)
+        soap.tr_soap = output["data"]
+
+        # gloabl tensor reduced soap
+        quippy_str += " average=T"
+        desc = descriptors.Descriptor(quippy_str)
+        output = desc.calc(atoms)
+        soap.global_tr_soap = output["data"][0]
+
+        # add the soap descriptor to the system
[…]
