### 2023-07-08 f7f01bd
Fixed indexing of NaN in dynamic search properties.
Changelog: Changed
