### 2023-07-08 25de71a
Added new archive config options to helm chart.
Changelog: Added
