### 2023-07-08 73cdacf
Added helm config options for plugins and normalize.
Changelog: Added
