+++ pyproject.toml
@@ -41,7 +41,7 @@ dependencies = [
     'panedr>=0.2',
     'parmed>=3.0.0',
     'mdanalysis==2.5.0',
-    'xarray~=0.20.2',
+    'xarray>=0.20.2',
     'phonopy~=2.11.0',
     'pymatgen~=2023.2.28',
     'bitarray>=2.3.5',
