+++ docs/develop/contrib.md
@@ -1,0 +1,282 @@
+!!! note
+
+    The NOMAD source-code is maintained in two synchronized projects on [GitHub](https://github.com/nomad-coe/nomad) and
+    a [GitLab run by MPCDF](https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR).
+    Everyone can contribute on GitHub. The GitLab requires an account for active contribution.
+    This not an ideal situation: there are historic reasons and there is
+    a lot of buy-in into the GitLab CI/CD system. This guide addresses contributions
+    to both projects.
+
+## Issues
+
+### Issue trackers
+
+Everyone can open an issue in our main [GitHub project](https://github.com/nomad-coe/nomad):
+
+- [https://github.com/nomad-coe/nomad/issues/new](https://github.com/nomad-coe/nomad/issues/new)
+
+Use issues to ask questions, report bugs, or suggest features. If in doubt, use the main project
+to engage with us. If you address a specific
+plugin (e.g. parser), you can also post into the respective projects. See also the
+list of [parsers](../reference/parsers.md) and list of [built-in plugins](../reference/plugins.md).
+
+If you are a member of FAIRmat, the NOMAD CoE, or are a close collaborator, you
+probably have an MPCDF GitLab account (or should ask us for one). Please use the
+issue tracker on our main [GitLab project](https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR).
+This is where most of the implementation work is planned and executed.
+
+### Issue content
+A few tips that will help us to solve your issues quicker:
+
+- Focus on the issue. You don't need to greet us or say thx and goodbye. Let's keep it technical.
+- Use descriptive short issue titles. Use specific words over general words.
+- Describe the observed problem and not the assumed causes. Clearly separate speculation from the
+problem description.
+- **Bugs**: think how we could re-produce the problem:
+
+    - What nomad URL are you using (UI), which package version (Python)?
+    - Is there an upload or entry id that we can look at?
+    - Example files or code snippets?
+    - Don't screenshot code, copy&paste please.
+
+- **Features**: augment your feature descriptions with a use-case that helps us understand the feature and its scope.
+
+### Issues labels (GitLab)
+
+On the main GitLab project, there are three main categories for labels. Ideally each issue gets one of each category:
+
+- The *state* labels (grey). Those are used to manage when and how the issue is addressed.
+This should be given the NOMAD team member who is currently responsible to moderate the
+development. If its a simple fix and you want to do it yourself, assign yourself and use
+"bugfixes".
+
+- The *component* label (purple). These denote the part of the NOMAD software that is
+most likely effected. Multiple purple labels are possible.
+
+- The *kind* label (red). Wether this is a bug, feature, refactoring, or documentation issue.
+
+Unlabeled issues will get labeled by the NOMAD team as soon as possible. You can provide
+labels yourself.
+
+## Documentation
+
+The documentation is part of NOMAD's main source code project. You can raise issues
+about the documentation in the usual way (see [above](#issues)). To make changes,
+create a [merge](#merge-requests-gitlab) or [pull](#pull-requests-github) request.
+
+See also the [documentation part](./code.md#documentation) in our code navigation guide.
+
+## Plugins
+
+Also read the [how-to develop, publish, and distribute plugins guide](../plugins/plugins.md).
+
+### Built-in plugins (and submodules)
+
+Most plugins that are maintained by the NOMAD team are built-in plugins (e.g. all the parsers).
+These plugins are also available on the public NOMAD service.
+
+These plugins are tight to the main project's source-code via submodules. They are
+included in the built and therefore automatically distributed as part of the
+NOMAD docker images and Python package.
+
[…]
