+++ docs/oasis/customize.md
@@ -5,7 +5,7 @@
 This is an incomplete list of potential customizations. Please read the respective
 guides to learn more.
 
-- Installation specific changes (domain, path-prefix): [How-to install an Oasis](install.md)
+- Installation specific changes (domain, path-prefix): [How to install an Oasis](install.md)
 - [Restricting user access](admin.md#restricting-access-to-your-oasis)
 - Write .yaml based [schemas](../schemas/basics.md) and [ELNs](../schemas/elns.md)
 - Learn how to use the [tabular parser](../schemas/tabular.md) to manage data from .xls or .csv
