+++ docs/reference/parsers.md
@@ -4,8 +4,8 @@
 
     You might also want to read:
 
-    - [How-to run parsers locally](../apis/local_parsers.md)
-    - [How-to develop a parser plugin](../plugins/parsers.md)
+    - [How to run parsers locally](../apis/local_parsers.md)
+    - [How to develop a parser plugin](../plugins/parsers.md)
 
 This is a list of all available parsers and supported file formats:
 
