### 2023-07-12 cb7326b
Added compression suport to quantum expresso mainfiles.
Changelog: Added
