### 2023-07-12 nomad-coe/nomad#68 @markus1978
We just released nomad-lab==1.2.0 to PyPI. This should solve this issue. Please re-open, if you still run into problems.
