### 2023-07-12 035e1f4
Fixed the missing call to generate_docs_artifacts in the setup_dev_env.sh.
Fixes https://github.com/nomad-coe/nomad/issues/67.

Changelog: Fixed
