### 2023-07-12 nomad-coe/nomad#70 @markus1978
We now released nomad-lab==1.2.0 to PyPI. This package can be installed to google codelab without any of the tricks above. Here is an example notebook: https://colab.research.google.com/drive/1Zlf24eAxFiqtIGbp5NLm66vQR8QPt3AM?usp=sharing.
