### 2023-07-12 b41131f
Removed outdated information about using parsers from the parser projects.
Changelog: Changed
