### 2023-07-12 nomad-coe/nomad#63 @markus1978
We published nomad-lab==1.2.0 to PyPI. The imports and dependencies have been fixed. The requirements are much more unrestrictive and should make it easier to install the package in different environments. There also have been updates to the documentation. The guide to use parser can be found here: https://nomad-lab.eu/prod/v1/staging/docs/apis/local_parsers.html
