### 2023-07-12 #5388 @markus1978
Thank you again for the reviews and all your efforts. Below I aggregate your points and provide a respective response for each of them.

### Open bug reports

You raised these issues:

- The installation seems to be broken. There are open bug reports on the NOMAD main repository: [https://github.com/nomad-coe/nomad/issues/68](https://github.com/nomad-coe/nomad/issues/68)
- There is a missing link in the documentation: [https://github.com/nomad-coe/nomad/issues/67](https://github.com/nomad-coe/nomad/issues/67)
- `nomad-lab` cannot be pip installed or used on google colab:  [https://github.com/nomad-coe/nomad/issues/70](https://github.com/nomad-coe/nomad/issues/70)
- There are too many dependencies.

We released a new version of our Python package `nomad-lab` and provided an update to the documentation. We now have a CI/CD job that tests the installation and basic usage of the package. We also confirmed that it can be installed on google colab ([example notebook](https://colab.research.google.com/drive/1Zlf24eAxFiqtIGbp5NLm66vQR8QPt3AM?usp=sharing)). Furthermore, the new version uses version ranges on its dependencies decreasing the chance of conflicts. We also restructured the imports and "extra" dependencies: now a basic pip install will provide everything needed for most functionalities (ArchiveQuery and most parsers).

On the number of dependencies. Unfortunately, splitting them up and pushing them into the parsers is harder than one might think. Most of the dependencies that we require, even for the basic installation, are used by the definition of our shared schema already and not just by the parsers. This is a challenge for us in general, especially since the project is still growing.  We now introduced a plugin mechanism to mediate this problem. We will see in the upcoming months, when we apply this mechanism to our parsers and schemas, if and how we can manage dependencies better. We did the little we could for now. I went through the imports and dependencies and optimised what was possible. We also lowered the requirements on the dependency versions.

### Documentation

The reviewers pointed out that:

-   The codebase and its multiple repositories are difficult to navigate.
-   Part of the API documentation is missing.
-   There's a lack of documentation on running unit tests locally.
-   A better contributing guide would be helpful.

In response to these points, we fully restructured [the documentation]([https://nomad-lab.eu/docs](https://nomad-lab.eu/docs)) ground up. It now contains categorised how-to guides (on data management, programming interfaces, development, etc.).

This includes a new guide on [how to navigate the code]([https://nomad-lab.eu/prod/v1/staging/docs/develop/code.html](https://nomad-lab.eu/prod/v1/staging/docs/develop/code.html)). This guide explains the projects and organisations. It also gives a few notes on the most important modules and directories. It also shows where to find the documentation.

It also includes a new more detailed [contributing guide]([https://nomad-lab.eu/prod/v1/staging/docs/develop/contrib.html](https://nomad-lab.eu/prod/v1/staging/docs/develop/contrib.html)). It is more structured after your suggestions. Here we also break down the GitHub vs GitLab topic, how the flow is for external contributors, how to test locally, etc.

Our [developers guide]([https://nomad-lab.eu/prod/v1/staging/docs/develop/setup.html](https://nomad-lab.eu/prod/v1/staging/docs/develop/setup.html)) (also linked from the contributing guide) explains how to run backend and frontend tests locally, how to configure your IDE (or what commands to run) to enforce style checks, linting etc.

There are also guides on [installing `nomad-lab`]([https://nomad-lab.eu/prod/v1/staging/docs/apis/pythonlib.html](https://nomad-lab.eu/prod/v1/staging/docs/apis/pythonlib.html)), [running parsers]([https://nomad-lab.eu/prod/v1/staging/docs/apis/local_parsers.html](https://nomad-lab.eu/prod/v1/staging/docs/apis/local_parsers.html)) from command line or code, how to use the [REST API]([https://nomad-lab.eu/prod/v1/staging/docs/apis/api.html](https://nomad-lab.eu/prod/v1/staging/docs/apis/api.html)) with examples (in addition to the API dashboards). However, we do not have a full reference of all modules, classes, and functions. This is a conscious decision. We had this in the beginning, but forcing your documentation structure to fit the code structure was very confusing, and dedicated tutorials/guides proved to be more useful. There are docstring throughout the code, but this is intended for navigating the code and IDE intelli-sense. However, we added a reference section to the documentation with all config settings, schema annotations, available parsers and plugins, and all cli commands and arguments.

The respective parts of the documentation are linked from the README.md, CONTRIBUTING.md, and of course, on our homepage. We also managed to work with google correctly and a search like "how to run nomad-lab unit tests" should get you to the right docs.

### The paper itself

We changed the line (in the paper on github) about OPTIMADE to:

> [OPTIMADE]([https://www.optimade.org/](https://www.optimade.org/)) [@andersen:2021] is an API specification for materials science databases. NOMAD provides an implementation of the OPTIMADE specification and is an active part of the OPTIMADE consortium.

Currently there are ~45k entries from experiments in NOMAD representing a solar-cell and electron loss spectroscopy database. We expect that the specific number and magnitude of experiment entries is likely to change as there are a few other experimental databases that are currently developed and integrated. Also most of the non theory work is happening in NOMAD Oasis installations at (also an increasing number of) groups and research institutes. Currently we have ~20 of these installations.

I am not sure, if we can add more information to the paper due to size constraints (we a ~1000 words of the 250-1000 words JOSS margin)? I agree that many other systems could have been mentioned. Here are our thoughts on the selection.

We focused on materials science-specific systems. Of course, there are tons of other solutions that are interesting. We tried to have at least one example in each materials science-specific category: high throughput database, experimental database, workflow system, data standard, and programming library.

Regarding the comments about referencing other centralized databases and infrastructure where users can contribute materials science-related data, we would prefer not to extend more given the large and growing amount of potential candidates for this list. Beyond the MPContribs and Foundry examples highlighted by @sgbaird, one could extend the list to many more like the ICSD, the Chemotion Archive, or the [refractiveindex.info](http://refractiveindex.info/) to give some examples. It would feel unfair to include some and not others and we feel it would possibly be better addressed in a topical review paper, with the right scope and without being bound to the word limit recommended by JOSS.

We also opted for leaving out databases in the SQL/mongodb sense. Those are not specific to research data management (RDM) or any specific field, let alone materials science. Of course they are often the underpinning of systems like ours (we are using mongodb and elasticsearch under the hood) and you might use them to implement an RDM solution, but we focused on end-user (researcher) focused examples in the list of references.

MPDS is a nice database (based on paper published data, The Pauling file). But only a specific part of the data is open, the rest is commercialised. Of course, this is not really a criterion for not being mentioned, but the other DBs referenced are all open. We could mention so many more databases, these are just examples. Btw. Evgeny Blokhin is/was involved with this and he is active in OPTIMADE as well.

Regarding the question of when and how to connect databases: in principle, we are very open to collaborations and integrating with other systems. It all comes down to time and resources. For example there are the mentioned collaborations with AFLOW, OQMD, Materials Project (MP). Ideally,common interfaces are used (e.g. same file formats). This is why we are active in OPTIMADE. Another strategy for us in the future will be representing NOMAD as linked data (semantic web, RDF, etc.). In principle, users can add references to their data. We could think about formalizing relationships to MP-ids (or entries in other databases). This could help to integrate with platforms like MPContribs and foster more links between data.

### Conclusions

In conclusion, we hope we managed to address all your major points, particularly the installation and documentation problems, which have had many significant changes and improvements. With this, we hope we have also addressed all the JOSS-review-related issues. Even if our job does not fully finish here, we hope that these major changes make our code in good shape to be accepted for publication in JOSS.  Your feedback helped us very much in shaping the new documentation and pushing us towards a more usable Python package, which without a doubt, will keep on improving organically.
