### 2023-07-12 #5388 @arosen93
@markus1978: Thank you and the team for this very thorough response and the absolutely fantastic changes made to the code and documentation. It is much easier to navigate now, and my hope is that this will greatly increase the impact of the hard work you all have put into this. I am delighted to say that all my concerns have been addressed, and I wholeheartedly recommend this work for acceptance in JOSS. I look forward to using NOMAD in my future research group (including OASIS!), and I thank you for the opportunity to review this nice work.
