### 2023-07-12 nomad-coe/nomad#75 @markus1978
Please not that you most likely don't need the `[parsing]` extra. Most parsers will work just fine without. The parsing extras are only required for hfd5 and netCDF based codes and some optional normalising functionality like the new SOAP descriptor.

Unfortunately, there is a limit to what we can do here. The version restrictions we removed, are already a big trade-off between easy installs and the risks of new dependency versions introducing breaking changes. For example you a couple a days ago, the pydantic 2.0.0 release was breaking our build because we removed version restriction. We removed most version restrictions on the normal install, but i am afraid we need to keep them for the extra dependencies that are most likely only affecting the service operations anyways.
