### 2023-07-12 nomad-coe/nomad#75 @Andrew-S-Rosen
@markus1978 --- thanks, that makes sense! Anyway, the more important was more about just letting you all know that the `pip install` isn't working as expected, but that will presumably be resolved once the netCDF version is changed.
