### 2023-07-12 nomad-coe/nomad#75 @markus1978
Of course, i just wanted to explain why we have not just removed every version restriction. We will test  and see what happens if we pump the netCDF version.
