### 2023-07-18 651a2f3
Fixed issue with row action visuals.
Changelog: Fixed
