### 2023-07-19 1a68e76
Resolve "Refactoring tabular parser"
Refactored tabular parser. Entry mode improved to include creation of either one new entry or multiple new entries per data file. Also added the option to create a file on creating one new entry.

Changelog: Added
