+++ gui/src/components/DataStore.js
@@ -441,24 +441,29 @@ const DataStore = React.memo(({children}) => {
           const isJson = fileName.endsWith('json')
           if (isYaml || isJson) {
             const path = `/uploads/${archive.metadata.upload_id}/raw/${fileName}`
-            const content = await api.get(path,
+            let content
+            try {
+              content = await api.get(path,
               {decompress: true, ignore_mime_type: true}, {transformResponse: []})
-            let rawArchive = {}
-            if (isYaml) {
-              try {
-                rawArchive = YAML.parse(content)
-              } catch (error) {
-                throw Error(`Could not parse YAML Schema defined here: ${path}. Failed with the following parsing error: ${error.message}`)
+            } catch (e) {}
+            if (content) {
+              let rawArchive = {}
+              if (isYaml) {
+                try {
+                  rawArchive = YAML.parse(content)
+                } catch (error) {
+                  throw Error(`Could not parse YAML Schema defined here: ${path}. Failed with the following parsing error: ${error.message}`)
+                }
+              } else if (isJson) {
+                try {
+                  rawArchive = JSON.parse(content)
+                } catch (error) {
+                  throw Error(`Could not parse JSON Schema defined here: ${path}. Failed with the following parsing error: ${error.message}`)
+                }
               }
-            } else if (isJson) {
-              try {
-                rawArchive = JSON.parse(content)
-              } catch (error) {
-                throw Error(`Could not parse JSON Schema defined here: ${path}. Failed with the following parsing error: ${error.message}`)
+              if ('definitions' in rawArchive) {
+                dataToUpdate.originalDefinition = rawArchive.definitions
               }
-            }
-            if ('definitions' in rawArchive) {
-              dataToUpdate.originalDefinition = rawArchive.definitions
             }
           }
           archive.processing_logs = undefined
