+++ examples/data/docs/tabular-parser-row-mode.archive.yaml
@@ -12,10 +12,13 @@ definitions:
           type: str
           m_annotations:
             tabular_parser:
-              current_entry:
-                row_to_sections: # This is the reference to where the targeted (sub-)section lies within this example schema file
-                  - My_Subsection/My_Section
-              comment: '#' # Skipping lines in csv or excel file that start with the sign `#`
+              parsing_options:
+                comment: '#' # Skipping lines in csv or excel file that start with the sign `#`
+              mapping_options:
+                - mapping_mode: row
+                  file_mode: current_entry
+                  sections:
+                    - My_Subsection/My_Section
       sub_sections:
         My_Subsection:
           section:
