### 2023-07-20 ad3a56f
Refactored periodic table componeng using SVG.
Changelog: Changed
