+++ nomad/datamodel/metainfo/simulation/workflow.py
@@ -29,10 +29,11 @@ from nomad.datamodel.metainfo.common import FastAccess
 from nomad.datamodel.metainfo.workflow import Workflow, Link, Task
 from nomad.datamodel.metainfo.simulation.system import System, AtomsGroup
 from nomad.datamodel.metainfo.simulation.method import (
-    Method, XCFunctional, BasisSetContainer, GW as GWMethodology,
+    Method, XCFunctional, BasisSetContainer, GW as GWMethodology, Projection as ProjectionMethodology,
+    DMFT as DMFTMethodology
 )
 from nomad.datamodel.metainfo.simulation.calculation import (
-    Calculation, BandGap, Dos, BandStructure, BandEnergies, Density, Potential, Spectra,
+    Calculation, BandGap, Dos, BandStructure, BandEnergies, Density, Potential, Spectra, GreensFunctions,
     RadiusOfGyration as RadiusOfGyrationCalculation,
     RadiusOfGyrationValues as RadiusOfGyrationValuesCalculation)
 from nomad.atomutils import archive_to_universe
@@ -2352,42 +2353,42 @@ class GWResults(SimulationWorkflowResults):
         type=Reference(BandGap),
         shape=['*'],
         description='''
-        DFT band gap
+        DFT band gap.
         ''')
 
     band_gap_gw = Quantity(
         type=Reference(BandGap),
         shape=['*'],
         description='''
-        GW band gap
+        GW band gap.
         ''')
 
     dos_dft = Quantity(
         type=Reference(Dos),
         shape=['*'],
         description='''
-        DFT density of states
+        Ref to the DFT density of states.
         ''')
 
     dos_gw = Quantity(
         type=Reference(Dos),
         shape=['*'],
         description='''
-        GW density of states
+        Ref to the GW density of states.
         ''')
 
     band_structure_dft = Quantity(
         type=Reference(BandStructure),
         shape=['*'],
         description='''
-        DFT density of states
+        Ref to the DFT band structure.
         ''')
 
     band_structure_gw = Quantity(
         type=Reference(BandStructure),
         shape=['*'],
         description='''
-        DFT density of states
+        Ref to the GW band structure.
         ''')
 
 
@@ -2504,28 +2505,28 @@ class PhotonPolarization(ParallelSimulation):
 
 class ParticleHoleExcitationsResults(SimulationWorkflowResults):
 
-    dos_dft = Quantity(
-        type=Reference(Dos),
-        description='''
-        DFT density of states
-        ''')
-
-    dos_gw = Quantity(
-        type=Reference(Dos),
-        description='''
-        GW density of states
-        ''')
-
     band_structure_dft = Quantity(
         type=Reference(BandStructure),
         description='''
-        DFT density of states
+        Ref to the DFT band structure.
         ''')
 
     band_structure_gw = Quantity(
         type=Reference(BandStructure),
         description='''
-        DFT density of states
+        Ref to the GW band structure.
+        ''')
+
+    dos_dft = Quantity(
+        type=Reference(Dos),
+        description='''
+        Ref to the DFT density of states.
+        ''')
+
+    dos_gw = Quantity(
+        type=Reference(Dos),
+        description='''
+        Ref to the GW density of states.
         ''')
 
     spectra = SubSection(sub_section=PhotonPolarizationResults, repeats=True)
@@ -2558,6 +2559,137 @@ class ParticleHoleExcitations(SerialSimulation):
             # link results also to last task
             if self.tasks:
                 self.tasks[-1].inputs.append(Link(name=_workflow_results_name, section=self.results))
+
+
+class DMFTResults(SimulationWorkflowResults):
+
+    band_gap_dft = Quantity(
+        type=Reference(BandGap),
+        shape=['*'],
+        description='''
+        DFT band gap.
+        ''')
+
[…]
