+++ tests/normalizing/test_workflow.py
@@ -87,6 +87,37 @@ def test_gw_workflow(gw_workflow):
     assert results.properties.electronic.dos_electronic[1].label == 'GW'
 
 
+def test_dmft_workflow(dmft_workflow):
+    """Testing DMFT workflow entry"""
+    workflow = dmft_workflow.workflow2
+    assert workflow.name == 'DMFT'
+    assert not workflow.method.projection_method_ref.wannier.is_maximally_localized
+    assert workflow.method.dmft_method_ref.n_atoms_per_unit_cell == 1
+    assert workflow.method.dmft_method_ref.n_correlated_orbitals[0] == 3
+    assert workflow.method.dmft_method_ref.n_correlated_electrons == 1.0
+    assert workflow.method.dmft_method_ref.inverse_temperature.magnitude == 60.0
+    assert workflow.method.dmft_method_ref.magnetic_state == 'paramagnetic'
+    assert workflow.method.dmft_method_ref.impurity_solver == 'CT-HYB'
+    results = dmft_workflow.results
+    assert results.method.method_name == 'DMFT'
+    assert results.method.workflow_name == 'DMFT'
+    assert results.method.simulation.program_name == 'w2dynamics'
+    assert results.method.simulation.dmft.impurity_solver_type == 'CT-HYB'
+    assert results.method.simulation.dmft.total_filling == 0.5 / 3.0
+    assert results.method.simulation.dmft.inverse_temperature.magnitude == 60.0
+    assert results.method.simulation.dmft.magnetic_state == 'paramagnetic'
+    assert results.method.simulation.dmft.u.magnitude == 4.0e-19
+    assert results.method.simulation.dmft.jh.magnitude == 0.6e-19
+    assert results.m_xpath('properties.electronic.band_gap')
+    assert len(results.properties.electronic.band_gap) == 1
+    assert results.properties.electronic.band_gap[0].label == 'PROJECTION'
+    assert results.m_xpath('properties.electronic.band_structure_electronic')
+    assert len(results.properties.electronic.band_structure_electronic) == 1
+    # TODO check why this testing is not passing
+    #   * conftest seems to not be able to normalize the archive_dmft for the Greens functions, despite self_energy_iw is defined.
+    # assert results.m_xpath('properties.electronic.greens_function_electronic')
+
+
 def test_geometry_optimization_workflow(workflow_archive):
     vasp_archive = workflow_archive(
         'parsers/vasp', 'tests/data/normalizers/workflow/vasp/vasprun.xml')
