### 2023-07-20 1e47e7e
Fixed issue with reading boolean values from VASP INCAR and OUTCAR files.
Changelog: Fixed
