### 2023-07-23 1ea52c6
Fixed issue with loading the custom quantities menu.
Changelog: Fixed
