### 2023-07-24 d97d90e
Added better search filters when searching for custom schemas.
Changelog: Added
