+++ gui/src/components/search/FilterRegistry.js
@@ -318,6 +318,8 @@ registerFilter('entry_id', idMetadata, termQuantity)
 registerFilter('entry_name', idMetadata, termQuantity)
 registerFilter('mainfile', idMetadata, termQuantity)
 registerFilter('upload_id', idMetadata, termQuantity)
+registerFilter('published', idMetadata, termQuantity)
+registerFilter('main_author.user_id', idMetadata, termQuantity)
 registerFilter('quantities', idMetadata, {...noAggQuantity, label: 'Metainfo definition', queryMode: 'all'})
 registerFilter('sections', idMetadata, {...noAggQuantity, label: 'Metainfo sections', queryMode: 'all'})
 registerFilter('section_defs.definition_qualified_name', idMetadata, {...noAggQuantity, label: 'Section defs qualified name', queryMode: 'all'})
