+++ gui/src/components/search/SearchContext.js
@@ -315,6 +315,7 @@ export const SearchContext = React.memo(({
     useResetFilters,
     useUpdateQueryString,
     queryState,
+    dynamicQueryModesState,
     aggsFamily,
     aggsState,
     paginationState,
@@ -343,6 +344,10 @@ export const SearchContext = React.memo(({
       key: `queryFamily_${contextID}`,
       default: (name) => initialQuery[name]
     })
+    const dynamicQueryModesFamily = atomFamily({
+      key: `dynamicQueryModesFamily_${contextID}`,
+      default: (name) => undefined
+    })
     // Used to get/set the state of all filters at once
     const filtersState = selector({
       key: `filtersState_${contextID}`,
@@ -354,8 +359,9 @@ export const SearchContext = React.memo(({
         }
         return query
       },
-      set: ({set}, [key, value]) => {
+      set: ({set}, [key, value, queryMode]) => {
         set(queryFamily(key), value)
+        if (queryMode) set(dynamicQueryModesFamily(key), queryMode)
       }
     })
 
@@ -401,6 +407,33 @@ export const SearchContext = React.memo(({
         if (data) {
           for (const [key, value] of Object.entries(data)) {
             set(queryFamily(key), value)
+          }
+        }
+      }
+    })
+
+    /**
+     * A Recoil.js selector that return the dynamic query modes state
+     */
+    const dynamicQueryModesState = selector({
+      key: `dynamicQueryModes_${contextID}`,
+      get: ({get}) => {
+        const query = {}
+        for (const key of filters) {
+          const filter = get(dynamicQueryModesFamily(key))
+          if (filter !== undefined) {
+            query[key] = filter
+          }
+        }
+        return query
+      },
+      set: ({ set }, data) => {
+        for (const filter of filters) {
+          set(dynamicQueryModesFamily(filter), undefined)
+        }
+        if (data) {
+          for (const [key, value] of Object.entries(data)) {
+            set(dynamicQueryModesFamily(key), value)
           }
         }
       }
@@ -657,6 +690,7 @@ export const SearchContext = React.memo(({
       const subname = useMemo(() => section ? name.slice(section.length + 1) : undefined, [name, section])
 
       const value = useRecoilValue(queryFamily(section || name))
+
       return section
         ? value?.[subname]
         : value
@@ -676,9 +710,11 @@ export const SearchContext = React.memo(({
       const section = sectionContext?.section
       const subname = useMemo(() => section ? name.slice(section.length + 1) : undefined, [name, section])
       const setter = useSetRecoilState(queryFamily(section || name))
+      const dynamicQueryModeSetter = useSetRecoilState(dynamicQueryModesFamily(section || name))
 
-      const handleSet = useCallback((value) => {
+      const handleSet = useCallback((value, config = undefined) => {
         updatedFilters.current.add(name)
+        dynamicQueryModeSetter(config?.queryMode)
         section
           ? setter(old => {
             const newValue = isNil(old) ? {} : {...old}
@@ -691,7 +727,7 @@ export const SearchContext = React.memo(({
           : setter(isFunction(value)
             ? (old) => clearEmpty(value(old))
             : clearEmpty(value))
-      }, [section, subname, setter, name])
+      }, [name, dynamicQueryModeSetter, section, setter, subname])
       return handleSet
     }
 
@@ -702,9 +738,9 @@ export const SearchContext = React.memo(({
      * @param {string} name Name of the filter.
      * @returns Array containing the filter value and setter function for it.
      */
-    const useFilterState = (name, section) => {
[…]
