### 2023-07-25 c0b0747
Updated the documentation on mounting plugins within docker.
Changelog: Changed
