### 2023-07-27 3b4fbf3
Resolve "ELN for integration (elabFTW, labfolder, chemotion)"
