### 2023-07-28 8d3a249
Changed the coloring and texts in the upload page to make typical upload workflows more easy to understand.
Changelog: Changed
