### 2023-07-28 23b7026
Spectra normalizer
