### 2023-07-31 ee5228b
Fixed issues with zero values and redundant API calls in the user defined quantities menu.
Changelog: Fixed
