+++ nomad/atomutils.py
@@ -988,7 +988,8 @@ class Formula():
 
 def create_empty_universe(n_atoms: int, n_frames: int = 1, n_residues: int = 1, n_segments: int = 1,
                           atom_resindex: NDArray[Int] = None, residue_segindex: NDArray[Int] = None,
-                          flag_trajectory: bool = False, flag_velocities: bool = False, flag_forces: bool = False):
+                          flag_trajectory: bool = False, flag_velocities: bool = False, flag_forces: bool = False,
+                          timestep: float = None):
     '''Create a blank Universe
 
     This function was adapted from the function empty() within the MDA class Universe().
@@ -1075,7 +1076,7 @@ def create_empty_universe(n_atoms: int, n_frames: int = 1, n_residues: int = 1, 
         # grab and attach a MemoryReader
         universe.trajectory = get_reader_for(coords)(
             coords, order='fac', n_atoms=n_atoms,
-            velocities=vels, forces=forces)
+            velocities=vels, forces=forces, dt=timestep)
 
     return universe
 
@@ -1140,6 +1141,7 @@ def archive_to_universe(archive, system_index: int = 0, method_index: int = -1, 
         sec_system_top = sec_run.system[system_index]
         sec_atoms = sec_system_top.atoms
         sec_atoms_group = sec_system_top.atoms_group
+        sec_calculation = sec_run.calculation
         sec_method = sec_run.method[method_index] if sec_run.get('method') is not None else None
         sec_force_field = sec_method.force_field if sec_method is not None else None
         sec_model = sec_force_field.model[model_index] if sec_force_field is not None else None
@@ -1260,11 +1262,34 @@ def archive_to_universe(archive, system_index: int = 0, method_index: int = -1, 
         if contribution.type == 'bond':  # and contribution.atom_indices is not None:
             bonds.append(tuple(contribution.atom_indices))
 
+    # get the system times
+    system_timestep = None
+
+    def approx(a, b, rel_tol=1e-09, abs_tol=0.0):
+        return abs(a-b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
+
+    system_times = [calc.time for calc in sec_calculation if calc.system_ref]
+    if system_times is []:
+        try:
+            integration_timestep = archive.workflow2.method.integration_timestep
+            coordinate_save_frequency = archive.workflow2.method.coordinate_save_frequency
+            system_timestep = integration_timestep * coordinate_save_frequency
+        except Exception:
+            logging.warning('Cannot find the system times. MDA universe will contain non-physical times and timestep.')
+    else:
+        time_steps = []
+        for i_time in range(1, len(system_times)):
+            time_steps.append(system_times[i_time] - system_times[i_time - 1])
+        if all([approx(time_steps[0], time_step) for time_step in time_steps]):
+            system_timestep = ureg.convert(time_steps[0].magnitude, ureg.second, ureg.picosecond)
+        else:
+            logging.warning('System times are not equally spaced. Cannot set system times in MDA universe. MDA universe will contain non-physical times and timestep.')
+
     # create the Universe
     metainfo_universe = create_empty_universe(
         n_atoms, n_frames=n_frames, n_residues=n_residues, n_segments=n_segments,
         atom_resindex=atom_resindex, residue_segindex=residue_segindex, flag_trajectory=True,
-        flag_velocities=True)
+        flag_velocities=True, timestep=system_timestep)
 
     # set the positions and velocities
     for frame_ind, frame in enumerate(metainfo_universe.trajectory):
[…]
