### 2023-08-02 b8f68de
updated atomistic ref for testing
