### 2023-08-02 c4d1b65
fixed some issues with workflow2 calc of rdf and msd
