### 2023-08-03 91f57ea
Fixed issue with inconsistent upload status icon usage, added more consistent handling of upload deletions.
Changelog: Fixed
