+++ gui/src/components/uploads/DeleteUploadsButton.js
@@ -15,51 +15,86 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React, {useState, useCallback, useEffect} from 'react'
+import React, {useState, useCallback, useEffect, useMemo} from 'react'
 import PropTypes from 'prop-types'
-import { Tooltip, IconButton, Dialog, DialogContent, DialogContentText, DialogActions, Button } from '@material-ui/core'
+import {
+  Tooltip,
+  IconButton,
+  Dialog,
+  DialogContent,
+  DialogContentText,
+  DialogActions,
+  Button
+} from '@material-ui/core'
 import DeleteIcon from '@material-ui/icons/Delete'
 import {pluralize} from '../../utils'
 import {useApi} from '../api'
 import {useErrors} from '../errors'
 import DeletingReferencesTable from './DeletingReferencesTable'
 
-const DeleteUploadsButton = React.memo(({tooltip, disabled, buttonProps, dark, selectedUploads, selectedCount, onSubmitted}) => {
-  const [openDeleteConfirmDialog, setOpenDeleteConfirmDialog] = useState(false)
+/**
+ * Button for deleting uploads. Shows a confirmation dialog with additional
+ * information about the effects of the deletion.
+ */
+const DeleteUploadsButton = React.memo(({
+  tooltip,
+  disabled,
+  buttonProps,
+  dark,
+  uploads,
+  onConfirm,
+  'data-testid': testID
+}) => {
+  const [openDialog, setOpenDialog] = useState(false)
   const {api} = useApi()
   const {raiseError} = useErrors()
   const [entryReferences, setEntryReferences] = useState([])
   const [brokenEntries, setBrokenEntries] = useState([])
-
-  useEffect(() => {
-    if (!openDeleteConfirmDialog) {
-      return
+  const {uploadIDs, nUploads, shared, uploadText, haveText} = useMemo(() => {
+    const nUploads = uploads?.length || 0
+    const uploadText = pluralize('upload', nUploads, false)
+    const haveText = pluralize('has', nUploads, false)
+    return {
+      uploadIDs: uploads?.map(upload => upload.upload_id) || [],
+      nUploads,
+      shared: uploads?.some(entry => entry?.coauthors?.length > 0 || entry?.reviewers?.length > 0),
+      uploadText,
+      haveText
     }
-    const requestBody = {owner: 'visible', query: {entry_references: {'target_upload_id:any': selectedUploads || []}}}
-    api.post(`entries/query`, requestBody)
+  }, [uploads])
+
+  // Fetch information about whether deleting the uploads will result in broken
+  // references.
+  useEffect(() => {
+    if (!openDialog) return
+
+    api.post(`entries/query`,
+      {
+        owner: 'visible',
+        query: {entry_references: {'target_upload_id:any': uploadIDs}}
+      }
+    )
       .then(results => {
         const data = results?.data || []
         if (data.length > 0) {
-          const brokenEntries = data.filter(entry => !selectedUploads.includes(entry.upload_id))
+          const brokenEntries = data.filter(entry => !uploadIDs.includes(entry.upload_id))
           setBrokenEntries(brokenEntries)
           const allReferences = data.map(entry => entry.entry_references).flat()
           const references = [...new Map(allReferences.map(entry => [entry.target_entry_id, entry])).values()]
           setEntryReferences(references)
         }
       })
-      .catch(err =>
-        raiseError(err)
-      )
-  }, [api, openDeleteConfirmDialog, raiseError, selectedUploads])
+      .catch(raiseError)
[…]
