### 2023-08-04 cb90a23
Remove docker image which doesn't seem to be available anymore
