+++ docs/develop/guides.md
@@ -1,6 +1,4 @@
----
-title: Code guidelines
----
+# Code guidelines
 
 NOMAD has a long history and many people are involved in its development. These
 guidelines are set out to keep the code quality high and consistent. Please read
@@ -9,70 +7,81 @@ them carefully.
 ## Principles and rules
 
 - simple first, complicated only when necessary
-- search and adopt generic established 3rd party solutions before implementing specific solutions
-- only uni directional dependencies between components/modules, no circles
-- only one language: Python (except, GUI of course)
+
+- search and adopt generic established 3rd-party solutions before implementing specific
+  solutions
+
+- only unidirectional dependencies between components/modules, no circles
+
+- only one language: Python (except GUI of course)
 
 The are some *rules* or better strong *guidelines* for writing code. The following
-applies to all python code (and were applicable, also to JS and other code):
+applies to all Python code (and where applicable, also to Javascript and other code):
 
-- Use an IDE (e.g. [vscode](https://code.visualstudio.com/) or otherwise automatically
-  enforce code ([formatting and linting](https://code.visualstudio.com/docs/python/linting)).
-  Use `nomad qa` before committing. This will run all tests, static type checks, linting, etc.
+- Use an IDE (e.g. [VS Code](https://code.visualstudio.com/)) or otherwise automatically
+  enforce
+  [code formatting and linting](https://code.visualstudio.com/docs/python/linting).
 
-- There is a style guide to python. Write [pep-8](https://www.python.org/dev/peps/pep-0008/)
-  compliant python code. An exception is the line cap at 79, which can be broken but keep it 90-ish.
+- Use `nomad qa` before committing. This will run all tests, static type checks, linting,
+  etc.
 
-- Test the public interface of each sub-module (i.e. python file)
+- Test the public interface of each submodule (i.e. Python file).
 
-- Be [pythonic](https://docs.python-guide.org/writing/style/) and watch
-  [this](https://www.youtube.com/watch?v=wf-BqAjZb8M).
+- There is a style guide to Python. Write
+  [PEP 8](https://www.python.org/dev/peps/pep-0008/)-compliant Python code. An exception
+  is the line cap at 79, which can be broken but keep it 90-ish.
 
-- Add doc-strings to the *public* interface of each sub-module (e.g. python file). Public meaning API that
-  is exposed to other sub-modules (i.e. other python files).
+- Be [Pythonic](https://docs.python-guide.org/writing/style/) and watch
+  [this talk about best practices](https://www.youtube.com/watch?v=wf-BqAjZb8M).
 
-- The project structure is according to [this guide](https://docs.python-guide.org/writing/structure/).
-  Keep it!
+- Add docstrings to the *public* interface of each submodule (i.e. Python file). This
+  includes APIs that are exposed to other submodules (i.e. other Python files).
+
+- The project structure follows
+  [this guide](https://docs.python-guide.org/writing/structure/). Keep it!
 
 - Write tests for all contributions.
 
-- Adopt *Clean Code* practices. Here is a [good introduction](https://youtu.be/7EmboKQH8lM).
+- Adopt *Clean Code* practices. Here is a good
+  [introductory talk to Clean Code](https://youtu.be/7EmboKQH8lM).
 
-
-## Enforcing Rules with CI/CD
+## Enforcing rules with CI/CD
 
 These *guidelines* are partially enforced by CI/CD. As part of CI all tests are run on all
-branches; further we run a *linter*, *pep8* checker, and *mypy* (static type checker). You can
-run `nomad qa` to run all these tests and checks before committing.
+branches; further we run a *linter*, *PEP 8* checker, and *mypy* (static type checker).
+You can run `nomad qa` to run all these tests and checks before committing.
 
[…]
