+++ docs/develop/parsers.md
@@ -1,24 +1,30 @@
-NOMAD uses parsers to convert raw code input and output files into NOMAD's common Archive format. This is the documentation on how to develop such a parser.
+# How to write a parser
+
+NOMAD uses parsers to convert raw code input and output files into NOMAD's common archive
+format. This is the documentation on how to develop such a parser.
 
 ## Getting started
 
 Let's assume we need to write a new parser from scratch.
 
-First we need to install *nomad-lab* Python package to get the necessary libraries:
-```sh
+First we need to install the `nomad-lab` Python package to get the necessary libraries:
+
+```shell
 pip install nomad-lab
 ```
 
-We prepared an example parser project that you can work with.
-```sh
+We have prepared an example parser project that you can work with:
+
+```shell
 git clone https://github.com/nomad-coe/nomad-parser-example.git --branch hello-world
 ```
 
 Alternatively, you can fork the example project on GitHub to create your own parser. Clone
 your fork accordingly.
 
-The project structure should be
-```none
+The project structure should be:
+
+```text
 example/exampleparser/__init__.py
 example/exampleparser/__main__.py
 example/exampleparser/metainfo.py
@@ -28,14 +34,16 @@ example/README.md
 example/setup.py
 ```
 
-Next you should install your new parser with pip. The `-e` parameter installs the parser
-in *development*. This means you can change the sources without having to reinstall.
-```sh
+Next, you should install your new parser with pip. The `-e` parameter installs the parser
+in *development*. This means you can change the sources without having to reinstall:
+
+```shell
 cd example
 pip install -e .
 ```
 
 The main code file `exampleparser/parser.py` should look like this:
+
 ```python
 class ExampleParser(MatchingParser):
     def __init__(self):
@@ -52,18 +60,20 @@ class ExampleParser(MatchingParser):
 A parser is a simple program with a single class. The base class `MatchingParser`
 provides the necessary interface to NOMAD. We provide some basic information
 about our parser in the constructor. The *main* function `run` simply takes a filepath
-and an empty archive as input. Now its up to you, to open the given file and populate the
+and an empty archive as input. Now it's up to you, to open the given file and populate the
 given archive accordingly. In the plain *hello world*, we simple create a log entry,
 populate the archive with a *root section* `Run`, and set the program name to `EXAMPLE`.
 
 You can run the parser with the included `__main__.py`. It takes a file as argument and
 you can run it like this:
-```sh
+
+```shell
 python -m exampleparser tests/data/example.out
 ```
 
 The output should show the log entry and the minimal archive with one `section_run` and
-the respective `program_name`.
+the respective `program_name`:
+
 ```json
 {
   "section_run": [
@@ -77,15 +87,18 @@ the respective `program_name`.
 ## Parsing test files
 
 Let's do some actual parsing. Here we demonstrate how to parse ASCII files with some
-structure information in it. As it is typically used by materials science codes.
+structure information in it, as it is typically used by materials science codes.
 
 On the `master` branch of the example project, we have a more 'realistic' example:
-```sh
+
+```shell
 git checkout master
 ```
 
-This example imagines a potential code output that looks like this (`tests/data/example.out`):
-```none
+This example imagines a potential code output that looks like this
+(`tests/data/example.out`):
+
+```text
 2020/05/15
                *** super_code v2 ***
 
@@ -106,14 +119,15 @@ cell: (0, 0, 0), (1, 0, 0), (1, 1, 0)
 energy: 1.29372
 ```
 
-At the top there are some general informations. Below that is a list of simulated systems with
-sites and lattice describing crystal structures as well as a computed energy value as an example for
-a code specific quantity from a 'magic source'.
[…]

[…]
