### 2023-08-05 nomad-coe/nomad#81 @berquist
Running `./scripts/setup_dev_env.sh` fails at MDAnalysis build
This is part of [the JOSS review](https://github.com/openjournals/joss-reviews/issues/5388).

Following https://nomad-lab.eu/prod/v1/docs/develop/setup.html#install-nomad, running the script after performing
```bash
pyenv install 3.9.17
pyenv local 3.9.17
python -m venv ./venv
source ./venv/bin/activate
```
gives the following `pip` traceback, failing to compile MDAnalysis 2.1.0:
```
Collecting mdanalysis==2.1.0 (from -r requirements-dev.txt (line 144))
  Downloading MDAnalysis-2.1.0.tar.gz (3.5 MB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [58 lines of output]

      Error compiling Cython file:
      ------------------------------------------------------------
      ...
          array_wrapper = ArrayWrapper()
          array_wrapper.set_data(<void*> data_ptr, <int*> &dim[0], dim.size, data_type)

          cdef np.ndarray ndarray = np.array(array_wrapper, copy=False)
          # Assign our object to the 'base' of the ndarray object
          ndarray.base = <PyObject*> array_wrapper
                 ^
      ------------------------------------------------------------

      MDAnalysis/lib/formats/cython_util.pyx:115:11: Assignment to a read-only property
      Attempting to autodetect OpenMP support... Compiler supports OpenMP
      Will attempt to use Cython.
      Compiling MDAnalysis/lib/formats/libdcd.pyx because it changed.
      Compiling MDAnalysis/lib/c_distances.pyx because it changed.
      Compiling MDAnalysis/lib/c_distances_openmp.pyx because it changed.
      Compiling MDAnalysis/lib/qcprot.pyx because it changed.
      Compiling MDAnalysis/lib/formats/libmdaxdr.pyx because it changed.
      Compiling MDAnalysis/lib/formats/cython_util.pyx because it changed.
      Compiling MDAnalysis/analysis/encore/cutils.pyx because it changed.
      Compiling MDAnalysis/analysis/encore/clustering/affinityprop.pyx because it changed.
      Compiling MDAnalysis/analysis/encore/dimensionality_reduction/stochasticproxembed.pyx because it changed.
      Compiling MDAnalysis/lib/_cutil.pyx because it changed.
      Compiling MDAnalysis/lib/_augment.pyx because it changed.
      Compiling MDAnalysis/lib/nsgrid.pyx because it changed.
      [ 1/12] Cythonizing MDAnalysis/analysis/encore/clustering/affinityprop.pyx
      [ 2/12] Cythonizing MDAnalysis/analysis/encore/cutils.pyx
      [ 3/12] Cythonizing MDAnalysis/analysis/encore/dimensionality_reduction/stochasticproxembed.pyx
      [ 4/12] Cythonizing MDAnalysis/lib/_augment.pyx
      [ 5/12] Cythonizing MDAnalysis/lib/_cutil.pyx
      [ 6/12] Cythonizing MDAnalysis/lib/c_distances.pyx
      [ 7/12] Cythonizing MDAnalysis/lib/c_distances_openmp.pyx
      [ 8/12] Cythonizing MDAnalysis/lib/formats/cython_util.pyx
      Traceback (most recent call last):
        File "/home/eric/repositories/nomad-coe/nomad/venv/lib/python3.9/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
        File "/home/eric/repositories/nomad-coe/nomad/venv/lib/python3.9/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/home/eric/repositories/nomad-coe/nomad/venv/lib/python3.9/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 118, in get_requires_for_build_wheel
          return hook(config_settings)
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/setuptools/build_meta.py", line 341, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=['wheel'])
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/setuptools/build_meta.py", line 323, in _get_build_requires
          self.run_setup()
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/setuptools/build_meta.py", line 487, in run_setup
          super(_BuildMetaLegacyBackend,
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/setuptools/build_meta.py", line 338, in run_setup
          exec(code, locals())
        File "<string>", line 590, in <module>
        File "<string>", line 449, in extensions
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/Cython/Build/Dependencies.py", line 1134, in cythonize
          cythonize_one(*args)
        File "/tmp/pip-build-env-f6kxxutk/overlay/lib/python3.9/site-packages/Cython/Build/Dependencies.py", line 1301, in cythonize_one
          raise CompileError(None, pyx_file)
      Cython.Compiler.Errors.CompileError: MDAnalysis/lib/formats/cython_util.pyx
      [end of output]
```
I haven't confirmed, but I wouldn't be surprised if it's one of the many issues related to the new Cython 3.0.0 release, since it's probably picking up my system install, and https://github.com/MDAnalysis/mdanalysis/blob/2bf55dd714292bbf7c12023a28153a6443a52ec6/package/CHANGELOG#L86 shows this may not have been fixed until MDAnalysis 2.5.0.

Two questions:
- Can the pinned version of MDAnalysis be updated?
- If not, can a workaround be added?
