### 2023-08-05 nomad-coe/nomad#81 @berquist
Regarding the pin, I've read https://github.com/nomad-coe/nomad/issues/75#issuecomment-1633223069 and do agree, but you should also know that since many projects bring in Cython as a transitive dependency those projects don't necessarily pin Cython versions, their upgrade from from 0.29 to 3.0 broke a _lot_ of code.
