### 2023-08-07 nomad-coe/nomad#81 @lauri-codes
Thanks @berquist for this report and all of the details.

@JFRudzinski: I remember that you were trying out updating the MDAnalysis version: Did you update the mda version in the issue you are working on: [#1643](https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR/-/issues/1643)? In principle, I don't see any problems updating our pinned version to e.g. 2.5.0.
