### 2023-08-07 nomad-coe/nomad#81 @JFRudzinski
@lauri-codes Actually, I had updated to 2.5.0 some time ago (not sure which issue or branch). However, in my tests for [#1643](https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR/-/issues/1643) I found that 2.5.0 has some issues reading certain versions of .trr and .xtc files.

So, I was planning to revert now back to 2.1.0. However, once I am done with my dev, I can do some more testing if it would help to see what the root of the issues are with 2.5.0 in my case?
