### 2023-08-07 nomad-coe/nomad#81 @lauri-codes
@JFRudzinski: Yes, it would be great if you could debug the issue in reading .trr and .xtc files. If the problem is with MDA itself, then you should search for any existing issues in their [GitHub page](https://github.com/MDAnalysis/mdanalysis/issues), and possibly file a new one if nothing exists already.

It is a bit unfortunate that MDA has a cython build step: [the official recommendation](https://cython.readthedocs.io/en/latest/src/userguide/source_files_and_compilation.html#distributing-cython-modules) is to ship the resulting .c files in a release and use them primarily even if the user has cython installed. But looking at their setup.py, they seem to prefer the cython compilation if cython exists in the installation environment.
