### 2023-08-07 11a84ca
adjusted calculation metainfo for MD sims
