### 2023-08-07 e3248fd
Improved the handling of dynamically resolved filters in SearchContext.
Changelog: Changed
