+++ gui/src/components/search/SearchContext.js
@@ -29,32 +29,32 @@ import {
 } from 'recoil'
 import {
   debounce,
-  isEmpty,
-  isArray,
   isBoolean,
+  isEmpty,
+  isEqual,
+  isArray,
   isPlainObject,
   isNil,
   isSet,
   isFunction,
   size,
-  isEqual,
   cloneDeep
 } from 'lodash'
-import { Link } from '@material-ui/core'
 import qs from 'qs'
+import { Link } from '@material-ui/core'
 import { v4 as uuidv4 } from 'uuid'
 import PropTypes from 'prop-types'
 import { useHistory } from 'react-router-dom'
 import { useApi } from '../api'
 import {
-  setToArray,
   authorList,
   entryName,
   entryType,
   formatTimestamp,
   getDeep,
   formatNumber,
-  getDatatype
+  getDatatype,
+  setToArray
 } from '../../utils'
 import { Quantity, Unit } from '../../units'
 import { useErrors } from '../errors'
@@ -63,12 +63,7 @@ import UploadStatusIcon from '../uploads/UploadStatusIcon'
 import { getWidgetsObject } from './widgets/Widget'
 import { inputSectionContext } from './input/InputSection'
 import { searchQuantities } from '../../config'
-import {
-  filterData as filterDataGlobal,
-  filterAbbreviations,
-  filterFullnames,
-  materialNames
-} from './FilterRegistry'
+import { filterData as filterDataStatic } from './FilterRegistry'
 
 const debounceTime = 450
 
@@ -91,23 +86,6 @@ const debounceTime = 450
  * depending on the context would re-render for each filter change regardless if
  * it actually changes the state of that component or not.
  */
-
-/**
- * Used to turn empty containers into undefined which is used to indicate that a
- * search filter has not been specified.
- */
-function clearEmpty(value) {
-  function isEmpty(value) {
-    if (isPlainObject(value)) {
-      return Object.values(value).every(isEmpty)
-    } else if ((isSet(value) || isArray(value)) && size(value) === 0) {
-      return true
-    }
-    return false
-  }
-  return isEmpty(value) ? undefined : value
-}
-
 export const searchContext = React.createContext()
 export const SearchContext = React.memo(({
   resource,
@@ -222,7 +200,7 @@ export const SearchContext = React.memo(({
         }
       },
       published: {
-        render: (entry) => <UploadStatusIcon data={entry} user={user}/>
+        render: (entry) => <UploadStatusIcon data={entry} user={user} />
       }
     }
 
@@ -260,25 +238,26 @@ export const SearchContext = React.memo(({
       : undefined
   }, [initialDashboard])
 
-  // Initialize the set of available filters. This may depend on the resource.
-  const [filtersLocal, filterDataLocal] = useMemo(() => {
-    const include = initialFilters?.include || [...Object.keys(filterDataGlobal)]
+  // Initialize the set of filters that are available directly from the
+  // metainfo.
+  const {filtersStatic, filtersStaticData, filtersStaticAbbreviations} = useMemo(() => {
+    const staticFilterNames = Object.keys(filterDataStatic)
+    const include = initialFilters?.include || [...staticFilterNames]
     const exclude = initialFilters?.exclude || []
     const filters = include.filter((key) => !exclude?.includes(key))
-    const filtersLocal = new Set(filters)
-    const filterDataLocal = Object.fromEntries(filters.map(
-      (name) => [name, filterDataGlobal[name]]
+    const filtersStatic = new Set(filters)
+    const filtersStaticData = Object.fromEntries(filters.map(
+      (name) => [name, filterDataStatic[name]]
     ))
-    return [filtersLocal, filterDataLocal]
+    const filtersStaticAbbreviations = getAbbreviations([...Object.keys(filtersStaticData)])
[…]
