### 2023-08-08 b6a54c6
added gromacs specific energy quants metainfo, and changed mda version back to 2.5.0
