+++ nomad/atomutils.py
@@ -51,6 +51,7 @@ import MDAnalysis.analysis.rdf as MDA_RDF
 from nomad.aflow_prototypes import aflow_prototypes
 from nomad.constants import atomic_masses
 from nomad.units import ureg
+# from nomad.parsing.file_parser import Quantity
 
 
 valid_elements = set(ase.data.chemical_symbols[1:])
@@ -1265,23 +1266,21 @@ def archive_to_universe(archive, system_index: int = 0, method_index: int = -1, 
             bonds.append(tuple(contribution.atom_indices))
 
     # get the system times
-    system_timestep = None
+    system_timestep = 1.0
 
     def approx(a, b, rel_tol=1e-09, abs_tol=0.0):
-        return abs(a-b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
+        return abs(a - b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
 
     system_times = [calc.time for calc in sec_calculation if calc.system_ref]
-    if system_times is []:
+    if system_times:
         try:
-            integration_timestep = archive.workflow2.method.integration_timestep
+            method = archive.workflow2.method
             coordinate_save_frequency = archive.workflow2.method.coordinate_save_frequency
-            system_timestep = integration_timestep * coordinate_save_frequency
+            system_timestep = method.integration_timestep * coordinate_save_frequency
         except Exception:
             logging.warning('Cannot find the system times. MDA universe will contain non-physical times and timestep.')
     else:
-        time_steps = []
-        for i_time in range(1, len(system_times)):
-            time_steps.append(system_times[i_time] - system_times[i_time - 1])
+        time_steps = [system_times[i_time] - system_times[i_time - 1] for i_time in range(1, len(system_times))]
         if all([approx(time_steps[0], time_step) for time_step in time_steps]):
             system_timestep = ureg.convert(time_steps[0].magnitude, ureg.second, ureg.picosecond)
         else:
@@ -1451,7 +1450,7 @@ def shifted_correlation_average(function: Callable, times: NDArray, positions: N
     return correlation_times, result
 
 
-def __calc_diffusion_constant(times: NDArray, values: NDArray, dim: int = 3):
+def _calc_diffusion_constant(times: NDArray, values: NDArray, dim: int = 3):
     '''
     Determines the diffusion constant from a fit of the mean squared displacement
     vs. time according to the Einstein relation.
@@ -1462,7 +1461,7 @@ def __calc_diffusion_constant(times: NDArray, values: NDArray, dim: int = 3):
     return slope * 1 / (2 * dim), error
 
 
-def __get_molecular_bead_groups(universe: MDAnalysis.Universe, moltypes: List[str] = None):
+def _get_molecular_bead_groups(universe: MDAnalysis.Universe, moltypes: List[str] = None):
 
     if moltypes is None:
         atoms_moltypes = getattr(universe.atoms, 'moltypes', [])
@@ -1483,14 +1482,10 @@ def calc_molecular_rdf(universe: MDAnalysis.Universe, n_traj_split: int = 10, n_
 
     interval_indices: 2D array specifying the groups of the n_traj_split intervals to be averaged
     max_mols: the maximum number of molecules per bead group for calculating the rdf, for efficiency purposes.
-    5k was set after > 50k was giving problems. Should do further testing to see where the appropriate limit should be set.
     '''
 
-    if universe is None:
-        return
-    trajectory = universe.trajectory[0] if universe.trajectory else None
-    dimensions = getattr(trajectory, 'dimensions', None) if trajectory else None
-    if dimensions is None:
+    # TODO 5k default for max_mols was set after > 50k was giving problems. Should do further testing to see where the appropriate limit should be set.
[…]
