### 2023-08-10 nomad-coe/nomad#81 @JFRudzinski
@lauri-codes I am fairly confident now that the only issue with MDA version 2.5.0 is something to do with a sort of advanced usage that I am performing when calculating the msds. Additionally, this is only happening for 1 trajectory file obtained from a user. (As I said before, I had been using 2.5.0 for a while without any issues on many other data).

I would say then that you can pretty safely go ahead and update the MDA version. I will address my own specific issue in the mean time, but I don't really think you need to wait for this, as it is so specific (and possibly unique to this single example).
