### 2023-08-11 5a4184d
Improved styling and navigation when browsing nexus files
Changelog: Changed
