+++ gui/tests/artifacts.js
@@ -56088,6 +56088,7 @@ window.nomadArtifacts = {
                     "apm",
                     "ellips",
                     "em_nion",
+                    "em_om",
                     "em_spctrscpy",
                     "example",
                     "hall",
@@ -56151,9 +56152,13 @@ window.nomadArtifacts = {
                     "NXxnb",
                     "NXxrot",
                     "NXaberration",
-                    "NXaperture",
+                    "NXaberration_model",
+                    "NXaberration_model_ceos",
+                    "NXaberration_model_nion",
+                    "NXadc",
                     "NXaperture_em",
                     "NXapm",
+                    "NXapm_composition_space_results",
                     "NXapm_input_ranging",
                     "NXapm_input_reconstruction",
                     "NXapm_paraprobe_config_clusterer",
@@ -56176,7 +56181,8 @@ window.nomadArtifacts = {
                     "NXapm_paraprobe_results_surfacer",
                     "NXapm_paraprobe_results_tessellator",
                     "NXapm_paraprobe_results_transcoder",
-                    "NXbeam",
+                    "NXbeam_path",
+                    "NXbeam_splitter",
                     "NXcalibration",
                     "NXcg_alpha_complex",
                     "NXcg_cylinder_set",
@@ -56200,6 +56206,7 @@ window.nomadArtifacts = {
                     "NXcg_unit_normal_set",
                     "NXchamber",
                     "NXchemical_composition",
+                    "NXcircuit_board",
                     "NXclustering",
                     "NXcollectioncolumn",
                     "NXcontainer",
@@ -56217,9 +56224,9 @@ window.nomadArtifacts = {
                     "NXcs_profiling_event",
                     "NXcsg",
                     "NXcxi_ptycho",
+                    "NXdac",
                     "NXdeflector",
                     "NXdelocalization",
-                    "NXdetector",
                     "NXdispersion",
                     "NXdispersion_function",
                     "NXdispersion_repeated_parameter",
@@ -56236,26 +56243,17 @@ window.nomadArtifacts = {
                     "NXem_ebsd_conventions",
                     "NXem_ebsd_crystal_structure_model",
                     "NXenergydispersion",
-                    "NXentry",
                     "NXevent_data_em",
                     "NXevent_data_em_set",
                     "NXfabrication",
+                    "NXfiber",
                     "NXgraph_edge_set",
                     "NXgraph_node_set",
                     "NXgraph_root",
                     "NXibeam_column",
-                    "NXimage_set_em",
+                    "NXimage_set",
                     "NXimage_set_em_adf",
-                    "NXimage_set_em_bf",
-                    "NXimage_set_em_bse",
-                    "NXimage_set_em_chamber",
-                    "NXimage_set_em_df",
-                    "NXimage_set_em_diffrac",
-                    "NXimage_set_em_ecci",
                     "NXimage_set_em_kikuchi",
-                    "NXimage_set_em_ronchigram",
-                    "NXimage_set_em_se",
-                    "NXinstrument",
                     "NXinteraction_vol_em",
                     "NXion",
                     "NXisocontour",
@@ -56263,24 +56261,23 @@ window.nomadArtifacts = {
                     "NXlab_electro_chemo_mechanical_preparation",
                     "NXlab_sample_mounting",
                     "NXlens_em",
+                    "NXlens_opt",
                     "NXmagnetic_kicker",
                     "NXmanipulator",
                     "NXmatch_filter",
                     "NXmpes",
                     "NXms",
-                    "NXms_atom_set",
-                    "NXms_crystal_set",
-                    "NXms_dislocation_set",
-                    "NXms_interface_set",
+                    "NXms_feature_set",
                     "NXms_score_config",
[…]
