### 2023-08-14 bb727db
Read all once
