### 2023-08-15 nomad-coe/nomad#81 @lauri-codes
The MDA version has now been updated to v2.5.0.
