### 2023-08-15 #5388 @berquist
Again, apologies for being slow.  I don't have a good excuse, but it was not the fault of the JOSS process, which has been painless.  In this spirit of not letting the perfect be the enemy of the good...

# Review

## Paper

The paper itself is well-written, and while there might be some grammar issues or awkward wording in a handful of places, the intent is clear.  The referenced literature touches upon the main work in the same space that I'm familiar with and touches upon the conceptual difference, but makes it sound like only the only users of those group's software infrastructure (not to be confused with the databases containing calculation results).  This is odd because they explicitly mention pymatgen as a dependency, and pymatgen is originally from the Materials Project.  This might be one of the aforementioned wording issues.  There are no specific corrections I would make unless further clarification is necessary.

While there are references to other databases, except for the last paragraph, there are no technical comparisons to the frameworks behind these databases.  That may not be where NOMAD shines, which to me looks like the FAIR support, the web frontend, and the ease of hosting.

My principal concern is over the author list.  https://github.com/openjournals/joss-reviews/issues/5388#issuecomment-1548937979 did mention that JOSS allows authors other than code contributors in the list (which I don't recall seeing codified anywhere), and that does alleviate some concerns.  Going by the GitHub contributor ranking, while an imperfect metric, shows that there are handful of people with code contributions, some appearing substantial, that are not in the list at all.  I have not looked at the GitLab instance, and my notes may be stale now (from 2023-04-22), but it isn't clear why these people are not authors when they're high on the GH contributor list:

- Luigi Sbailò
- Cuauhtemoc Salazar

and these authors in the repository `AUTHORS`:

- Daniel Speckhard

and it is not clear why these people are not equal authors when they are high on the GH contributor list:

- José A. Márquez
- possibly others above 15.

## Functionality

The first place I usually look for how to build any piece of software is the CI definition, since if tests are passing, it's more likely to work than sometimes outdated documentation.  I extracted this from `.gitlab-ci.yaml` and had no problems building the image(s):

```bash
#!/bin/bash

set -euo pipefail

CI_REGISTRY_IMAGE=nomad
DOCKER_TAG=1.1.8

docker build -t ${CI_REGISTRY_IMAGE}:${DOCKER_TAG} .
docker build --target dev_node -t ${CI_REGISTRY_IMAGE}/dev_node:${DOCKER_TAG} .
docker build --target dev_python -t ${CI_REGISTRY_IMAGE}/dev_python:${DOCKER_TAG} .
```

Having a Docker image really serves as a nice standard now that Python packaging is so difficult to deal with.  The Dockerfile itself follows good practices by not only having useful comments and tidying up after certain steps, but performing a multi-stage build with multiple available tags, and the final images having contents not owned as root.  With not much work it should even be a foundation for a development environment using [devcontainers](https://containers.dev/).

It's not a requirement, but unfortunately replacing Docker with Podman doesn't quite work:

```
[6/6] STEP 4/15: COPY --chown=nomad:1000 scripts/run.sh .
Error: building at STEP "COPY --chown=nomad:1000 scripts/run.sh .": looking up UID/GID for "nomad:1000": determining run uid: user: unknown user error looking up user "nomad"
```

Being able to build and run with Podman and/or Singularity/Apptainer could open up NOMAD containers to more permissions-restrained environments.  Apptainer can at least ingest the already-built image (`apptainer build nomad-fair-latest.sif docker://gitlab-registry.mpcdf.mpg.de/nomad-lab/nomad-fair:latest`).

Going back to look at the [documentation](https://nomad-lab.eu/prod/v1/docs/develop/setup.html#build-the-docker-image), it covers exactly this, which is a good sign.

Under the [OASIS](http://nomad-lab.eu/prod/v1/docs/oasis.html), in addition to "Download our basic configuration files", there could be a link to the identical docker-compose infrastructure in the Git repository.  The ideal would be if MkDocs had functionality similar to Doxygen where you can put code comments in files that have a named anchor or tag around a span of text which the doc build tool will read and then insert that block directly into the docs, so that the docs never fall out of data with regards to the code.  (I don't know MkDocs well so it may or may not be possible.)

Deploying goes exactly as stated in [Operating an OASIS](https://nomad-lab.eu/prod/v1/docs/oasis.html#operating-an-oasis).  A smaller Docker-based Digital Ocean droplet with the [Docker UID changed](https://github.com/nomad-coe/nomad/blob/5b0fccd7362732a208b7a95d3052830f745f7951/ops/docker-compose/nomad-oasis/docker-compose.yaml#L150) and port 80 opened in the firewall had me uploading data in minutes.

Regarding CI, the number of GitLab job definitions are considerable but have reasonably descriptive names.  It also looks like not much remains that could be abstracted into scripts which do not depend on the CI platform, which is a common trap.  Having scripts shared between CI and local development aids consistency between the two environments.  The scripts themselves mostly handle egregious errors (`set -e`) but they could use some high-level documentation.

The dedicated `./scripts/setup_dev_env.sh` unfortunately [failed for me](https://github.com/nomad-coe/nomad/issues/81), and highlights the growing need for something such as Docker- or Spack-based development environments, particularly for Python-based or mixed-language projects.  Similarly, I appreciate not being tied to conda in any way.

## Documentation

Unfortunately, the GitHub repository is only a mirror for the self-hosted repository.  This is understandable, but is confusing for contributors since GitHub--GitLab integration is not perfect.  At least this section of the [README](https://github.com/nomad-coe/nomad#contributing) is right at the top and succinctly explains exactly how contributing works.

Along with the Docker-relevant documentation mentioned above, the general quality of the developer documentation is high.  Having "How to do X" or "How to write Y" pages is helpful for onboarding when projects either grow large or are highly domain specific.  "How to write a parser" even has a worked example for a hypothetical output.  An additional thing that some might find sloppy or distracting but I find informative, at least as a developer, is that the TODOs for documentation are simply left directly on the rendered pages.  It would be interesting to know (not in the paper) if any issues or requests have been made as a result.

## Conclusions

I completely support NOMAD's publication in JOSS.
