### 2023-08-16 03b23a0
Added System and PureSubstance components for CompositeSystem, added Experiment, Analysis, and Collection base sections, added "How to use basesections" documentation, changed name of Ensemble to CompositeSystem and name of Substance to PureSubstance.
Changelog: Added
