+++ gui/tests/artifacts.js
@@ -56623,6 +56623,1029 @@ window.nomadArtifacts = {
             "m_def": "nomad.metainfo.metainfo.Section",
             "m_parent_index": 3,
             "m_parent_sub_section": "section_definitions",
+            "m_annotations": {
+              "eln": [
+                {
+                  "lane_width": "600px"
+                }
+              ]
+            },
+            "name": "ELNProcess",
+            "label": "Material Processing ELN",
+            "description": "A basic electronic lab notebook for a material processing activity.",
+            "categories": [
+              "/packages/14/category_definitions/2"
+            ],
+            "base_sections": [
+              "/packages/15/section_definitions/20",
+              "/packages/14/section_definitions/1"
+            ],
+            "quantities": [
+              {
+                "m_def": "nomad.metainfo.metainfo.Quantity",
+                "m_parent_index": 0,
+                "m_parent_sub_section": "quantities",
+                "m_annotations": {
+                  "eln": [
+                    {
+                      "component": "StringEditQuantity"
+                    }
+                  ],
+                  "template": [
+                    {
+                      "process_identifiers": {}
+                    }
+                  ]
+                },
+                "name": "tags",
+                "description": "Add a tag that can be used for search.",
+                "type": {
+                  "type_kind": "python",
+                  "type_data": "str"
+                },
+                "shape": [
+                  "*"
+                ]
+              }
+            ],
+            "sub_sections": [
+              {
+                "m_def": "nomad.metainfo.metainfo.SubSection",
+                "m_parent_index": 0,
+                "m_parent_sub_section": "sub_sections",
+                "name": "process_identifiers",
+                "sub_section": "/packages/15/section_definitions/31"
+              }
+            ]
+          },
+          {
+            "m_def": "nomad.metainfo.metainfo.Section",
+            "m_parent_index": 4,
+            "m_parent_sub_section": "section_definitions",
+            "name": "BasicMeasurementResult",
+            "label": "Measurement Result",
+            "description": "A basic section for describing the result of a measurement.",
+            "base_sections": [
+              "/packages/15/section_definitions/24"
+            ],
+            "quantities": [
+              {
+                "m_def": "nomad.metainfo.metainfo.Quantity",
+                "m_parent_index": 0,
+                "m_parent_sub_section": "quantities",
+                "m_annotations": {
+                  "eln": [
+                    {
+                      "component": "RichTextEditQuantity"
+                    }
+                  ]
+                },
+                "name": "result",
+                "type": {
+                  "type_kind": "python",
+                  "type_data": "str"
+                }
+              }
+            ]
+          },
+          {
+            "m_def": "nomad.metainfo.metainfo.Section",
+            "m_parent_index": 5,
+            "m_parent_sub_section": "section_definitions",
+            "m_annotations": {
+              "eln": [
+                {
+                  "lane_width": "600px"
+                }
+              ],
+              "template": [
+                {
+                  "measurement_identifiers": {}
+                }
+              ]
+            },
+            "name": "ELNMeasurement",
+            "label": "Measurement ELN",
+            "description": "A basic electronic lab notebook for a measurement activity.",
+            "categories": [
+              "/packages/14/category_definitions/2"
+            ],
+            "base_sections": [
+              "/packages/15/section_definitions/25",
+              "/packages/14/section_definitions/1"
+            ],
+            "quantities": [
+              {
+                "m_def": "nomad.metainfo.metainfo.Quantity",
+                "m_parent_index": 0,
+                "m_parent_sub_section": "quantities",
+                "m_annotations": {
+                  "eln": [
+                    {
+                      "component": "StringEditQuantity"
+                    }
+                  ]
+                },
+                "name": "tags",
[…]
