+++ nomad/datamodel/metainfo/basesections.py
@@ -16,7 +16,7 @@
 # limitations under the License.
 #
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Iterable
 import datetime
 import re
 from typing import (
@@ -30,11 +30,14 @@ from ase.data import (
     atomic_numbers,
     atomic_masses,
 )
+import requests
 if TYPE_CHECKING:
     from structlog.stdlib import (
         BoundLogger,
     )
-
+from nomad.atomutils import (
+    Formula,
+)
 from nomad import (
     utils,
 )
@@ -61,6 +64,125 @@ from nomad.datamodel.results import (
     ElementalComposition as ResultsElementalComposition,
     Material,
 )
+from nomad.datamodel.metainfo.annotations import (
+    ELNAnnotation,
+)
+
+
+PUB_CHEM_PUG_PATH = 'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound'
+CAS_API_PATH = 'https://commonchemistry.cas.org/api'
+EXTERNAL_API_TIMEOUT = 5
+
+
+def pub_chem_api_get_properties(cid: int, properties: Iterable[str]) -> requests.Response:
+    '''
+    Function for performing a get request to the PubChem PUG API to get properties for a
+    given compound identifier.
+
+    Args:
+        cid (int): The compound identifier of the compound of interest.
+        properties (Iterable[str]): The properties to retrieve the value for.
+
+    Returns:
+        requests.Response: The response as returned from the PubChem PUG API.
+    '''
+    return requests.get(
+        url=f'{PUB_CHEM_PUG_PATH}/cid/{cid}/property/{str.join(",", properties)}/JSON',
+        timeout=EXTERNAL_API_TIMEOUT,
+    )
+
+
+def pub_chem_api_get_synonyms(cid: int) -> requests.Response:
+    '''
+    Function for performing a get request to the PubChem PUG API to get properties for a
+    given compound identifier.
+
+    Args:
+        cid (int): The compound identifier of the compound of interest.
+
+    Returns:
+        requests.Response: The response as returned from the PubChem PUG API.
+    '''
+    return requests.get(
+        url=f'{PUB_CHEM_PUG_PATH}/cid/{cid}/synonyms/JSON',
+        timeout=EXTERNAL_API_TIMEOUT,
+    )
+
+
+def pub_chem_api_search(path: str, search: str) -> requests.Response:
+    '''
+    Function for performing a get request to the PubChem PUG API to search the given path
+    for a given string.
+
+    Args:
+        path (str): The path (property) to search for.
+        search (str): The string to search for a match with.
+
+    Returns:
+        requests.Response: The response as returned from the PubChem PUG API.
+    '''
+    return requests.get(
+        url=f'{PUB_CHEM_PUG_PATH}/{path}/{search}/cids/JSON',
+        timeout=EXTERNAL_API_TIMEOUT,
+    )
+
+
+def cas_api_search(search: str) -> requests.Response:
+    '''
+    Function for performing a get request to the CAS API to search for a match with the
+    given string.
+
+    Args:
+        search (str): The string to search for a match with.
+
+    Returns:
+        requests.Response: The response as returned from the CAS API.
+    '''
+    return requests.get(
+        f'{CAS_API_PATH}/search?q={search}',
+        timeout=EXTERNAL_API_TIMEOUT,
+    )
+
+
+def cas_api_details(cas_rn: str) -> requests.Response:
+    '''
+    Function for performing a get request to the CAS API to get the details for the
+    substance with the given CAS registry number.
+
+    Args:
+        cas_rn (str): The CAS registry number of the substance for which to get details.
+
[…]
