### 2023-08-16 6743000
Fixed the condition for maximum numbers of uploads.
With this fix, only the created uploads count towards
the maximum possible amount of unpublished uploads
for an user.

Changelog: Fixed
