### 2023-08-22 b230af6
Command line `--parser` now skips parser matching
The behavior of `nomad parse --parser <parser_name> mainfile` has changed.
Previously, the option `--parse` would just restrict the list of parsers to match.
These would still have to matched, which could fail in cases of alternative main files.

Now, `--parse` skips matching altogether and will apply the specified parser, given it exists.
Note that parsers may still have their own prerequisites for deciding on format or main file type.
These are not bypassed.

Changelog: Changed
