### 2023-08-24 e50dc0f
Removed obsolete elk csv to md script.
Changelog: Changed
