### 2023-08-24 #5388 @zhubonan
Thanks for the review report @berquist ! If my understanding is correct, the overall view is positive, but there are some minor issues that should be addressed.

As for the authorship concern, I think for a large open source projects there can be many contributors (such as fixing bugs), and the changes in the lines does not always reflect the work done (for example, running a code-formatter would result in many line changes). So I think it is OK that the author list does not include all contributors.
PS: I don't seem to be able to find the two authors: Cuauhtemoc Salazar and Daniel Speckhard?

@markus1978 Could you please address comments raised?
