### 2023-08-31 e9726fa
Resolve "ReadableIdentifiers Normalizers Fails"
Fixed an issue where the `ReadableIdentifiers` normalizer would fail when using a custom keycloak that does not specify the user affiliation.

Also, removed old uses of `creation_datetime` and changed it to `datetime` in the `ReadableIdentifiers` normalizer.

Changelog: Fixed
