+++ scripts/generate_example_uploads.sh
@@ -20,3 +20,4 @@ zip -r -j uploads/ellips.zip $REMOTE_TOOLS_PATH/ellips/example/*
 zip -r -j uploads/em_spctrscpy.zip em_spctrscpy/*
 zip -r -j uploads/iv_temp.zip iv_temp/*
 zip -r -j uploads/xps.zip $REMOTE_TOOLS_PATH/xps/example/*
+zip -r -j uploads/sts.zip $REMOTE_TOOLS_PATH/sts/example/*
