+++ examples/data/uploads/example_uploads.yml
@@ -49,6 +49,13 @@ FAIRmat examples:
       It shows the generation of a NeXus file according to the
       [NXmpes](https://manual.nexusformat.org/classes/contributed_definitions/NXmpes.html#nxmpes)
       application definition and a successive analysis of an example data set.
+  sts:
+    path: examples/data/uploads/sts.zip
+    title: STS
+    description: |
+      AT this moment, the reader works for two types of experiments Scanning Tunneling Microscopy (STM) and Scanning Tunneling Spectroscopy (STS) from Scanning Probe Microscopy.
+      It can only transform the data from Nanonis machine generate files into standarized nexus application definition NXsts. In this zip file, the data file are collected from
+      two specific software versions generic 5e and genric 4.5.
   apm:
     path: examples/data/uploads/apm.zip
     title: Electronic Lab Notebook for Atom Probe Microscopy (APM)
