### 2023-09-01 #5388 @markus1978
@berquist Thanks also from me for the review and all the efforts put into it. These comments really help us to improve.

In respond to the authors list.

- Luigi Sbailo's contribution, belongs to another project (AI Toolkit). His commits  to this Github, were experiments in putting a clone of the AI Toolkit homepage into our GUI for a more seamless flow on the central NOMAD installation. The AI Toolkit has its own paper publications where Luigi is recognized as (first) author. It is also referenced in our paper.
- Daniel Speckhard was assigned to the project for a brief moment, but did not really had any commits. I think he should also be removed from the AUTHORS list.
- Regarding Temok Salazar, I agree that he belongs on the authors list and I will add him.

Some authors are placed lower in the GH list (or not there at all), because their contributions are in submodules or not in git at all (design docs, workshops, meetings, etc.). We honestly tried to reflect contributions to NOMAD accurately, even though some choices and the order might not appear as objective as commit metrics. I hope this levitates the concerns about the authors list and is also compliant with JOSS's author policy?

On the conceptual difference to other databases. Maybe it would be clearer if we change

> Other materials science databases (and the respective software) focus on **only** publishing data that were **produced** by the group behind the database itself.

into

> Other materials science databases (and the respective software) focus on publishing data that were produced with a specific framework and **carefully curated** by the group behind the database.

I believe this would also cover Aiida or Material Project contribution (e.g. via [MPContribsAPI](https://docs.materialsproject.org/uploading-data/what-is-mpcontribs)).

On the detailed comparison with other databases/frameworks. A meaningful comparison on the technical and framework differences would be hard to fit into the constrains of a JOSS paper. I agree that it could be a very interesting publication on its own though.

There has been some work on the `setup_dev_env.sh` [issue](https://github.com/nomad-coe/nomad/issues/81). I hope this addresses it sufficiently. As you mentioned, it is really hard to get everything on all platforms outside the docker environment.

We are in the process of moving the central NOMAD installation to a new cloud, i.e. a "more permissions-restrained environment". This will force us to address podman/sigularity issues somewhere in the near future.

Explaining certain NOMAD features to users, partners, and other parties constantly shows new wholes in the documentation. We put them as public todos to also put some pressure on our selfes to fill them soon.

I hope, I could address everything sufficiently. Please let me know, if I missed something. If you (@zhubonan) agree, I will make the mentioned changes to the paper and let you know.
