+++ requirements.txt
@@ -130,7 +130,7 @@ matplotlib==3.5.3         # via ase, asr, diffsims, hyperspy, kikuchipy, matplot
 matplotlib-inline==0.1.6  # via ipykernel, ipython
 matplotlib-scalebar==0.8.1  # via orix
 mdanalysis==2.5.0         # via atomisticparsers (dependencies/parsers/atomistic/pyproject.toml), nomad-lab (pyproject.toml)
-mdtraj==1.9.6             # via nomad-lab (pyproject.toml)
+mdtraj==1.9.9             # via nomad-lab (pyproject.toml)
 memoization==0.4.0        # via nomad-lab (pyproject.toml)
 mistune==2.0.4            # via m2r
 mmtf-python==1.1.3        # via mdanalysis
