### 2023-09-04 #5388 @berquist
> I hope this levitates the concerns about the authors list and is also compliant with JOSS's author policy?

Yes, it does.  Thank you for examining this.  All my other comments are addressed as well, and I'm happy with everything.
