### 2023-09-05 4fbb9be
eDMFT parser and revisit DMFT metainfo
