+++ gui/tests/artifacts.js
@@ -2952,15 +2952,6 @@ window.nomadArtifacts = {
       "aggregatable": true,
       "suggestion": true
     },
-    "results.method.simulation.dmft.total_filling": {
-      "name": "total_filling",
-      "description": "Total filling of the correlated atoms in the unit cell per spin \u2208[0.0, 1.0]. E.g., half-filling\nis defined as 0.5.",
-      "type": {
-        "type_kind": "numpy",
-        "type_data": "float64"
-      },
-      "aggregatable": false
-    },
     "results.method.simulation.dmft.inverse_temperature": {
       "name": "inverse_temperature",
       "description": "Inverse temperature = 1/(kB*T).",
@@ -3008,6 +2999,21 @@ window.nomadArtifacts = {
       "unit": "joule",
       "shape": [],
       "aggregatable": false
+    },
+    "results.method.simulation.dmft.analytical_continuation": {
+      "name": "analytical_continuation",
+      "description": "Analytical continuation used to continuate the imaginary space Green's functions into\nthe real frequencies space.\n\n| Name           | Description         | Reference                        |\n\n| -------------- | ------------------- | -------------------------------- |\n\n| `'Pade'` | Pade's approximant  | https://www.sciencedirect.com/science/article/pii/0021999173901277?via%3Dihub |\n\n| `'MaxEnt'` | Maximum Entropy method | https://journals.aps.org/prb/abstract/10.1103/PhysRevB.41.2380 |\n\n| `'SVD'` | Singular value decomposition | https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.75.517 |\n\n| `'Stochastic'` | Stochastic method | https://journals.aps.org/prb/abstract/10.1103/PhysRevB.57.10287 |",
+      "type": {
+        "type_kind": "Enum",
+        "type_data": [
+          "Pade",
+          "MaxEnt",
+          "SVD",
+          ""
+        ]
+      },
+      "shape": [],
+      "aggregatable": true
     },
     "results.method.simulation.precision.k_line_density": {
       "name": "k_line_density",
@@ -4736,7 +4742,7 @@ window.nomadArtifacts = {
                 "m_parent_index": 1,
                 "m_parent_sub_section": "quantities",
                 "name": "sampling_method",
-                "description": "Method used to generate the mesh:\n\n| Name      | Description                      | Reference             |\n\n| --------- | -------------------------------- | --------------------- |\n\n| `'Gamma-centered'` | Regular mesh is centered around Gamma. No offset. |\n\n| `'Monkhorst-Pack'` | Regular mesh with an offset of half the reciprocal lattice vector. |\n\n| `'Gamma-offcenter'` | Regular mesh with an offset that is neither `'Gamma-centered'`, nor `'Monkhorst-Pack'`. |\n\n| `'Line-path'` | Line path along high-symmetry points. Typically employed for simualting band structures. |\n\n| `'Equidistant'`  | Equidistant 1D grid (also known as 'Newton-Cotes')                      |\n\n| `'Logarithmic'`  | log distance 1D grid               |\n\n| `'Gauss-Legendre'` | Quadrature rule for integration using Legendre polynomials |\n\n| `'Gauss-Laguerre'` | Quadrature rule for integration using Laguerre polynomials |\n\n| `'Clenshaw-Curtis'`  | Quadrature rule for integration using Chebyshev polynomials using discrete cosine transformations |\n\n| `'Gauss-Hermite'`  | Quadrature rule for integration using Hermite polynomials |",
[…]
