+++ tests/normalizing/test_method.py
@@ -172,7 +172,6 @@ def test_method_dmft(dmft):
     assert method.method_name == "DMFT"
     assert method.simulation.program_name == "w2dynamics"
     assert method.simulation.dmft.impurity_solver_type == "CT-HYB"
-    assert method.simulation.dmft.total_filling == 0.5 / 3.0
     assert method.simulation.dmft.inverse_temperature.magnitude == 60.0
     assert method.simulation.dmft.magnetic_state == "paramagnetic"
     assert method.simulation.dmft.u.magnitude == 4.0e-19
