+++ nomad/normalizing/results.py
@@ -25,7 +25,7 @@ import matid.geometry  # pylint: disable=import-error
 
 from nomad import config
 from nomad import atomutils
-from nomad.utils import traverse_reversed
+from nomad.utils import traverse_reversed, extract_section
 from nomad.atomutils import Formula
 from nomad.normalizing.normalizer import Normalizer
 from nomad.normalizing.method import MethodNormalizer
@@ -324,6 +324,7 @@ class ResultsNormalizer(Normalizer):
                 n_channels = values.shape[0]
                 dos_results.spin_polarized = n_channels > 1
                 dos_results.energy_fermi = dos.energy_fermi
+                dos.label = dos.kind
                 for info in dos.band_gap:
                     info_new = BandGapDeprecated().m_from_dict(info.m_to_dict())
                     dos_results.m_add_sub_section(DOSElectronic.band_gap, info_new)
@@ -332,36 +333,34 @@ class ResultsNormalizer(Normalizer):
 
     def resolve_greens_functions(self, path: list[str]) -> Union[List[GreensFunctionsElectronic], None]:
         """Returns a reference to the section containing the electronic Green's functions.
-        In the case of multiple valid Green's functions sections, only the latest one is reported.
-
-        Green's functions are reported only under the following conditions:
-            - There is a non-empty array of greens_function_tau or self_energy_iw or occupancies.
-            - There is a non-empty array of tau or matsubara_freq.
         """
-        greens_functions = traverse_reversed(self.entry_archive, path)
+        greens_functions = extract_section(self.entry_archive, path, full_list=True)
         if not greens_functions:
-            return None
+            return []
         gfs_root: List[GreensFunctionsElectronic] = []
         for gfs in greens_functions:
+            gfs_results = GreensFunctionsElectronic()
             tau = gfs.tau
-            iw = gfs.matsubara_freq
-            values_gtau = np.array([np.absolute(gtau) for gtau in gfs.greens_function_tau.real])
-            values_siw = np.array([siw for siw in gfs.self_energy_iw.imag])
-            if (valid_array(tau) and valid_array(values_gtau)) or (valid_array(iw) and valid_array(values_siw)):
-                gfs_results = GreensFunctionsElectronic()
-                gfs_results.chemical_potential = gfs.chemical_potential
-                if valid_array(tau) and valid_array(values_gtau):
-                    gfs_results.tau = tau
-                    gfs_results.real_greens_function_tau = values_gtau
-                if valid_array(iw) and valid_array(values_siw):
-                    gfs_results.matsubara_freq = iw
-                    gfs_results.imag_self_energy_iw = values_siw
-                if valid_array(gfs.orbital_occupations):
-                    gfs_results.orbital_occupations = gfs.orbital_occupations
-                if valid_array(gfs.quasiparticle_weights):
-                    gfs_results.quasiparticle_weights = gfs.quasiparticle_weights
-
-                gfs_root.insert(0, gfs_results)
+            if valid_array(tau):
+                gfs_results.tau = gfs
+                gfs_results.greens_function_tau = gfs if valid_array(gfs.greens_function_tau) else None
+            matsubara_freq = gfs.matsubara_freq
+            if valid_array(matsubara_freq):
+                gfs_results.matsubara_freq = gfs
+                gfs_results.greens_function_iw = gfs if valid_array(gfs.greens_function_iw) else None
+                gfs_results.self_energy_iw = gfs if valid_array(gfs.self_energy_iw) else None
+            frequencies = gfs.frequencies
+            if valid_array(frequencies):
+                gfs_results.frequencies = gfs
[…]

[…]
