+++ examples/data/em_nion/README.md
@@ -1,0 +1,78 @@
+# Tutorial for parsing nionswift projects for electron microscopy
+
+## Scope
+This is a draft example which combines a set of configured components of 
+NOMAD to collect metadata specific for electron microscopy and combine these with
+images and spectrum data processed using the electron-microscopy-centric image
+processing software nionswift. This software is used to control electron
+microscopes from the Nion Co. Specifically, the tutorial shows how
+functionalities of a NOMAD electronic lab notebook (ELN) instance
+customized for electron microscopy (currently using NXem as the guiding application
+definition), the `em_nion` dataconverter offered through the `pynxtools` NeXus
+parser, the `NXem` NeXus data schema, and several NOMAD core functionalities
+work together to create nionswift-specific NOMAD entries with respective default
+plots for `h5web`.
+
+## Getting started
+1. Click the respective button to execute the example.
+2. Wait for NOMAD to instantiate the example for you.
+3. Check that you are in the `Uploads` section (menu bar).
+4. Select the generated upload.
+5. Explore its content by clicking on the arrow keys (to the right side of a file).
+6. Modify the input file in the NOMAD ELN to match your dataset.
+   You can also use the drag-and-drop functionality to drop your
+   reconstruction and ranging definition file inside the ELN and remove
+   the example reconstruction and ranging file.
+   Please note that this is a draft version of the nionswift parser
+   and thus it is expected that many specific examples may not work out-of-the
+   box.
+7. Click `save` to trigger a reprocessing of your upload which will
+   update and reprocess the NeXus/HDF5 NXem entries for you.
+8. Explore its content via e.g. the preview or via the `Entries`
+   (select from the menu bar).
+
+## Which input is supported by this example?
+This example works with a ZIP file of the nionswift project directory.
+This project dump which can be generated with nionswift (we tested this with
+version 0.16.8 of nionswift) contains a guiding nsproj file which resolves
+individual locations of files in sub-ordinary directories. These can be nionswift
+specific data files, so called ndata files or HDF5 files with the ending h5.
+The ZIP file of the project dump can easily be created by just zipping the
+entire location with the nsproj file at the top.
+
+The ELN is used to collect metadata which are currently not stored typically in
+nionswift. The ELN is configured in this example to collect details about the
+specimen, some exemplar additional metadata about the instrument, and users.
+
+## Which data artefacts should I expect to find?
+This example upload contains the following entries:
+- A schema in NOMAD's *archive.yaml* format: *nxem.schema.archive.yaml*
+- A schema instance file as used and generated by NOMAD *emnion.archive.json*.
+  For educational purposes this file is filled with values but it is the
+  user's responsibility to check that these metadata match their dataset.
+- Another schema instance file *eln_data.yaml*. This file contains a plain
+  view of all entered/modified quantities including their units. This file is
+  updated with clicking the `save` button. This file should not be edited
+  manually. Instead, edit the schema instance using the NOMAD GUI.
+
+## Which data schema is used?
+Specifically the ELN template is configured such that its terminology matches to
+the [NXem](https://fairmat-nfdi.github.io/nexus-fairmat-proposal) NeXus data schema.
+
+## Acknowledgements
+This example comes with a real world dataset collected by Benedikt Haas which
+contains data and metadata collected for different imaging modes.
+[Details can be found here:](https://www.zenodo.org/record/7986279#.ZHh5HKVBxro)
+
+## Where to raise questions, comments, and suggestions?
+You are very welcome to approach us with all sorts of questions, and
+feature requests. [Markus Kühbach](https://www.fairmat-nfdi.eu/fairmat/about-fairmat/team-fairmat)
[…]
