+++ examples/data/apm/apm.oasis.specific.yaml
@@ -1,0 +1,1 @@
+location: Leoben

[…]
