### 2023-09-06 68bd9d2
Added XRD as method name in results.method
