### 2023-09-06 ff09a1c
Added measurements and xrd method sections
