+++ nomad/datamodel/results.py
@@ -1935,6 +1935,53 @@ class Simulation(MSection):
     quantum_cms = SubSection(sub_section=QuantumCMS.m_def, repeats=False)
     precision = SubSection(sub_section=Precision.m_def, repeats=False)
 
+class XRD(MSection):
+    m_def = Section(
+        description='''
+        Methodology for an X-Ray Diffraction measurement.
+        '''
+    )
+    incident_beam_wavelength = Quantity(
+        links=['http://purl.obolibrary.org/obo/BFO_0000001'], #TODO put link to incident_wavelength nexus concept
+        description='''
+        The wavelength of the incident beam.
+        ''',
+        type=np.float64,
+        unit='m',
+        a_elasticsearch=Elasticsearch(material_entry_type),
+    )
+    diffraction_method = Quantity(
+        type=MEnum(
+            [
+                "Powder X-ray Diffraction (PXRD)",
+                "Single Crystal X-ray Diffraction (SCXRD)",
+                "High-Resolution X-ray Diffraction (HRXRD)",
+                "Small-Angle X-ray Scattering (SAXS)",
+                "X-ray Reflectivity (XRR)",
+                "Grazing Incidence X-ray Diffraction (GIXRD)",
+            ]
+        ),
+        description='''
+        The diffraction method used to obtain the diffraction pattern.
+        | X-ray Diffraction Method                                   | Description                                                                                                                                                                                                 |
+        |------------------------------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+        | **Powder X-ray Diffraction (PXRD)**                        | The term "powder" refers more to the random orientation of small crystallites than to the physical form of the sample. Can be used with non-powder samples if they present random crystallite orientations. |
+        | **Single Crystal X-ray Diffraction (SCXRD)**               | Used for determining the atomic structure of a single crystal.                                                                                                                                              |
+        | **High-Resolution X-ray Diffraction (HRXRD)**              | A technique typically used for detailed characterization of epitaxial thin films using precise diffraction measurements.                                                                                    |
+        | **Small-Angle X-ray Scattering (SAXS)**                    | Used for studying nanostructures in the size range of 1-100 nm. Provides information on particle size, shape, and distribution.                                                                             |
+        | **X-ray Reflectivity (XRR)**                               | Used to study thin film layers, interfaces, and multilayers. Provides info on film thickness, density, and roughness.                                                                                       |
+        | **Grazing Incidence X-ray Diffraction (GIXRD)**            | Primarily used for the analysis of thin films with the incident beam at a fixed shallow angle.                                                                                                              |
+        ''',
+        a_elasticsearch=Elasticsearch(material_entry_type),
+    )
+
+class Measurement(MSection):
+    m_def = Section(
+        description='''
+        Contains method details for a measurement entry.
+        '''
+    )
+    xrd = SubSection(sub_section=XRD.m_def, repeats=False)
 
 class Method(MSection):
     m_def = Section(
