### 2023-09-06 2433c60
Added Diffraction Pattern section to structural properties
