+++ gui/tests/artifacts.js
@@ -2505,6 +2505,7 @@ window.nomadArtifacts = {
           "BSE",
           "EELS",
           "XPS",
+          "XRD",
           "unavailable"
         ]
       },
@@ -3080,6 +3081,33 @@ window.nomadArtifacts = {
       },
       "shape": [],
       "aggregatable": false
+    },
+    "results.method.measurement.xrd.incident_beam_wavelength": {
+      "name": "incident_beam_wavelength",
+      "description": "The wavelength of the incident beam.",
+      "type": {
+        "type_kind": "numpy",
+        "type_data": "float64"
+      },
+      "unit": "meter",
+      "aggregatable": false
+    },
+    "results.method.measurement.xrd.diffraction_method_name": {
+      "name": "diffraction_method_name",
+      "description": "The diffraction method used to obtain the diffraction pattern.\n| X-ray Diffraction Method                                   | Description                                                                                                                                                                                                 |\n|------------------------------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|\n| **Powder X-ray Diffraction (PXRD)**                        | The term \"powder\" refers more to the random orientation of small crystallites than to the physical form of the sample. Can be used with non-powder samples if they present random crystallite orientations. |\n| **Single Crystal X-ray Diffraction (SCXRD)**               | Used for determining the atomic structure of a single crystal.                                                                                                                                              |\n| **High-Resolution X-ray Diffraction (HRXRD)**              | A technique typically used for detailed characterization of epitaxial thin films using precise diffraction measurements.                                                                                    |\n| **Small-Angle X-ray Scattering (SAXS)**                    | Used for studying nanostructures in the size range of 1-100 nm. Provides information on particle size, shape, and distribution.                                                                             |\n| **X-ray Reflectivity (XRR)**                               | Used to study thin film layers, interfaces, and multilayers. Provides info on film thickness, density, and roughness.                                                                                       |\n| **Grazing Incidence X-ray Diffraction (GIXRD)**            | Primarily used for the analysis of thin films with the incident beam at a fixed shallow angle.                                                                                                              |",
+      "type": {
+        "type_kind": "Enum",
+        "type_data": [
+          "Powder X-ray Diffraction (PXRD)",
+          "Single Crystal X-ray Diffraction (SCXRD)",
+          "High-Resolution X-ray Diffraction (HRXRD)",
+          "Small-Angle X-ray Scattering (SAXS)",
+          "X-ray Reflectivity (XRR)",
+          "Grazing Incidence X-ray Diffraction (GIXRD)",
+          "unavailable"
+        ]
+      },
+      "aggregatable": true
     },
     "results.properties.n_calculations": {
       "name": "n_calculations",
@@ -4349,6 +4377,18 @@ window.nomadArtifacts = {
       "nested": false,
       "repeats": false
     },
+    "results.method.measurement": {
+      "name": "measurement",
+      "description": "\n        Contains method details for a measurement entry.\n        ",
+      "nested": false,
+      "repeats": false
+    },
+    "results.method.measurement.xrd": {
+      "name": "xrd",
+      "description": "\n        Methodology for an X-Ray Diffraction measurement.\n        ",
+      "nested": false,
+      "repeats": false
+    },
     "results.properties": {
       "name": "properties",
       "description": "\n        Contains the physical properties that have been calculated or used in\n        this entry.\n        ",
@@ -4396,6 +4436,12 @@ window.nomadArtifacts = {
       "description": "\n        Methodology for molecular dynamics.\n        ",
       "nested": false,
[…]
