### 2023-09-06 2e3c770
Added scattering vector quantity and moved incidence_beam_wavelength
