### 2023-09-11 #5388 @markus1978
I made the discussed changes to the paper (https://github.com/nomad-coe/nomad/commit/d734515f71258a1d8f2dcf3b8e0fdac28c6d8092).

I wanted to thank you all the reviewers and the editor again for all the work. Thank you very much for everything.

@berquist What are the next steps?
