### 2023-09-11 434d91d
Add `disabled` status to `Normalize Intensities` option in the DOS kebab menu. Add tests for when option is enabled and disabled.
