### 2023-09-12 #5388 @zhubonan
@markus1978 I have one small comment:

> Materials scientists require effective solutions for managing their research data, but *they cannot
and should not* develop their own individual solution

Sounds a bit extreme to me, as there are developers who are also materials scientists, perhaps changing it to:

> *they should not have to* to develop their own individual solution.
