+++ requirements.txt
@@ -42,6 +42,7 @@ click-didyoumean==0.3.0   # via celery
 click-plugins==1.1.1      # via celery
 click-repl==0.2.0         # via celery
 cloudpickle==2.2.0        # via dask
+comm==0.1.4               # via ipywidgets
 commonmark==0.9.1         # via recommonmark
 cryptography==39.0.0      # via nomad-lab (pyproject.toml), pyjwt, pyopenssl, rfc3161ng
 cycler==0.11.0            # via matplotlib
@@ -96,10 +97,10 @@ imageio==2.24.0           # via hyperspy, kikuchipy, nionswift, nionswift-io, ni
 imagesize==1.4.1          # via sphinx
 importlib-metadata==4.13.0  # via flask, hyperspy, nomad-lab (pyproject.toml), sphinx
 inflection==0.5.1         # via nomad-lab (pyproject.toml)
-ipykernel==6.16.2         # via ipyparallel, ipywidgets
+ipykernel==6.16.2         # via ipyparallel
 ipyparallel==8.4.1        # via hyperspy
 ipython==7.34.0           # via hyperspy, ipykernel, ipyparallel, ipywidgets
-ipywidgets==8.0.7         # via pyxem
+ipywidgets==8.1.1         # via pyxem
 isodate==0.6.1            # via rdflib
 isoduration==20.11.0      # via jsonschema
 itsdangerous==2.1.2       # via flask, nomad-lab (pyproject.toml)
@@ -113,8 +114,8 @@ jupyter-client==7.4.8     # via ipykernel, ipyparallel
 jupyter-core==4.12.0      # via jupyter-client
 jupyter-telemetry==0.1.0  # via jupyterhub
 jupyterhub==1.4.2         # via dockerspawner, nomad-lab (pyproject.toml), oauthenticator
-jupyterlab-widgets==3.0.8  # via ipywidgets
-kikuchipy==0.8.6          # via pynxtools (dependencies/parsers/nexus/pyproject.toml)
+jupyterlab-widgets==3.0.9  # via ipywidgets
+kikuchipy==0.8.7          # via pynxtools (dependencies/parsers/nexus/pyproject.toml)
 kiwisolver==1.4.4         # via matplotlib
 kombu==5.2.4              # via celery
 lark==1.1.5               # via optimade, pynxtools (dependencies/parsers/nexus/pyproject.toml)
@@ -150,7 +151,7 @@ niondata==0.15.5          # via nionswift, nionswift-io
 nionswift==0.16.8         # via pynxtools (dependencies/parsers/nexus/pyproject.toml)
 nionswift-io==0.15.1      # via nionswift
 nionui==0.6.11            # via nionswift
-nionutils==0.4.7          # via niondata, nionswift, nionswift-io, nionui
+nionutils==0.4.8          # via niondata, nionswift, nionswift-io, nionui
 nptyping==1.4.4           # via nomad-lab (pyproject.toml)
 numba==0.56.4             # via diffsims, hyperspy, kikuchipy, orix, pyxem, sparse
 numcodecs==0.10.2         # via zarr
@@ -162,7 +163,7 @@ oauthlib==3.2.2           # via jupyterhub
 openpyxl==3.1.2           # via nomad-lab (pyproject.toml)
 optimade[mongo]==0.22.1   # via nomad-lab (pyproject.toml)
 orix==0.11.1              # via diffsims, kikuchipy, pyxem
-orjson==3.6.0             # via h5grove, nomad-lab (pyproject.toml)
+orjson==3.9.4             # via h5grove, nomad-lab (pyproject.toml)
[…]
