### 2023-09-18 da016c0
added rescaling woodstock enum
