### 2023-09-18 1ff8bb0
Generated gui artifacts.
