### 2023-09-18 520d38c
pylinting, gui artifacts, and atomistic submodule link
