### 2023-09-19 51a40ba
Changed diffraction_pattern to be repeating
