### 2023-09-19 16094f2
Generated gui artifacts
