### 2023-09-19 8713445
Resolve "Application Token"
