### 2023-09-21 183ac18
Add changelog for version v1.2.1 [skip-ci]
