### 2023-09-21 536a657
Add changelog for version v1.2.0 [skip-ci]
