+++ CHANGELOG.md
@@ -39,6 +39,144 @@
 - [Refactored periodic table componeng using SVG.](nomad-lab/nomad-FAIR@ad3a56f8ab0a6a2fabcd0ad1f0a9334eaa8a4b53) ([merge request](nomad-lab/nomad-FAIR!1415))
 - [Removed outdated information about using parsers from the parser projects.](nomad-lab/nomad-FAIR@b41131f576d3fd3574d0e81f67b3c180ebec48b3)
 
+## 1.2.0 (2023-09-21)
+
+### Added (39 changes)
+
+- [CLI utility to export archive data.](nomad-lab/nomad-FAIR@88c59a988e2cfafa321466b2aadef70ef16f74ed) ([merge request](nomad-lab/nomad-FAIR!1452))
+- [Added Activity ID to ExperiementStep](nomad-lab/nomad-FAIR@1ff3610d114820df24b5cfd7a77415b1707a2893) ([merge request](nomad-lab/nomad-FAIR!1456))
+- [Added oasis registration link to the docs.](nomad-lab/nomad-FAIR@3b06ce9a083414da1340be14bbbc48a525436c10) ([merge request](nomad-lab/nomad-FAIR!1450))
+- [Added authentication for mongo and elastic use.](nomad-lab/nomad-FAIR@c1f4147dddbaf6def6e831bd924a170a401b7db8) ([merge request](nomad-lab/nomad-FAIR!1332))
+- [Added YAML linting plugin for NOMAD schemas.](nomad-lab/nomad-FAIR@d6ae88b6a84a7505d0f85d43a9e12a544c3780a0) ([merge request](nomad-lab/nomad-FAIR!1439))
+- [Added System and PureSubstance components for CompositeSystem, added...](nomad-lab/nomad-FAIR@03b23a052eb6d6c2e0cf06596c73b64ec2518f8a) ([merge request](nomad-lab/nomad-FAIR!1385))
+- [Added better search filters when searching for custom schemas.](nomad-lab/nomad-FAIR@d97d90e2c993fad1e5a199e6e7d03c36757578f4) ([merge request](nomad-lab/nomad-FAIR!1335))
+- [Resolve "Refactoring tabular parser"](nomad-lab/nomad-FAIR@1a68e76252f6da76aab40eff7250d99119d9b8e1) ([merge request](nomad-lab/nomad-FAIR!1377))
+- [Added compression suport to quantum expresso mainfiles.](nomad-lab/nomad-FAIR@cb7326b31a3f13fcb9c204c2c45c68144becec80) ([merge request](nomad-lab/nomad-FAIR!1401))
+- [Added a new code contribution guide to the docs.](nomad-lab/nomad-FAIR@5fef1c2eabdc6fcf14f32236ec38d4b57e269793) ([merge request](nomad-lab/nomad-FAIR!1396))
+- [Added helm config options for plugins and normalize.](nomad-lab/nomad-FAIR@73cdacf80c4b1abc54911dfb706791b6fd400be2) ([merge request](nomad-lab/nomad-FAIR!1393))
+- [Added more documentation on data, schemas, and plugins.](nomad-lab/nomad-FAIR@b0d590503e3eea3dff0e30672257cd4c0ab2318d) ([merge request](nomad-lab/nomad-FAIR!1390))
+- [Added new archive config options to helm chart.](nomad-lab/nomad-FAIR@25de71a4e72252e3236c5a6131c69deddf81a2a1) ([merge request](nomad-lab/nomad-FAIR!1251))
+- [Added a soap normalizer normalizer.](nomad-lab/nomad-FAIR@f6c5064f27056e70f920d8aeb5efc2ed4629124f) ([merge request](nomad-lab/nomad-FAIR!1391))
[…]
