+++ paper.md
@@ -21,12 +21,13 @@ authors:
     equal-contrib: true
     affiliation: 1
   - name: Alvin Noe Ladines
-    email: ladinesalvinnoe@gmail.com
+    email: alvin.noe.ladines@physik.hu-berlin.de
     orcid: 0000-0003-0077-2097
     equal-contrib: true
     affiliation: 1
   - name: David Sikter
     email: david.sikter@physik.hu-berlin.de
+    orcid: 0000-0002-2102-7160
     equal-contrib: true
     affiliation: 1
   - name: Mohammad Nakhaee
@@ -36,6 +37,7 @@ authors:
     affiliation: 1
   - name: Ádám Fekete
     email: adam@fekete.co.uk
+    orcid: 0000-0002-6263-897X
     equal-contrib: true
     affiliation: 1
   - name: Theodore Chang
@@ -45,10 +47,12 @@ authors:
     affiliation: 1
   - name: Amir Golparvar
     email: amir.golparvar@physik.hu-berlin.de
+    orcid: 0000-0002-2412-6615
     equal-contrib: true
     affiliation: 1
   - name: José A. Márquez
     email: jose.marquez@physik.hu-berlin.de
+    orcid: 0000-0002-8173-2566
     affiliation: 1
   - name: Sandor Brockhauser
     email: sandor.brockhauser@physik.hu-berlin.de
@@ -56,13 +60,15 @@ authors:
     affiliation: 1
   - name: Sebastian Brückner
     email: sebastian.brueckner@ikz-berlin.de
+    orcid: 0000-0002-5969-847X
     affiliation: 2
   - name: Luca M. Ghiringhelli
     email: luca.ghiringhelli@physik.hu-berlin.de
     orcid: 0000-0001-5099-3029
     affiliation: 1
   - name: Felix Dietrich
-    email: felix.dietrich@tum.de    
+    email: felix.dietrich@tum.de
+    orcid: 0000-0002-2906-1769
     affiliation: 3
   - name: Daniel Lehmberg
     email: d.lehmberg@tum.de
@@ -70,15 +76,19 @@ authors:
     affiliation: 3
   - name: Thea Denell
     email: denell@physik.hu-berlin.de
+    orcid: 0009-0000-7185-9363
     affiliation: 1
   - name: Andrea Albino
     email: andrea.albino@physik.hu-berlin.de
+    orcid: 0000-0001-9280-7431
     affiliation: 1
   - name: Hampus Näsström
     email: hampus.naesstroem@physik.hu-berlin.de
+    orcid: 0000-0002-3264-1692
     affiliation: 1
   - name: Sherjeel Shabih
     email: sherjeel.shabih@physik.hu-berlin.de
+    orcid: 0009-0008-6635-4465
     affiliation: 1
   - name: Florian Dobener
     email: florian.dobener@physik.hu-berlin.de
@@ -90,6 +100,7 @@ authors:
     affiliation: 1
   - name: Rubel Mozumder
     email: mozumder@physik.hu-berlin.de
+    orcid: 0009-0007-5926-6646
     affiliation: 1
   - name: Joseph F. Rudzinski
     email: joseph.rudzinski@physik.hu-berlin.de
@@ -101,9 +112,11 @@ authors:
     affiliation: 1
   - name: José M. Pizarro
     email: jose.pizarro@physik.hu-berlin.de
+    orcid: 0000-0002-6751-8192
     affiliation: 1
   - name: Martin Kuban
     email: kuban@physik.hu-berlin.de
+    orcid: 0000-0002-1619-2460
     affiliation: 1
   - name: Cuauhtemoc Salazar
     email: cuauhtemoc.salazar@physik.hu-berlin.de
@@ -128,7 +141,7 @@ affiliations:
     index: 2
[…]
