+++ paper.bib
@@ -22,12 +22,16 @@
   doi={10.1038/s41586-022-04501-x}
 }
 
-@article{ghiringhelli:2022,
-  title={Shared Metadata for Data-Centric Materials Science},
+@article{ghiringhelli:2023,
+  title={Shared metadata for data-centric materials science},
   author={Ghiringhelli, Luca M and Baldauf, Carsten and Bereau, Tristan and Brockhauser, Sandor and Carbogno, Christian and Chamanara, Javad and Cozzini, Stefano and Curtarolo, Stefano and Draxl, Claudia and Dwaraknath, Shyam and others},
-  journal={arXiv preprint arXiv:2205.14774},
-  year={2022},
-  doi={10.48550/arXiv.2205.14774}
+  journal={Scientific Data},
+  volume={10},
+  number={1},
+  pages={626},
+  year={2023},
+  doi={10.1038/s41597-023-02501-8},
+  publisher={Nature Publishing Group UK London}
 }
 
 @article{andersen:2021,
