### 2023-09-21 #5388 @markus1978
- I pushed the final paper version with all the ORCIDs
- I made a release and the release version is `1.2.1`
- I published the release on zenodo, the DOI is: https://doi.org/10.5281/zenodo.8366163
