### 2023-09-25 21e8e94
Fixed issue in serializing and deserializing boolean values.
