### 2023-09-26 1d03ce3
Minor fix solid dmft
