### 2023-09-27 d7d36dc
Removed service worker.
