### 2023-09-28 c4035ca
Improved the handling of layered systems in PorosityNormalizer.
