+++ nomad/normalizing/mof_deconstructor.py
@@ -231,6 +231,13 @@ def dfsutil_graph_method(graph, temp, node, visited):
     return temp
 
 
+def longest_list(lst):
+    '''
+    return longest list in list of list
+    '''
+    return max(lst, key=len)
+
+
 def remove_unbound_guest(ase_atom):
     '''
     A simple script to remove guest from a metal organic framework.
@@ -264,7 +271,6 @@ def remove_unbound_guest(ase_atom):
             pymat_graph = connected_components(coordination_graph)
             if len(pymat_graph) == 1:
                 polymeric_indices.append(i)
-
         if len(polymeric_indices) > 0:
             Graphs = [StructureGraph.with_local_env_strategy(AseAtomsAdaptor.get_structure(
                 ase_atom[fragments[i]]), JmolNN()) for i in polymeric_indices]
@@ -278,9 +284,12 @@ def remove_unbound_guest(ase_atom):
             mof_indices = []
             for frag_indices in temp_indices:
                 mof_indices.extend(fragments[frag_indices])
-            return mof_indices
+            if len(mof_indices) == 0:
+                return longest_list(fragments)
+            else:
+                return mof_indices
         else:
-            return sum(sum(fragments, []))
+            return sum(fragments, [])
 
 
 def connected_components(graph):
@@ -528,6 +537,68 @@ def find_carbonyl_sulphate(ase_atom):
     return sulphate
 
 
+def find_sulfides(ase_atom):
+    '''
+    A simple aglorimth to search for sulfides.
+     S
+     |
+    -C
+     |
+     S
+
+    Parameters:
+    -----------
+    ase_atom: ASE atom
+
+    Returns
+    -------
+    dictionary of key = carbon index and values = sulphur index
+    '''
+    graph, _ = compute_ase_neighbour(ase_atom)
+    sulfides = {}
+    for atoms in ase_atom:
+        if atoms.symbol == 'C':
+            index = atoms.index
+            sulphure_atoms = [i for i in graph[index]
+                              if ase_atom[i].symbol == 'S']
+            if len(sulphure_atoms) == 2:
+                sulphure_to_metal = sum(
+                    [[j for j in graph[i] if ase_atom[j].symbol in transition_metals()] for i in sulphure_atoms], [])
+                if len(sulphure_to_metal) > 0:
+                    sulfides[index] = sulphure_atoms
+    return sulfides
+
+
+def find_phosphate(ase_atom):
+    '''
+    A simple algorithm to search for Carbonyl sulphate found in the system.
+       O
+       |
+      -P-o
+       |
+       O
+    Parameters:
+    -----------
+    ase_atom: ASE atom
+
+    Returns
+    -------
+    dictionary of key = carbon index and values = oxygen index
+    '''
+    graph, _ = compute_ase_neighbour(ase_atom)
+    phosphate = {}
+    for atoms in ase_atom:
+        if atoms.symbol == 'P':
+            index = atoms.index
+            oxygen = [i for i in graph[index] if ase_atom[i].symbol == 'O']
+            if len(oxygen) >= 1:
+                oxy_metal = sum(
+                    [[j for j in graph[i] if ase_atom[j].symbol in transition_metals()] for i in oxygen], [])
+                if len(oxy_metal) > 0:
+                    phosphate[index] = oxygen
+    return phosphate
+
+
 def secondary_building_units(ase_atom):
     """
     1) Search for all carboxylate that are connected to a metal.
@@ -553,8 +624,13 @@ def secondary_building_units(ase_atom):
     bonds_to_break = []
     carboxylates = find_carboxylates(ase_atom)
     all_sulphates = find_carbonyl_sulphate(ase_atom)
[…]
