### 2023-09-28 079219e
Fixes the NOMAD app only running one worker.
No matter what the --workers parameter, the nomad run command is only
running one process. I added gunicorn as a process manager for production
settings. The run.sh command is now using gunicorn and defaults to
4 workers.

Changelog: Fixed
