+++ ops/kubernetes/deployments/prod-test-values.yaml
@@ -20,7 +20,6 @@ gui:
 
 app:
   replicas: 2
-  worker: 4
 
 worker:
   replicas: 1
