+++ ops/kubernetes/deployments/prod-develop-values.yaml
@@ -24,8 +24,6 @@ gui:
 
 app:
   replicas: 2
-  worker: 4
-  threads: 4
   nomadNodeType: "public"
 
 worker:
