### 2023-10-01 #5388 @Kevin-Mattheus-Moerman
@markus1978 I am the AEiC for this track and here to help process final steps. I have check this review, the repository, the archive link and the paper. Most seems in order but below are some minor points that needs your attention:

- [ ] Please update the ZENODO archive listed version to `v1.2.1`.
- [ ] You mention the authors are doing a final review. Please let me know when this is finished, so we can proceed here.
