### 2023-10-02 85afaf8
Fixed issue in displaying exclusive and open-ended range queries in the Range component.
