### 2023-10-02 6fef92a
Fixed issue with parseQuery not picking up the units correctly, added unit tests for parseQuery.
