### 2023-10-02 a7f2363
Fixed missing cache-control and etags.
We removed the service worker from the frontend and rely now
browsers caching gui (and other) resources. Fixed the cache control
headers on all static resources and added cache control and
etags for artifacts.js and env.js.

Changelog: Fixed
