+++ gui/src/components/search/FilterRegistry.js
@@ -316,6 +316,7 @@ registerFilter('entry_id', idMetadata, termQuantity)
 registerFilter('entry_name', idMetadata, termQuantity)
 registerFilter('mainfile', idMetadata, termQuantity)
 registerFilter('upload_id', idMetadata, termQuantity)
+registerFilter('upload_name', idMetadata, termQuantity)
 registerFilter('published', idMetadata, termQuantity)
 registerFilter('main_author.user_id', idMetadata, termQuantity)
 registerFilter('quantities', idMetadata, {...noAggQuantity, label: 'Metainfo definition', queryMode: 'all'})
