### 2023-10-03 2038464
Added new input component for configs, started using YAML in dashboard config, improved the visuals of the dashboard config dialog.
Changelog: Added YAML serialization for dashboard export/import.
