+++ gui/src/components/search/widgets/Dashboard.js
@@ -41,7 +41,7 @@ import { WidgetScatterPlotEdit, schemaWidgetScatterPlot } from './WidgetScatterP
 import { WidgetHistogramEdit, schemaWidgetHistogram } from './WidgetHistogram'
 import { WidgetTermsEdit, schemaWidgetTerms } from './WidgetTerms'
 import { WidgetPeriodicTableEdit, schemaWidgetPeriodicTable } from './WidgetPeriodicTable'
-import { JsonEditor } from '../../archive/SectionEditor'
+import InputConfig from '../input/InputConfig'
 import Markdown from '../../Markdown'
 import { isArray } from 'lodash'
 import { getWidgetsObject } from './Widget'
@@ -106,9 +106,9 @@ const Dashboard = React.memo(() => {
         xl: {...layout},
         xxl: {...layout}
       },
-      type: 'scatterplot',
       size: 1000,
-      autorange: true
+      autorange: true,
+      type: 'scatterplot'
     }
     addWidget(id, value)
   }, [addWidget])
@@ -129,9 +129,9 @@ const Dashboard = React.memo(() => {
         xl: {...layout},
         xxl: {...layout}
       },
-      type: 'periodictable',
+      scale: 'linear',
       quantity: 'results.material.elements',
-      scale: 'linear'
+      type: 'periodictable'
     }
     addWidget(id, value)
   }, [addWidget])
@@ -152,10 +152,10 @@ const Dashboard = React.memo(() => {
         xl: {...layout},
         xxl: {...layout}
       },
-      type: 'histogram',
       autorange: false,
       nbins: 30,
-      scale: 'linear'
+      scale: 'linear',
+      type: 'histogram'
     }
     addWidget(id, value)
   }, [addWidget])
@@ -176,8 +176,8 @@ const Dashboard = React.memo(() => {
         xl: {...layout},
         xxl: {...layout}
       },
-      type: 'terms',
-      scale: 'linear'
+      scale: 'linear',
+      type: 'terms'
     }
     addWidget(id, value)
   }, [addWidget])
@@ -224,12 +224,12 @@ const Dashboard = React.memo(() => {
         </Action>
         <Action
           tooltip=""
-          ButtonComponent={DashboardJSONButton}
+          ButtonComponent={DashboardExportButton}
           ButtonProps={{
             tooltip: "Dashboard export/import",
             title: "Dashboard export/import",
             DialogProps: {
-              maxWidth: "lg",
+              maxWidth: "md",
               fullWidth: true
             },
             ButtonProps: {
@@ -279,13 +279,20 @@ DashboardAction.propTypes = {
   children: PropTypes.node
 }
 
+const schemas = {
+  scatterplot: schemaWidgetScatterPlot,
+  periodictable: schemaWidgetPeriodicTable,
+  histogram: schemaWidgetHistogram,
+  terms: schemaWidgetTerms
+}
+
 /**
  * A button that displays a dialog that can be used to modify the current
  * dashboard setup.
  */
-export const DashboardJSONButton = React.memo((props) => {
+export const DashboardExportButton = React.memo((props) => {
   const {tooltip, title, DialogProps, ButtonProps} = props
-  const { useWidgetsState } = useSearchContext()
+  const {useWidgetsState} = useSearchContext()
   const [widgets, setWidgets] = useWidgetsState()
   const [widgetExport, setWidgetExport] = useState()
   const [widgetImport, setWidgetImport] = useState()
@@ -298,13 +305,19 @@ export const DashboardJSONButton = React.memo((props) => {
     const exp = Object
       .values(widgets)
       .map((widget) => {
-        const schemas = {
-          scatterplot: schemaWidgetScatterPlot,
[…]
