### 2023-10-03 b9560e8
Resolve "Implement ordering of simulation workflow tasks by execution time"
Implemented creation of tasks in normalization of SimulationWorkflow based on
time_calculation and time_physical.
Update atomistic and electronic parsers to parse these quantities.

Changelog: Added
