### 2023-10-03 #5388 @zhubonan
Or the alternative is to re-release with a `1.2.1` tag?
