### 2023-10-05 a30ae5b
Resolve "Revising Tabular parser features"
Tabular parser is updated with several new features and bug fixes: new entries can be created directly from the schema by providing mapping_options under tabular_parser annotation, each mapping_options section now create a new archive file, added further customization of the entry name, added comprehensive documentation under How-to and References, deprecated the old TableDataParser

Changelog: Added
