+++ nomad/parsing/tabular.py
@@ -16,8 +16,7 @@
 # limitations under the License.
 #
 import os
-import time
-from typing import List, Dict, Callable, Set, Any, Tuple, Iterator, Union, Iterable, cast
+from typing import List, Dict, Callable, Set, Any, Tuple, Iterator, Union, Iterable
 
 import pandas as pd
 from memoization import cached
@@ -30,11 +29,9 @@ import yaml
 from nomad import utils
 from nomad.parsing import MatchingParser
 from nomad.units import ureg
-from nomad.datamodel import EntryArchive, EntryMetadata
-from nomad.datamodel.data import ArchiveSection, EntryData
+from nomad.datamodel.data import ArchiveSection
 from nomad.metainfo import Section, Quantity, Package, Reference, MSection, Property
-from nomad.metainfo.metainfo import MetainfoError, SubSection, MProxy, SectionProxy
-from nomad.datamodel.context import Context
+from nomad.metainfo.metainfo import MetainfoError, SubSection, MProxy
 from nomad.datamodel.metainfo.annotations import TabularAnnotation, TabularParserAnnotation, TabularFileModeEnum, \
     TabularMode
 from nomad.metainfo.util import MSubSectionList
@@ -47,7 +44,12 @@ from nomad.utils import generate_entry_id
 m_package = Package()
 
 
-def create_archive(entry_dict, context, file_name, file_type):
+root_mapping = {
+    'root': '#root'
+}
+
+
+def create_archive(entry_dict, context, file_name, file_type, logger):
     if not context.raw_path_exists(file_name):
         with context.raw_file(file_name, 'w') as outfile:
             if file_type == 'json':
@@ -55,6 +57,10 @@ def create_archive(entry_dict, context, file_name, file_type):
             elif file_type == 'yaml':
                 yaml.dump(entry_dict, outfile)
         context.upload.process_updated_raw_file(file_name, allow_modify=True)
+    else:
+        logger.error(
+            f'{file_name} archive file already exists.'
+            f'If you intend to reprocess the older archive file, remove the existing one and run reprocessing again.')
 
 
 def traverse_to_target_data_file(section, path_list: List[str]):
@@ -66,20 +72,6 @@ def traverse_to_target_data_file(section, path_list: List[str]):
             return traverse_to_target_data_file(getattr(section, temp), path_list)
         except AttributeError:
             raise MetainfoError(f'The path {temp} in path_to_data_file does not exist')
-
-
-class TableRow(EntryData):
-    ''' Represents the data in one row of a table. '''
-    table_ref = Quantity(
-        type=Reference(SectionProxy('Table')),
-        description='A reference to the table that this row is contained in.')
-
-
-class Table(EntryData):
-    ''' Represents a table with many rows. '''
-    row_refs = Quantity(
-        type=Reference(TableRow.m_def), shape=['*'],
-        description='References that connect to each row. Each row is stored in it individual entry.')
 
 
 class TabularParserError(Exception):
@@ -106,7 +98,9 @@ class TableData(ArchiveSection):
             for quantity_def in self.m_def.all_quantities.values():
                 annotation = quantity_def.m_get_annotations('tabular_parser')
                 annotation = annotation[0] if isinstance(annotation, list) else annotation
-                if annotation:
+                # this normalizer potentially creates new archives,.to avoid recursive call of the normalizer created by
+                # this normalizer, we check for if the parent is already parsed
+                if annotation and not archive.data.m_parent.metadata.entry_name:
                     self.tabular_parser(quantity_def, archive, logger, annotation)
 
     def tabular_parser(self, quantity_def: Quantity, archive, logger, annotation: TabularParserAnnotation):
@@ -138,30 +132,6 @@ class TableData(ArchiveSection):
[…]
