### 2023-10-05 #5388 @markus1978
@Kevin-Mattheus-Moerman

- I comited a new version of the paper with spelling/grammar things, and one affiliation change.
- I updated the version on zenodo and added the "v".

I am sorry for the late changes to the paper, I hope this does not cause too many issues. But i think all the co-authors are now happy and we can proceed.
