### 2023-10-05 #5388 @Kevin-Mattheus-Moerman
@markus1978 thanks for making those changes. I have some more minor points below:

- [ ] Since the acronym `(ELN)` is not used, you can remove it from the text.
- [ ] Check `... following the instruction ...` which should perhaps be `... following the instructions ...`
- [ ] Your text mentions: `NOMAD Oasis can be used freely as per our OSI license...`, I recommend that you actually mention the license you use (Apache 2.0).
- [ ] In fact, given the above two points I would recommend to rephrase the following sentence to target the above:
`NOMAD Oasis can be used freely as per our OSI license following the instruction in the [NOMAD documentation](https://nomad-lab.eu/docs).`
to something like:
`NOMAD Oasis can be used freely, is available open source under the [Apache License 2.0](https://github.com/nomad-coe/nomad/blob/develop/LICENSE), and detailed instructions are available in the [NOMAD documentation](https://nomad-lab.eu/docs).`
