### 2023-10-05 84a96d3
Fixed etags produced by fastapi to work with nginx.
Changelog: Fixed
