### 2023-10-05 #5388 @markus1978
@Kevin-Mattheus-Moerman thanks you very much for your suggestions. We now follow your suggestion for the last 3 points. For the first point, I would like to keep the acronym if possible. "ELN" is so prevalent in our community that for many readers "ELN" is more recognisable than "electronic lab notebook".
