+++ gui/tests/env.js
@@ -11,7 +11,7 @@ window.nomadEnv = {
   "version": {},
   "globalLoginRequired": false,
   "servicesUploadLimit": 10,
-  "appTokenMaxExpiresIn": 86400,
+  "appTokenMaxExpiresIn": 2592000,
   "ui": {
     "app_base": "http://localhost:8000/fairdi/nomad/latest",
     "north_base": "http://localhost:9000/fairdi/nomad/latest/north",
