### 2023-10-09 ac07c9b
Added v1.2 API skeleton and graph API pydantic models.
