+++ nomad/graph/graph_reader.py
@@ -29,15 +29,16 @@ from mongoengine import Q
 
 from nomad import utils
 from nomad.app.v1.models import (
-    Direction, WithQuery, MetadataPagination, Pagination, PaginationResponse, MetadataRequired
+    MetadataPagination, Pagination, PaginationResponse, Metadata
 )
 from nomad.app.v1.routers.datasets import DatasetPagination
 from nomad.app.v1.routers.entries import perform_search
 from nomad.app.v1.routers.uploads import (
-    get_upload_with_read_access, upload_to_pydantic, entry_to_pydantic, UploadProcDataQuery, UploadProcDataPagination
+    get_upload_with_read_access, upload_to_pydantic, entry_to_pydantic, UploadProcDataQuery, UploadProcDataPagination,
+    EntryProcDataPagination
 )
 from nomad.archive import ArchiveList, ArchiveDict, to_json
-from nomad.archive.model import RequestConfig, DefinitionType, DirectiveType, ResolveType, DatasetQuery
+from nomad.graph.model import RequestConfig, DefinitionType, DirectiveType, ResolveType, DatasetQuery, EntryQuery
 from nomad.datamodel import ServerContext, User, EntryArchive, Dataset
 from nomad.datamodel.util import parse_path
 from nomad.files import UploadFiles, RawPathInfo
@@ -58,14 +59,22 @@ class Token:
     It is thus recommended to use a prefix to avoid collision.
     '''
     DEF = 'm_def'
-    RAW = 'm_raw'
-    ARCHIVE = 'm_archive'
-    ENTRY = 'm_entries'
-    UPLOAD = 'm_uploads'
-    USER = 'm_users'
-    DATASET = 'm_datasets'
-    METAINFO = 'm_metainfo'
+    RAW = 'files'
+    ARCHIVE = 'archive'
+    ENTRY = 'entry'
+    ENTRIES = 'entries'
+    UPLOAD = 'upload'
+    UPLOADS = 'uploads'
+    USER = 'user'
+    USERS = 'users'
+    DATASET = 'dataset'
+    DATASETS = 'm_datasets'
+    METAINFO = 'metainfo'
+    SEARCH = 'search'
     ERROR = 'm_errors'
+    METADATA = 'metadata'
+    MAINFILE = 'mainfile'
+    RESPONSE = 'm_response'
 
 
 @dataclasses.dataclass(frozen=True)
@@ -73,6 +82,7 @@ class QueryError:
     NOACCESS = 'NOACCESS'
     NOTFOUND = 'NOTFOUND'
     ARCHIVEERROR = 'ARCHIVEERROR'
+    GENERAL = 'GENERAL'
 
 
 def dataset_to_pydantic(item):
@@ -98,7 +108,7 @@ class ConfigError(Exception):
 
 
 @dataclasses.dataclass(frozen=True)
-class ArchiveNode:
+class GraphNode:
     upload_id: str  # the upload id of the current node
     entry_id: str  # the entry id of the current node
     current_path: list[str]  # the path in the result container
@@ -128,13 +138,13 @@ class ArchiveNode:
     def _generate_prefix(self) -> str:
         return f'../uploads/{self.upload_id}/archive/{self.entry_id}'
 
-    def goto(self, reference: str, resolve_inplace: bool) -> ArchiveNode:
+    def goto(self, reference: str, resolve_inplace: bool) -> GraphNode:
         if reference.startswith(('/', '#')):
             return self._goto_local(reference, resolve_inplace)
 
         return self._goto_remote(reference, resolve_inplace)
 
-    def _goto_local(self, reference: str, resolve_inplace: bool) -> ArchiveNode:
+    def _goto_local(self, reference: str, resolve_inplace: bool) -> GraphNode:
         '''
         Go to a local reference.
[…]
