### 2023-10-10 7215bc3
Resolve "Graph model generation, metadata token support"
