+++ nomad/app/v1/models/__main__.py
@@ -1,5 +1,0 @@
-import sys
-
-setattr(sys.modules["nomad.app.v1.models.utils"], "ref_prefix", "#/definitions")
-model = getattr(sys.modules["nomad.app.v1.graph_models"], sys.argv[1])
-print(model.schema_json(indent=2))
