### 2023-10-13 c3304ff
Update file parser and parser documentation
Updates the example parser and example parser plug in.
Updates the documentation on how to write a parser.
Minor improvements and documentation for file parser module.

Changelog: Fixed
